@@ -1,10 +1,8 @@
-# type: ignore
 # Tests for aiohttp/protocol.py
 
 import asyncio
 import re
-from contextlib import nullcontext
-from typing import Any, Dict, List
+from typing import Any, Dict, Iterable, List, Type
 from unittest import mock
 from urllib.parse import quote
 
@@ -14,14 +12,18 @@ from yarl import URL
 
 import aiohttp
 from aiohttp import http_exceptions, streams
+from aiohttp.base_protocol import BaseProtocol
+from aiohttp.helpers import NO_EXTENSIONS
 from aiohttp.http_parser import (
-    NO_EXTENSIONS,
     DeflateBuffer,
+    HttpParser,
     HttpPayloadParser,
+    HttpRequestParser,
     HttpRequestParserPy,
+    HttpResponseParser,
     HttpResponseParserPy,
-    HttpVersion,
 )
+from aiohttp.http_writer import HttpVersion
 
 try:
     try:
@@ -31,8 +33,8 @@ try:
 except ImportError:
     brotli = None
 
-REQUEST_PARSERS: Any = [HttpRequestParserPy]
-RESPONSE_PARSERS: Any = [HttpResponseParserPy]
+REQUEST_PARSERS = [HttpRequestParserPy]
+RESPONSE_PARSERS = [HttpResponseParserPy]
 
 try:
     from aiohttp.http_parser import HttpRequestParserC, HttpResponseParserC
@@ -44,11 +46,11 @@ except ImportError:  # pragma: no cover
 
 
 @pytest.fixture
-def protocol():
-    return mock.Mock()
+def protocol() -> Any:
+    return mock.create_autospec(BaseProtocol, spec_set=True, instance=True)
 
 
-def _gen_ids(parsers: List[Any]) -> List[str]:
+def _gen_ids(parsers: Iterable[Type[HttpParser[Any]]]) -> List[str]:
     return [
         "py-parser" if parser.__module__ == "aiohttp.http_parser" else "c-parser"
         for parser in parsers
@@ -56,9 +58,13 @@ def _gen_ids(parsers: List[Any]) -> List[str]:
 
 
 @pytest.fixture(params=REQUEST_PARSERS, ids=_gen_ids(REQUEST_PARSERS))
-def parser(loop: Any, protocol: Any, request: Any):
+def parser(
+    loop: asyncio.AbstractEventLoop,
+    protocol: BaseProtocol,
+    request: pytest.FixtureRequest,
+) -> HttpRequestParser:
     # Parser implementations
-    return request.param(
+    return request.param(  # type: ignore[no-any-return]
         protocol,
         loop,
         2**16,
@@ -68,43 +74,43 @@ def parser(loop: Any, protocol: Any, request: Any):
 
 
 @pytest.fixture(params=REQUEST_PARSERS, ids=_gen_ids(REQUEST_PARSERS))
-def request_cls(request: Any):
+def request_cls(request: pytest.FixtureRequest) -> Type[HttpRequestParser]:
     # Request Parser class
-    return request.param
+    return request.param  # type: ignore[no-any-return]
 
 
 @pytest.fixture(params=RESPONSE_PARSERS, ids=_gen_ids(RESPONSE_PARSERS))
-def response(loop: Any, protocol: Any, request: Any):
+def response(
+    loop: asyncio.AbstractEventLoop,
+    protocol: BaseProtocol,
+    request: pytest.FixtureRequest,
+) -> HttpResponseParser:
     # Parser implementations
-    return request.param(
+    return request.param(  # type: ignore[no-any-return]
         protocol,
         loop,
         2**16,
         max_line_size=8190,
         max_field_size=8190,
+        read_until_eof=True,
     )
 
 
 @pytest.fixture(params=RESPONSE_PARSERS, ids=_gen_ids(RESPONSE_PARSERS))
-def response_cls(request: Any):
+def response_cls(request: pytest.FixtureRequest) -> Type[HttpResponseParser]:
     # Parser implementations
-    return request.param
-
-
-@pytest.fixture
-def stream():
-    return mock.Mock()
+    return request.param  # type: ignore[no-any-return]
 
 
 @pytest.mark.skipif(NO_EXTENSIONS, reason="Extensions available but not imported")
-def test_c_parser_loaded():
+def test_c_parser_loaded() -> None:
     assert "HttpRequestParserC" in dir(aiohttp.http_parser)
     assert "HttpResponseParserC" in dir(aiohttp.http_parser)
     assert "RawRequestMessageC" in dir(aiohttp.http_parser)
     assert "RawResponseMessageC" in dir(aiohttp.http_parser)
 
 
-def test_parse_headers(parser: Any) -> None:
+def test_parse_headers(parser: HttpRequestParser) -> None:
     text = b"""GET /test HTTP/1.1\r
 test: a line\r
 test2: data\r
@@ -121,7 +127,7 @@ test2: data\r
     assert not msg.upgrade
 
 
-def test_reject_obsolete_line_folding(parser: Any) -> None:
+def test_reject_obsolete_line_folding(parser: HttpRequestParser) -> None:
     text = b"""GET /test HTTP/1.1\r
 test: line\r
  Content-Length: 48\r
@@ -133,7 +139,11 @@ test2: data\r
 
 
 @pytest.mark.skipif(NO_EXTENSIONS, reason="Only tests C parser.")
-def test_invalid_character(loop: Any, protocol: Any, request: Any) -> None:
+def test_invalid_character(
+    loop: asyncio.AbstractEventLoop,
+    protocol: BaseProtocol,
+    request: pytest.FixtureRequest,
+) -> None:
     parser = HttpRequestParserC(
         protocol,
         loop,
@@ -153,7 +163,11 @@ def test_invalid_character(loop: Any, protocol: Any, request: Any) -> None:
 
 
 @pytest.mark.skipif(NO_EXTENSIONS, reason="Only tests C parser.")
-def test_invalid_linebreak(loop: Any, protocol: Any, request: Any) -> None:
+def test_invalid_linebreak(
+    loop: asyncio.AbstractEventLoop,
+    protocol: BaseProtocol,
+    request: pytest.FixtureRequest,
+) -> None:
     parser = HttpRequestParserC(
         protocol,
         loop,
@@ -172,7 +186,7 @@ def test_invalid_linebreak(loop: Any, protocol: Any, request: Any) -> None:
         parser.feed_data(text)
 
 
-def test_cve_2023_37276(parser: Any) -> None:
+def test_cve_2023_37276(parser: HttpRequestParser) -> None:
     text = b"""POST / HTTP/1.1\r\nHost: localhost:8080\r\nX-Abc: \rxTransfer-Encoding: chunked\r\n\r\n"""
     with pytest.raises(http_exceptions.BadHttpMessage):
         parser.feed_data(text)
@@ -182,13 +196,16 @@ def test_cve_2023_37276(parser: Any) -> None:
     "rfc9110_5_6_2_token_delim",
     r'"(),/:;<=>?@[\]{}',
 )
-def test_bad_header_name(parser: Any, rfc9110_5_6_2_token_delim: str) -> None:
+def test_bad_header_name(
+    parser: HttpRequestParser, rfc9110_5_6_2_token_delim: str
+) -> None:
     text = f"POST / HTTP/1.1\r\nhead{rfc9110_5_6_2_token_delim}er: val\r\n\r\n".encode()
-    expectation = pytest.raises(http_exceptions.BadHttpMessage)
     if rfc9110_5_6_2_token_delim == ":":
         # Inserting colon into header just splits name/value earlier.
-        expectation = nullcontext()
-    with expectation:
+        parser.feed_data(text)
+        return
+
+    with pytest.raises(http_exceptions.BadHttpMessage):
         parser.feed_data(text)
 
 
@@ -207,13 +224,15 @@ def test_bad_header_name(parser: Any, rfc9110_5_6_2_token_delim: str) -> None:
         "\xffoo: bar",
     ),
 )
-def test_bad_headers(parser: Any, hdr: str) -> None:
+def test_bad_headers(parser: HttpRequestParser, hdr: str) -> None:
     text = f"POST / HTTP/1.1\r\n{hdr}\r\n\r\n".encode()
     with pytest.raises(http_exceptions.BadHttpMessage):
         parser.feed_data(text)
 
 
-def test_unpaired_surrogate_in_header_py(loop: Any, protocol: Any) -> None:
+def test_unpaired_surrogate_in_header_py(
+    loop: asyncio.AbstractEventLoop, protocol: BaseProtocol
+) -> None:
     parser = HttpRequestParserPy(
         protocol,
         loop,
@@ -230,7 +249,7 @@ def test_unpaired_surrogate_in_header_py(loop: Any, protocol: Any) -> None:
     assert message is not None
 
 
-def test_content_length_transfer_encoding(parser: Any) -> None:
+def test_content_length_transfer_encoding(parser: HttpRequestParser) -> None:
     text = (
         b"GET / HTTP/1.1\r\nHost: a\r\nContent-Length: 5\r\nTransfer-Encoding: a\r\n\r\n"
         + b"apple\r\n"
@@ -239,7 +258,9 @@ def test_content_length_transfer_encoding(parser: Any) -> None:
         parser.feed_data(text)
 
 
-def test_bad_chunked_py(loop: Any, protocol: Any) -> None:
+def test_bad_chunked_py(
+    loop: asyncio.AbstractEventLoop, protocol: BaseProtocol
+) -> None:
     """Test that invalid chunked encoding doesn't allow content-length to be used."""
     parser = HttpRequestParserPy(
         protocol,
@@ -260,7 +281,7 @@ def test_bad_chunked_py(loop: Any, protocol: Any) -> None:
     "HttpRequestParserC" not in dir(aiohttp.http_parser),
     reason="C based HTTP parser not available",
 )
-def test_bad_chunked_c(loop: Any, protocol: Any) -> None:
+def test_bad_chunked_c(loop: asyncio.AbstractEventLoop, protocol: BaseProtocol) -> None:
     """C parser behaves differently. Maybe we should align them later."""
     parser = HttpRequestParserC(
         protocol,
@@ -277,13 +298,13 @@ def test_bad_chunked_c(loop: Any, protocol: Any) -> None:
         parser.feed_data(text)
 
 
-def test_whitespace_before_header(parser: Any) -> None:
+def test_whitespace_before_header(parser: HttpRequestParser) -> None:
     text = b"GET / HTTP/1.1\r\n\tContent-Length: 1\r\n\r\nX"
     with pytest.raises(http_exceptions.BadHttpMessage):
         parser.feed_data(text)
 
 
-def test_parse_headers_longline(parser: Any) -> None:
+def test_parse_headers_longline(parser: HttpRequestParser) -> None:
     invalid_unicode_byte = b"\xd9"
     header_name = b"Test" + invalid_unicode_byte + b"Header" + b"A" * 8192
     text = b"GET /test HTTP/1.1\r\n" + header_name + b": test\r\n" + b"\r\n" + b"\r\n"
@@ -293,7 +314,7 @@ def test_parse_headers_longline(parser: Any) -> None:
 
 
 @pytest.fixture
-def xfail_c_parser_status(request) -> None:
+def xfail_c_parser_status(request: pytest.FixtureRequest) -> None:
     if isinstance(request.getfixturevalue("parser"), HttpRequestParserPy):
         return
     request.node.add_marker(
@@ -305,7 +326,7 @@ def xfail_c_parser_status(request) -> None:
 
 
 @pytest.mark.usefixtures("xfail_c_parser_status")
-def test_parse_unusual_request_line(parser: Any) -> None:
+def test_parse_unusual_request_line(parser: HttpRequestParser) -> None:
     text = b"#smol //a HTTP/1.3\r\n\r\n"
     messages, upgrade, tail = parser.feed_data(text)
     assert len(messages) == 1
@@ -317,7 +338,7 @@ def test_parse_unusual_request_line(parser: Any) -> None:
     assert msg.version == (1, 3)
 
 
-def test_parse(parser: Any) -> None:
+def test_parse(parser: HttpRequestParser) -> None:
     text = b"GET /test HTTP/1.1\r\n\r\n"
     messages, upgrade, tail = parser.feed_data(text)
     assert len(messages) == 1
@@ -329,7 +350,7 @@ def test_parse(parser: Any) -> None:
     assert msg.version == (1, 1)
 
 
-async def test_parse_body(parser: Any) -> None:
+async def test_parse_body(parser: HttpRequestParser) -> None:
     text = b"GET /test HTTP/1.1\r\nContent-Length: 4\r\n\r\nbody"
     messages, upgrade, tail = parser.feed_data(text)
     assert len(messages) == 1
@@ -338,7 +359,7 @@ async def test_parse_body(parser: Any) -> None:
     assert body == b"body"
 
 
-async def test_parse_body_with_CRLF(parser: Any) -> None:
+async def test_parse_body_with_CRLF(parser: HttpRequestParser) -> None:
     text = b"\r\nGET /test HTTP/1.1\r\nContent-Length: 4\r\n\r\nbody"
     messages, upgrade, tail = parser.feed_data(text)
     assert len(messages) == 1
@@ -347,7 +368,7 @@ async def test_parse_body_with_CRLF(parser: Any) -> None:
     assert body == b"body"
 
 
-def test_parse_delayed(parser: Any) -> None:
+def test_parse_delayed(parser: HttpRequestParser) -> None:
     text = b"GET /test HTTP/1.1\r\n"
     messages, upgrade, tail = parser.feed_data(text)
     assert len(messages) == 0
@@ -359,7 +380,7 @@ def test_parse_delayed(parser: Any) -> None:
     assert msg.method == "GET"
 
 
-def test_headers_multi_feed(parser: Any) -> None:
+def test_headers_multi_feed(parser: HttpRequestParser) -> None:
     text1 = b"GET /test HTTP/1.1\r\n"
     text2 = b"test: line"
     text3 = b" continue\r\n\r\n"
@@ -381,7 +402,7 @@ def test_headers_multi_feed(parser: Any) -> None:
     assert not msg.upgrade
 
 
-def test_headers_split_field(parser: Any) -> None:
+def test_headers_split_field(parser: HttpRequestParser) -> None:
     text1 = b"GET /test HTTP/1.1\r\n"
     text2 = b"t"
     text3 = b"es"
@@ -402,7 +423,7 @@ def test_headers_split_field(parser: Any) -> None:
     assert not msg.upgrade
 
 
-def test_parse_headers_multi(parser: Any) -> None:
+def test_parse_headers_multi(parser: HttpRequestParser) -> None:
     text = (
         b"GET /test HTTP/1.1\r\n"
         b"Set-Cookie: c1=cookie1\r\n"
@@ -425,64 +446,64 @@ def test_parse_headers_multi(parser: Any) -> None:
     assert msg.compression is None
 
 
-def test_conn_default_1_0(parser: Any) -> None:
+def test_conn_default_1_0(parser: HttpRequestParser) -> None:
     text = b"GET /test HTTP/1.0\r\n\r\n"
     messages, upgrade, tail = parser.feed_data(text)
     msg = messages[0][0]
     assert msg.should_close
 
 
-def test_conn_default_1_1(parser: Any) -> None:
+def test_conn_default_1_1(parser: HttpRequestParser) -> None:
     text = b"GET /test HTTP/1.1\r\n\r\n"
     messages, upgrade, tail = parser.feed_data(text)
     msg = messages[0][0]
     assert not msg.should_close
 
 
-def test_conn_close(parser: Any) -> None:
-    text = b"GET /test HTTP/1.1\r\n" b"connection: close\r\n\r\n"
+def test_conn_close(parser: HttpRequestParser) -> None:
+    text = b"GET /test HTTP/1.1\r\nconnection: close\r\n\r\n"
     messages, upgrade, tail = parser.feed_data(text)
     msg = messages[0][0]
     assert msg.should_close
 
 
-def test_conn_close_1_0(parser: Any) -> None:
-    text = b"GET /test HTTP/1.0\r\n" b"connection: close\r\n\r\n"
+def test_conn_close_1_0(parser: HttpRequestParser) -> None:
+    text = b"GET /test HTTP/1.0\r\nconnection: close\r\n\r\n"
     messages, upgrade, tail = parser.feed_data(text)
     msg = messages[0][0]
     assert msg.should_close
 
 
-def test_conn_keep_alive_1_0(parser: Any) -> None:
-    text = b"GET /test HTTP/1.0\r\n" b"connection: keep-alive\r\n\r\n"
+def test_conn_keep_alive_1_0(parser: HttpRequestParser) -> None:
+    text = b"GET /test HTTP/1.0\r\nconnection: keep-alive\r\n\r\n"
     messages, upgrade, tail = parser.feed_data(text)
     msg = messages[0][0]
     assert not msg.should_close
 
 
-def test_conn_keep_alive_1_1(parser: Any) -> None:
-    text = b"GET /test HTTP/1.1\r\n" b"connection: keep-alive\r\n\r\n"
+def test_conn_keep_alive_1_1(parser: HttpRequestParser) -> None:
+    text = b"GET /test HTTP/1.1\r\nconnection: keep-alive\r\n\r\n"
     messages, upgrade, tail = parser.feed_data(text)
     msg = messages[0][0]
     assert not msg.should_close
 
 
-def test_conn_other_1_0(parser: Any) -> None:
-    text = b"GET /test HTTP/1.0\r\n" b"connection: test\r\n\r\n"
+def test_conn_other_1_0(parser: HttpRequestParser) -> None:
+    text = b"GET /test HTTP/1.0\r\nconnection: test\r\n\r\n"
     messages, upgrade, tail = parser.feed_data(text)
     msg = messages[0][0]
     assert msg.should_close
 
 
-def test_conn_other_1_1(parser: Any) -> None:
-    text = b"GET /test HTTP/1.1\r\n" b"connection: test\r\n\r\n"
+def test_conn_other_1_1(parser: HttpRequestParser) -> None:
+    text = b"GET /test HTTP/1.1\r\nconnection: test\r\n\r\n"
     messages, upgrade, tail = parser.feed_data(text)
     msg = messages[0][0]
     assert not msg.should_close
 
 
-def test_request_chunked(parser: Any) -> None:
-    text = b"GET /test HTTP/1.1\r\n" b"transfer-encoding: chunked\r\n\r\n"
+def test_request_chunked(parser: HttpRequestParser) -> None:
+    text = b"GET /test HTTP/1.1\r\ntransfer-encoding: chunked\r\n\r\n"
     messages, upgrade, tail = parser.feed_data(text)
     msg, payload = messages[0]
     assert msg.chunked
@@ -490,7 +511,7 @@ def test_request_chunked(parser: Any) -> None:
     assert isinstance(payload, streams.StreamReader)
 
 
-def test_request_te_chunked_with_content_length(parser: Any) -> None:
+def test_request_te_chunked_with_content_length(parser: HttpRequestParser) -> None:
     text = (
         b"GET /test HTTP/1.1\r\n"
         b"content-length: 1234\r\n"
@@ -503,8 +524,8 @@ def test_request_te_chunked_with_content_length(parser: Any) -> None:
         parser.feed_data(text)
 
 
-def test_request_te_chunked123(parser: Any) -> None:
-    text = b"GET /test HTTP/1.1\r\n" b"transfer-encoding: chunked123\r\n\r\n"
+def test_request_te_chunked123(parser: HttpRequestParser) -> None:
+    text = b"GET /test HTTP/1.1\r\ntransfer-encoding: chunked123\r\n\r\n"
     with pytest.raises(
         http_exceptions.BadHttpMessage,
         match="Request has invalid `Transfer-Encoding`",
@@ -512,7 +533,24 @@ def test_request_te_chunked123(parser: Any) -> None:
         parser.feed_data(text)
 
 
-def test_conn_upgrade(parser: Any) -> None:
+async def test_request_te_last_chunked(parser: HttpRequestParser) -> None:
+    text = b"GET /test HTTP/1.1\r\nTransfer-Encoding: not, chunked\r\n\r\n1\r\nT\r\n3\r\nest\r\n0\r\n\r\n"
+    messages, upgrade, tail = parser.feed_data(text)
+    # https://www.rfc-editor.org/rfc/rfc9112#section-6.3-2.4.3
+    assert await messages[0][1].read() == b"Test"
+
+
+def test_request_te_first_chunked(parser: HttpRequestParser) -> None:
+    text = b"GET /test HTTP/1.1\r\nTransfer-Encoding: chunked, not\r\n\r\n1\r\nT\r\n3\r\nest\r\n0\r\n\r\n"
+    # https://www.rfc-editor.org/rfc/rfc9112#section-6.3-2.4.3
+    with pytest.raises(
+        http_exceptions.BadHttpMessage,
+        match="nvalid `Transfer-Encoding`",
+    ):
+        parser.feed_data(text)
+
+
+def test_conn_upgrade(parser: HttpRequestParser) -> None:
     text = (
         b"GET /test HTTP/1.1\r\n"
         b"connection: upgrade\r\n"
@@ -525,7 +563,7 @@ def test_conn_upgrade(parser: Any) -> None:
     assert upgrade
 
 
-def test_bad_upgrade(parser: Any) -> None:
+def test_bad_upgrade(parser: HttpRequestParser) -> None:
     """Test not upgraded if missing Upgrade header."""
     text = b"GET /test HTTP/1.1\r\nconnection: upgrade\r\n\r\n"
     messages, upgrade, tail = parser.feed_data(text)
@@ -534,59 +572,59 @@ def test_bad_upgrade(parser: Any) -> None:
     assert not upgrade
 
 
-def test_compression_empty(parser: Any) -> None:
-    text = b"GET /test HTTP/1.1\r\n" b"content-encoding: \r\n\r\n"
+def test_compression_empty(parser: HttpRequestParser) -> None:
+    text = b"GET /test HTTP/1.1\r\ncontent-encoding: \r\n\r\n"
     messages, upgrade, tail = parser.feed_data(text)
     msg = messages[0][0]
     assert msg.compression is None
 
 
-def test_compression_deflate(parser: Any) -> None:
-    text = b"GET /test HTTP/1.1\r\n" b"content-encoding: deflate\r\n\r\n"
+def test_compression_deflate(parser: HttpRequestParser) -> None:
+    text = b"GET /test HTTP/1.1\r\ncontent-encoding: deflate\r\n\r\n"
     messages, upgrade, tail = parser.feed_data(text)
     msg = messages[0][0]
     assert msg.compression == "deflate"
 
 
-def test_compression_gzip(parser: Any) -> None:
-    text = b"GET /test HTTP/1.1\r\n" b"content-encoding: gzip\r\n\r\n"
+def test_compression_gzip(parser: HttpRequestParser) -> None:
+    text = b"GET /test HTTP/1.1\r\ncontent-encoding: gzip\r\n\r\n"
     messages, upgrade, tail = parser.feed_data(text)
     msg = messages[0][0]
     assert msg.compression == "gzip"
 
 
 @pytest.mark.skipif(brotli is None, reason="brotli is not installed")
-def test_compression_brotli(parser: Any) -> None:
-    text = b"GET /test HTTP/1.1\r\n" b"content-encoding: br\r\n\r\n"
+def test_compression_brotli(parser: HttpRequestParser) -> None:
+    text = b"GET /test HTTP/1.1\r\ncontent-encoding: br\r\n\r\n"
     messages, upgrade, tail = parser.feed_data(text)
     msg = messages[0][0]
     assert msg.compression == "br"
 
 
-def test_compression_unknown(parser: Any) -> None:
-    text = b"GET /test HTTP/1.1\r\n" b"content-encoding: compress\r\n\r\n"
+def test_compression_unknown(parser: HttpRequestParser) -> None:
+    text = b"GET /test HTTP/1.1\r\ncontent-encoding: compress\r\n\r\n"
     messages, upgrade, tail = parser.feed_data(text)
     msg = messages[0][0]
     assert msg.compression is None
 
 
-def test_url_connect(parser: Any) -> None:
-    text = b"CONNECT www.google.com HTTP/1.1\r\n" b"content-length: 0\r\n\r\n"
+def test_url_connect(parser: HttpRequestParser) -> None:
+    text = b"CONNECT www.google.com HTTP/1.1\r\ncontent-length: 0\r\n\r\n"
     messages, upgrade, tail = parser.feed_data(text)
     msg, payload = messages[0]
     assert upgrade
     assert msg.url == URL.build(authority="www.google.com")
 
 
-def test_headers_connect(parser: Any) -> None:
-    text = b"CONNECT www.google.com HTTP/1.1\r\n" b"content-length: 0\r\n\r\n"
+def test_headers_connect(parser: HttpRequestParser) -> None:
+    text = b"CONNECT www.google.com HTTP/1.1\r\ncontent-length: 0\r\n\r\n"
     messages, upgrade, tail = parser.feed_data(text)
     msg, payload = messages[0]
     assert upgrade
     assert isinstance(payload, streams.StreamReader)
 
 
-def test_url_absolute(parser: Any) -> None:
+def test_url_absolute(parser: HttpRequestParser) -> None:
     text = (
         b"GET https://www.google.com/path/to.html HTTP/1.1\r\n"
         b"content-length: 0\r\n\r\n"
@@ -598,22 +636,22 @@ def test_url_absolute(parser: Any) -> None:
     assert msg.url == URL("https://www.google.com/path/to.html")
 
 
-def test_headers_old_websocket_key1(parser: Any) -> None:
-    text = b"GET /test HTTP/1.1\r\n" b"SEC-WEBSOCKET-KEY1: line\r\n\r\n"
+def test_headers_old_websocket_key1(parser: HttpRequestParser) -> None:
+    text = b"GET /test HTTP/1.1\r\nSEC-WEBSOCKET-KEY1: line\r\n\r\n"
 
     with pytest.raises(http_exceptions.BadHttpMessage):
         parser.feed_data(text)
 
 
-def test_headers_content_length_err_1(parser: Any) -> None:
-    text = b"GET /test HTTP/1.1\r\n" b"content-length: line\r\n\r\n"
+def test_headers_content_length_err_1(parser: HttpRequestParser) -> None:
+    text = b"GET /test HTTP/1.1\r\ncontent-length: line\r\n\r\n"
 
     with pytest.raises(http_exceptions.BadHttpMessage):
         parser.feed_data(text)
 
 
-def test_headers_content_length_err_2(parser: Any) -> None:
-    text = b"GET /test HTTP/1.1\r\n" b"content-length: -1\r\n\r\n"
+def test_headers_content_length_err_2(parser: HttpRequestParser) -> None:
+    text = b"GET /test HTTP/1.1\r\ncontent-length: -1\r\n\r\n"
 
     with pytest.raises(http_exceptions.BadHttpMessage):
         parser.feed_data(text)
@@ -636,38 +674,39 @@ _pad: Dict[bytes, str] = {
 @pytest.mark.parametrize("pad2", _pad.keys(), ids=["post-" + n for n in _pad.values()])
 @pytest.mark.parametrize("pad1", _pad.keys(), ids=["pre-" + n for n in _pad.values()])
 def test_invalid_header_spacing(
-    parser: Any, pad1: bytes, pad2: bytes, hdr: bytes
+    parser: HttpRequestParser, pad1: bytes, pad2: bytes, hdr: bytes
 ) -> None:
-    text = b"GET /test HTTP/1.1\r\n" b"%s%s%s: value\r\n\r\n" % (pad1, hdr, pad2)
-    expectation = pytest.raises(http_exceptions.BadHttpMessage)
+    text = b"GET /test HTTP/1.1\r\n%s%s%s: value\r\n\r\n" % (pad1, hdr, pad2)
     if pad1 == pad2 == b"" and hdr != b"":
         # one entry in param matrix is correct: non-empty name, not padded
-        expectation = nullcontext()
-    with expectation:
         parser.feed_data(text)
+        return
 
-
-def test_empty_header_name(parser: Any) -> None:
-    text = b"GET /test HTTP/1.1\r\n" b":test\r\n\r\n"
     with pytest.raises(http_exceptions.BadHttpMessage):
         parser.feed_data(text)
 
 
-def test_invalid_header(parser: Any) -> None:
-    text = b"GET /test HTTP/1.1\r\n" b"test line\r\n\r\n"
+def test_empty_header_name(parser: HttpRequestParser) -> None:
+    text = b"GET /test HTTP/1.1\r\n:test\r\n\r\n"
     with pytest.raises(http_exceptions.BadHttpMessage):
         parser.feed_data(text)
 
 
-def test_invalid_name(parser: Any) -> None:
-    text = b"GET /test HTTP/1.1\r\n" b"test[]: line\r\n\r\n"
+def test_invalid_header(parser: HttpRequestParser) -> None:
+    text = b"GET /test HTTP/1.1\r\ntest line\r\n\r\n"
+    with pytest.raises(http_exceptions.BadHttpMessage):
+        parser.feed_data(text)
+
+
+def test_invalid_name(parser: HttpRequestParser) -> None:
+    text = b"GET /test HTTP/1.1\r\ntest[]: line\r\n\r\n"
 
     with pytest.raises(http_exceptions.BadHttpMessage):
         parser.feed_data(text)
 
 
 @pytest.mark.parametrize("size", [40960, 8191])
-def test_max_header_field_size(parser: Any, size: Any) -> None:
+def test_max_header_field_size(parser: HttpRequestParser, size: int) -> None:
     name = b"t" * size
     text = b"GET /test HTTP/1.1\r\n" + name + b":data\r\n\r\n"
 
@@ -676,7 +715,7 @@ def test_max_header_field_size(parser: Any, size: Any) -> None:
         parser.feed_data(text)
 
 
-def test_max_header_field_size_under_limit(parser: Any) -> None:
+def test_max_header_field_size_under_limit(parser: HttpRequestParser) -> None:
     name = b"t" * 8190
     text = b"GET /test HTTP/1.1\r\n" + name + b":data\r\n\r\n"
 
@@ -695,18 +734,18 @@ def test_max_header_field_size_under_limit(parser: Any) -> None:
 
 
 @pytest.mark.parametrize("size", [40960, 8191])
-def test_max_header_value_size(parser: Any, size: Any) -> None:
+def test_max_header_value_size(parser: HttpRequestParser, size: int) -> None:
     name = b"t" * size
-    text = b"GET /test HTTP/1.1\r\n" b"data:" + name + b"\r\n\r\n"
+    text = b"GET /test HTTP/1.1\r\ndata:" + name + b"\r\n\r\n"
 
     match = f"400, message:\n  Got more than 8190 bytes \\({size}\\) when reading"
     with pytest.raises(http_exceptions.LineTooLong, match=match):
         parser.feed_data(text)
 
 
-def test_max_header_value_size_under_limit(parser: Any) -> None:
+def test_max_header_value_size_under_limit(parser: HttpRequestParser) -> None:
     value = b"A" * 8190
-    text = b"GET /test HTTP/1.1\r\n" b"data:" + value + b"\r\n\r\n"
+    text = b"GET /test HTTP/1.1\r\ndata:" + value + b"\r\n\r\n"
 
     messages, upgrade, tail = parser.feed_data(text)
     msg = messages[0][0]
@@ -723,7 +762,9 @@ def test_max_header_value_size_under_limit(parser: Any) -> None:
 
 
 @pytest.mark.parametrize("size", [40965, 8191])
-def test_max_header_value_size_continuation(response: Any, size: Any) -> None:
+def test_max_header_value_size_continuation(
+    response: HttpResponseParser, size: int
+) -> None:
     name = b"T" * (size - 5)
     text = b"HTTP/1.1 200 Ok\r\ndata: test\r\n " + name + b"\r\n\r\n"
 
@@ -732,7 +773,9 @@ def test_max_header_value_size_continuation(response: Any, size: Any) -> None:
         response.feed_data(text)
 
 
-def test_max_header_value_size_continuation_under_limit(response: Any) -> None:
+def test_max_header_value_size_continuation_under_limit(
+    response: HttpResponseParser,
+) -> None:
     value = b"A" * 8185
     text = b"HTTP/1.1 200 Ok\r\ndata: test\r\n " + value + b"\r\n\r\n"
 
@@ -749,7 +792,7 @@ def test_max_header_value_size_continuation_under_limit(response: Any) -> None:
     assert not msg.chunked
 
 
-def test_http_request_parser(parser: Any) -> None:
+def test_http_request_parser(parser: HttpRequestParser) -> None:
     text = b"GET /path HTTP/1.1\r\n\r\n"
     messages, upgrade, tail = parser.feed_data(text)
     msg = messages[0][0]
@@ -766,7 +809,7 @@ def test_http_request_parser(parser: Any) -> None:
     assert msg.url == URL("/path")
 
 
-def test_http_request_bad_status_line(parser: Any) -> None:
+def test_http_request_bad_status_line(parser: HttpRequestParser) -> None:
     text = b"getpath \r\n\r\n"
     with pytest.raises(http_exceptions.BadStatusLine) as exc_info:
         parser.feed_data(text)
@@ -787,14 +830,14 @@ _num: Dict[bytes, str] = {
 
 @pytest.mark.parametrize("nonascii_digit", _num.keys(), ids=_num.values())
 def test_http_request_bad_status_line_number(
-    parser: Any, nonascii_digit: bytes
+    parser: HttpRequestParser, nonascii_digit: bytes
 ) -> None:
     text = b"GET /digit HTTP/1." + nonascii_digit + b"\r\n\r\n"
     with pytest.raises(http_exceptions.BadStatusLine):
         parser.feed_data(text)
 
 
-def test_http_request_bad_status_line_separator(parser: Any) -> None:
+def test_http_request_bad_status_line_separator(parser: HttpRequestParser) -> None:
     # single code point, old, multibyte NFKC, multibyte NFKD
     utf8sep = "\N{arabic ligature sallallahou alayhe wasallam}".encode()
     text = b"GET /ligature HTTP/1" + utf8sep + b"1\r\n\r\n"
@@ -802,13 +845,13 @@ def test_http_request_bad_status_line_separator(parser: Any) -> None:
         parser.feed_data(text)
 
 
-def test_http_request_bad_status_line_whitespace(parser: Any) -> None:
+def test_http_request_bad_status_line_whitespace(parser: HttpRequestParser) -> None:
     text = b"GET\n/path\fHTTP/1.1\r\n\r\n"
     with pytest.raises(http_exceptions.BadStatusLine):
         parser.feed_data(text)
 
 
-def test_http_request_upgrade(parser: Any) -> None:
+def test_http_request_upgrade(parser: HttpRequestParser) -> None:
     text = (
         b"GET /test HTTP/1.1\r\n"
         b"connection: upgrade\r\n"
@@ -823,8 +866,28 @@ def test_http_request_upgrade(parser: Any) -> None:
     assert tail == b"some raw data"
 
 
+async def test_http_request_upgrade_unknown(parser: HttpRequestParser) -> None:
+    text = (
+        b"POST / HTTP/1.1\r\n"
+        b"Connection: Upgrade\r\n"
+        b"Content-Length: 2\r\n"
+        b"Upgrade: unknown\r\n"
+        b"Content-Type: application/json\r\n\r\n"
+        b"{}"
+    )
+    messages, upgrade, tail = parser.feed_data(text)
+
+    msg = messages[0][0]
+    assert not msg.should_close
+    assert msg.upgrade
+    assert not upgrade
+    assert not msg.chunked
+    assert tail == b""
+    assert await messages[0][-1].read() == b"{}"
+
+
 @pytest.fixture
-def xfail_c_parser_url(request) -> None:
+def xfail_c_parser_url(request: pytest.FixtureRequest) -> None:
     if isinstance(request.getfixturevalue("parser"), HttpRequestParserPy):
         return
     request.node.add_marker(
@@ -836,7 +899,7 @@ def xfail_c_parser_url(request) -> None:
 
 
 @pytest.mark.usefixtures("xfail_c_parser_url")
-def test_http_request_parser_utf8_request_line(parser: Any) -> None:
+def test_http_request_parser_utf8_request_line(parser: HttpRequestParser) -> None:
     messages, upgrade, tail = parser.feed_data(
         # note the truncated unicode sequence
         b"GET /P\xc3\xbcnktchen\xa0\xef\xb7 HTTP/1.1\r\n" +
@@ -861,7 +924,7 @@ def test_http_request_parser_utf8_request_line(parser: Any) -> None:
     assert msg.url == URL.build(path="/Pünktchen\udca0\udcef\udcb7", encoded=True)
 
 
-def test_http_request_parser_utf8(parser: Any) -> None:
+def test_http_request_parser_utf8(parser: HttpRequestParser) -> None:
     text = "GET /path HTTP/1.1\r\nx-test:тест\r\n\r\n".encode()
     messages, upgrade, tail = parser.feed_data(text)
     msg = messages[0][0]
@@ -878,7 +941,7 @@ def test_http_request_parser_utf8(parser: Any) -> None:
     assert msg.url == URL("/path")
 
 
-def test_http_request_parser_non_utf8(parser: Any) -> None:
+def test_http_request_parser_non_utf8(parser: HttpRequestParser) -> None:
     text = "GET /path HTTP/1.1\r\nx-test:тест\r\n\r\n".encode("cp1251")
     msg = parser.feed_data(text)[0][0][0]
 
@@ -896,7 +959,7 @@ def test_http_request_parser_non_utf8(parser: Any) -> None:
     assert msg.url == URL("/path")
 
 
-def test_http_request_parser_two_slashes(parser: Any) -> None:
+def test_http_request_parser_two_slashes(parser: HttpRequestParser) -> None:
     text = b"GET //path HTTP/1.1\r\n\r\n"
     msg = parser.feed_data(text)[0][0][0]
 
@@ -915,41 +978,41 @@ def test_http_request_parser_two_slashes(parser: Any) -> None:
     [bytes([i]) for i in rb'"(),/:;<=>?@[\]{}'],
 )
 def test_http_request_parser_bad_method(
-    parser: Any, rfc9110_5_6_2_token_delim: bytes
+    parser: HttpRequestParser, rfc9110_5_6_2_token_delim: bytes
 ) -> None:
     with pytest.raises(http_exceptions.BadStatusLine):
         parser.feed_data(rfc9110_5_6_2_token_delim + b'ET" /get HTTP/1.1\r\n\r\n')
 
 
-def test_http_request_parser_bad_version(parser: Any) -> None:
+def test_http_request_parser_bad_version(parser: HttpRequestParser) -> None:
     with pytest.raises(http_exceptions.BadHttpMessage):
         parser.feed_data(b"GET //get HT/11\r\n\r\n")
 
 
-def test_http_request_parser_bad_version_number(parser: Any) -> None:
+def test_http_request_parser_bad_version_number(parser: HttpRequestParser) -> None:
     with pytest.raises(http_exceptions.BadHttpMessage):
         parser.feed_data(b"GET /test HTTP/1.32\r\n\r\n")
 
 
-def test_http_request_parser_bad_ascii_uri(parser: Any) -> None:
+def test_http_request_parser_bad_ascii_uri(parser: HttpRequestParser) -> None:
     with pytest.raises(http_exceptions.InvalidURLError):
         parser.feed_data(b"GET ! HTTP/1.1\r\n\r\n")
 
 
-def test_http_request_parser_bad_nonascii_uri(parser: Any) -> None:
+def test_http_request_parser_bad_nonascii_uri(parser: HttpRequestParser) -> None:
     with pytest.raises(http_exceptions.InvalidURLError):
         parser.feed_data(b"GET \xff HTTP/1.1\r\n\r\n")
 
 
 @pytest.mark.parametrize("size", [40965, 8191])
-def test_http_request_max_status_line(parser: Any, size: Any) -> None:
+def test_http_request_max_status_line(parser: HttpRequestParser, size: int) -> None:
     path = b"t" * (size - 5)
     match = f"400, message:\n  Got more than 8190 bytes \\({size}\\) when reading"
     with pytest.raises(http_exceptions.LineTooLong, match=match):
         parser.feed_data(b"GET /path" + path + b" HTTP/1.1\r\n\r\n")
 
 
-def test_http_request_max_status_line_under_limit(parser: Any) -> None:
+def test_http_request_max_status_line_under_limit(parser: HttpRequestParser) -> None:
     path = b"t" * (8190 - 5)
     messages, upgraded, tail = parser.feed_data(
         b"GET /path" + path + b" HTTP/1.1\r\n\r\n"
@@ -968,7 +1031,7 @@ def test_http_request_max_status_line_under_limit(parser: Any) -> None:
     assert msg.url == URL("/path" + path.decode())
 
 
-def test_http_response_parser_utf8(response: Any) -> None:
+def test_http_response_parser_utf8(response: HttpResponseParser) -> None:
     text = "HTTP/1.1 200 Ok\r\nx-test:тест\r\n\r\n".encode()
 
     messages, upgraded, tail = response.feed_data(text)
@@ -984,7 +1047,7 @@ def test_http_response_parser_utf8(response: Any) -> None:
     assert not tail
 
 
-def test_http_response_parser_utf8_without_reason(response: Any) -> None:
+def test_http_response_parser_utf8_without_reason(response: HttpResponseParser) -> None:
     text = "HTTP/1.1 200 \r\nx-test:тест\r\n\r\n".encode()
 
     messages, upgraded, tail = response.feed_data(text)
@@ -1000,7 +1063,7 @@ def test_http_response_parser_utf8_without_reason(response: Any) -> None:
     assert not tail
 
 
-def test_http_response_parser_obs_line_folding(response: Any) -> None:
+def test_http_response_parser_obs_line_folding(response: HttpResponseParser) -> None:
     text = b"HTTP/1.1 200 Ok\r\ntest: line\r\n continue\r\n\r\n"
 
     messages, upgraded, tail = response.feed_data(text)
@@ -1017,7 +1080,9 @@ def test_http_response_parser_obs_line_folding(response: Any) -> None:
 
 
 @pytest.mark.dev_mode
-def test_http_response_parser_strict_obs_line_folding(response: Any) -> None:
+def test_http_response_parser_strict_obs_line_folding(
+    response: HttpResponseParser,
+) -> None:
     text = b"HTTP/1.1 200 Ok\r\ntest: line\r\n continue\r\n\r\n"
 
     with pytest.raises(http_exceptions.BadHttpMessage):
@@ -1026,7 +1091,7 @@ def test_http_response_parser_strict_obs_line_folding(response: Any) -> None:
 
 @pytest.mark.parametrize("size", [40962, 8191])
 def test_http_response_parser_bad_status_line_too_long(
-    response: Any, size: Any
+    response: HttpResponseParser, size: int
 ) -> None:
     reason = b"t" * (size - 2)
     match = f"400, message:\n  Got more than 8190 bytes \\({size}\\) when reading"
@@ -1034,7 +1099,9 @@ def test_http_response_parser_bad_status_line_too_long(
         response.feed_data(b"HTTP/1.1 200 Ok" + reason + b"\r\n\r\n")
 
 
-def test_http_response_parser_status_line_under_limit(response: Any) -> None:
+def test_http_response_parser_status_line_under_limit(
+    response: HttpResponseParser,
+) -> None:
     reason = b"O" * 8190
     messages, upgraded, tail = response.feed_data(
         b"HTTP/1.1 200 " + reason + b"\r\n\r\n"
@@ -1045,17 +1112,17 @@ def test_http_response_parser_status_line_under_limit(response: Any) -> None:
     assert msg.reason == reason.decode()
 
 
-def test_http_response_parser_bad_version(response: Any) -> None:
+def test_http_response_parser_bad_version(response: HttpResponseParser) -> None:
     with pytest.raises(http_exceptions.BadHttpMessage):
         response.feed_data(b"HT/11 200 Ok\r\n\r\n")
 
 
-def test_http_response_parser_bad_version_number(response: Any) -> None:
+def test_http_response_parser_bad_version_number(response: HttpResponseParser) -> None:
     with pytest.raises(http_exceptions.BadHttpMessage):
         response.feed_data(b"HTTP/12.3 200 Ok\r\n\r\n")
 
 
-def test_http_response_parser_no_reason(response: Any) -> None:
+def test_http_response_parser_no_reason(response: HttpResponseParser) -> None:
     msg = response.feed_data(b"HTTP/1.1 200\r\n\r\n")[0][0][0]
 
     assert msg.version == (1, 1)
@@ -1063,7 +1130,7 @@ def test_http_response_parser_no_reason(response: Any) -> None:
     assert msg.reason == ""
 
 
-def test_http_response_parser_lenient_headers(response: Any) -> None:
+def test_http_response_parser_lenient_headers(response: HttpResponseParser) -> None:
     messages, upgrade, tail = response.feed_data(
         b"HTTP/1.1 200 test\r\nFoo: abc\x01def\r\n\r\n"
     )
@@ -1073,14 +1140,14 @@ def test_http_response_parser_lenient_headers(response: Any) -> None:
 
 
 @pytest.mark.dev_mode
-def test_http_response_parser_strict_headers(response: Any) -> None:
+def test_http_response_parser_strict_headers(response: HttpResponseParser) -> None:
     if isinstance(response, HttpResponseParserPy):
         pytest.xfail("Py parser is lenient. May update py-parser later.")
     with pytest.raises(http_exceptions.BadHttpMessage):
         response.feed_data(b"HTTP/1.1 200 test\r\nFoo: abc\x01def\r\n\r\n")
 
 
-def test_http_response_parser_bad_crlf(response: Any) -> None:
+def test_http_response_parser_bad_crlf(response: HttpResponseParser) -> None:
     """Still a lot of dodgy servers sending bad requests like this."""
     messages, upgrade, tail = response.feed_data(
         b"HTTP/1.0 200 OK\nFoo: abc\nBar: def\n\nBODY\n"
@@ -1091,7 +1158,9 @@ def test_http_response_parser_bad_crlf(response: Any) -> None:
     assert msg.headers["Bar"] == "def"
 
 
-async def test_http_response_parser_bad_chunked_lax(response: Any) -> None:
+async def test_http_response_parser_bad_chunked_lax(
+    response: HttpResponseParser,
+) -> None:
     text = (
         b"HTTP/1.1 200 OK\r\nTransfer-Encoding: chunked\r\n\r\n5 \r\nabcde\r\n0\r\n\r\n"
     )
@@ -1102,7 +1171,7 @@ async def test_http_response_parser_bad_chunked_lax(response: Any) -> None:
 
 @pytest.mark.dev_mode
 async def test_http_response_parser_bad_chunked_strict_py(
-    loop: Any, protocol: Any
+    loop: asyncio.AbstractEventLoop, protocol: BaseProtocol
 ) -> None:
     response = HttpResponseParserPy(
         protocol,
@@ -1124,7 +1193,7 @@ async def test_http_response_parser_bad_chunked_strict_py(
     reason="C based HTTP parser not available",
 )
 async def test_http_response_parser_bad_chunked_strict_c(
-    loop: Any, protocol: Any
+    loop: asyncio.AbstractEventLoop, protocol: BaseProtocol
 ) -> None:
     response = HttpResponseParserC(
         protocol,
@@ -1140,36 +1209,57 @@ async def test_http_response_parser_bad_chunked_strict_c(
         response.feed_data(text)
 
 
-def test_http_response_parser_bad(response: Any) -> None:
+async def test_http_response_parser_notchunked(
+    response: HttpResponseParser,
+) -> None:
+    text = b"HTTP/1.1 200 OK\r\nTransfer-Encoding: notchunked\r\n\r\n1\r\nT\r\n3\r\nest\r\n0\r\n\r\n"
+    messages, upgrade, tail = response.feed_data(text)
+    response.feed_eof()
+
+    # https://www.rfc-editor.org/rfc/rfc9112#section-6.3-2.4.2
+    assert await messages[0][1].read() == b"1\r\nT\r\n3\r\nest\r\n0\r\n\r\n"
+
+
+async def test_http_response_parser_last_chunked(
+    response: HttpResponseParser,
+) -> None:
+    text = b"HTTP/1.1 200 OK\r\nTransfer-Encoding: not, chunked\r\n\r\n1\r\nT\r\n3\r\nest\r\n0\r\n\r\n"
+    messages, upgrade, tail = response.feed_data(text)
+
+    # https://www.rfc-editor.org/rfc/rfc9112#section-6.3-2.4.2
+    assert await messages[0][1].read() == b"Test"
+
+
+def test_http_response_parser_bad(response: HttpResponseParser) -> None:
     with pytest.raises(http_exceptions.BadHttpMessage):
         response.feed_data(b"HTT/1\r\n\r\n")
 
 
-def test_http_response_parser_code_under_100(response: Any) -> None:
+def test_http_response_parser_code_under_100(response: HttpResponseParser) -> None:
     with pytest.raises(http_exceptions.BadStatusLine):
         response.feed_data(b"HTTP/1.1 99 test\r\n\r\n")
 
 
-def test_http_response_parser_code_above_999(response: Any) -> None:
+def test_http_response_parser_code_above_999(response: HttpResponseParser) -> None:
     with pytest.raises(http_exceptions.BadStatusLine):
         response.feed_data(b"HTTP/1.1 9999 test\r\n\r\n")
 
 
-def test_http_response_parser_code_not_int(response: Any) -> None:
+def test_http_response_parser_code_not_int(response: HttpResponseParser) -> None:
     with pytest.raises(http_exceptions.BadStatusLine):
         response.feed_data(b"HTTP/1.1 ttt test\r\n\r\n")
 
 
 @pytest.mark.parametrize("nonascii_digit", _num.keys(), ids=_num.values())
 def test_http_response_parser_code_not_ascii(
-    response: Any, nonascii_digit: bytes
+    response: HttpResponseParser, nonascii_digit: bytes
 ) -> None:
     with pytest.raises(http_exceptions.BadStatusLine):
         response.feed_data(b"HTTP/1.1 20" + nonascii_digit + b" test\r\n\r\n")
 
 
-def test_http_request_chunked_payload(parser: Any) -> None:
-    text = b"GET /test HTTP/1.1\r\n" b"transfer-encoding: chunked\r\n\r\n"
+def test_http_request_chunked_payload(parser: HttpRequestParser) -> None:
+    text = b"GET /test HTTP/1.1\r\ntransfer-encoding: chunked\r\n\r\n"
     msg, payload = parser.feed_data(text)[0][0]
 
     assert msg.chunked
@@ -1183,8 +1273,10 @@ def test_http_request_chunked_payload(parser: Any) -> None:
     assert payload.is_eof()
 
 
-def test_http_request_chunked_payload_and_next_message(parser: Any) -> None:
-    text = b"GET /test HTTP/1.1\r\n" b"transfer-encoding: chunked\r\n\r\n"
+def test_http_request_chunked_payload_and_next_message(
+    parser: HttpRequestParser,
+) -> None:
+    text = b"GET /test HTTP/1.1\r\ntransfer-encoding: chunked\r\n\r\n"
     msg, payload = parser.feed_data(text)[0][0]
 
     messages, upgraded, tail = parser.feed_data(
@@ -1205,8 +1297,8 @@ def test_http_request_chunked_payload_and_next_message(parser: Any) -> None:
     assert not payload2.is_eof()
 
 
-def test_http_request_chunked_payload_chunks(parser: Any) -> None:
-    text = b"GET /test HTTP/1.1\r\n" b"transfer-encoding: chunked\r\n\r\n"
+def test_http_request_chunked_payload_chunks(parser: HttpRequestParser) -> None:
+    text = b"GET /test HTTP/1.1\r\ntransfer-encoding: chunked\r\n\r\n"
     msg, payload = parser.feed_data(text)[0][0]
 
     parser.feed_data(b"4\r\ndata\r")
@@ -1227,8 +1319,8 @@ def test_http_request_chunked_payload_chunks(parser: Any) -> None:
     assert payload.is_eof()
 
 
-def test_parse_chunked_payload_chunk_extension(parser: Any) -> None:
-    text = b"GET /test HTTP/1.1\r\n" b"transfer-encoding: chunked\r\n\r\n"
+def test_parse_chunked_payload_chunk_extension(parser: HttpRequestParser) -> None:
+    text = b"GET /test HTTP/1.1\r\ntransfer-encoding: chunked\r\n\r\n"
     msg, payload = parser.feed_data(text)[0][0]
 
     parser.feed_data(b"4;test\r\ndata\r\n4\r\nline\r\n0\r\ntest: test\r\n\r\n")
@@ -1238,7 +1330,11 @@ def test_parse_chunked_payload_chunk_extension(parser: Any) -> None:
     assert payload.is_eof()
 
 
-def test_parse_no_length_or_te_on_post(loop: Any, protocol: Any, request_cls: Any):
+def test_parse_no_length_or_te_on_post(
+    loop: asyncio.AbstractEventLoop,
+    protocol: BaseProtocol,
+    request_cls: Type[HttpRequestParser],
+) -> None:
     parser = request_cls(protocol, loop, limit=2**16)
     text = b"POST /test HTTP/1.1\r\n\r\n"
     msg, payload = parser.feed_data(text)[0][0]
@@ -1247,17 +1343,19 @@ def test_parse_no_length_or_te_on_post(loop: Any, protocol: Any, request_cls: An
 
 
 def test_parse_payload_response_without_body(
-    loop: Any, protocol: Any, response_cls: Any
+    loop: asyncio.AbstractEventLoop,
+    protocol: BaseProtocol,
+    response_cls: Type[HttpResponseParser],
 ) -> None:
     parser = response_cls(protocol, loop, 2**16, response_with_body=False)
-    text = b"HTTP/1.1 200 Ok\r\n" b"content-length: 10\r\n\r\n"
+    text = b"HTTP/1.1 200 Ok\r\ncontent-length: 10\r\n\r\n"
     msg, payload = parser.feed_data(text)[0][0]
 
     assert payload.is_eof()
 
 
-def test_parse_length_payload(response: Any) -> None:
-    text = b"HTTP/1.1 200 Ok\r\n" b"content-length: 4\r\n\r\n"
+def test_parse_length_payload(response: HttpResponseParser) -> None:
+    text = b"HTTP/1.1 200 Ok\r\ncontent-length: 4\r\n\r\n"
     msg, payload = response.feed_data(text)[0][0]
     assert not payload.is_eof()
 
@@ -1269,13 +1367,13 @@ def test_parse_length_payload(response: Any) -> None:
     assert b"data" == b"".join(d for d in payload._buffer)
 
 
-def test_parse_no_length_payload(parser: Any) -> None:
+def test_parse_no_length_payload(parser: HttpRequestParser) -> None:
     text = b"PUT / HTTP/1.1\r\n\r\n"
     msg, payload = parser.feed_data(text)[0][0]
     assert payload.is_eof()
 
 
-def test_parse_content_length_payload_multiple(response: Any) -> None:
+def test_parse_content_length_payload_multiple(response: HttpResponseParser) -> None:
     text = b"HTTP/1.1 200 OK\r\ncontent-length: 5\r\n\r\nfirst"
     msg, payload = response.feed_data(text)[0][0]
     assert msg.version == HttpVersion(major=1, minor=1)
@@ -1313,7 +1411,9 @@ def test_parse_content_length_payload_multiple(response: Any) -> None:
     assert b"second" == b"".join(d for d in payload._buffer)
 
 
-def test_parse_content_length_than_chunked_payload(response: Any) -> None:
+def test_parse_content_length_than_chunked_payload(
+    response: HttpResponseParser,
+) -> None:
     text = b"HTTP/1.1 200 OK\r\ncontent-length: 5\r\n\r\nfirst"
     msg, payload = response.feed_data(text)[0][0]
     assert msg.version == HttpVersion(major=1, minor=1)
@@ -1357,7 +1457,7 @@ def test_parse_content_length_than_chunked_payload(response: Any) -> None:
 
 @pytest.mark.parametrize("code", (204, 304, 101, 102))
 def test_parse_chunked_payload_empty_body_than_another_chunked(
-    response: Any, code: int
+    response: HttpResponseParser, code: int
 ) -> None:
     head = f"HTTP/1.1 {code} OK\r\n".encode()
     text = head + b"transfer-encoding: chunked\r\n\r\n"
@@ -1400,7 +1500,30 @@ def test_parse_chunked_payload_empty_body_than_another_chunked(
     assert b"second" == b"".join(d for d in payload._buffer)
 
 
-def test_partial_url(parser: Any) -> None:
+async def test_parse_chunked_payload_split_chunks(response: HttpResponseParser) -> None:
+    network_chunks = (
+        b"HTTP/1.1 200 OK\r\nTransfer-Encoding: chunked\r\n\r\n",
+        b"5\r\nfi",
+        b"rst",
+        # This simulates a bug in lax mode caused when the \r\n separator, before the
+        # next HTTP chunk, appears at the start of the next network chunk.
+        b"\r\n",
+        b"6",
+        b"\r",
+        b"\n",
+        b"second\r",
+        b"\n0\r\n\r\n",
+    )
+    reader = response.feed_data(network_chunks[0])[0][0][1]
+    for c in network_chunks[1:]:
+        response.feed_data(c)
+
+    assert response.feed_eof() is None
+    assert reader.is_eof()
+    assert await reader.read() == b"firstsecond"
+
+
+def test_partial_url(parser: HttpRequestParser) -> None:
     messages, upgrade, tail = parser.feed_data(b"GET /te")
     assert len(messages) == 0
     messages, upgrade, tail = parser.feed_data(b"st HTTP/1.1\r\n\r\n")
@@ -1426,7 +1549,7 @@ def test_partial_url(parser: Any) -> None:
     ],
 )
 def test_parse_uri_percent_encoded(
-    parser: Any, uri: Any, path: Any, query: Any, fragment: Any
+    parser: HttpRequestParser, uri: str, path: str, query: Dict[str, str], fragment: str
 ) -> None:
     text = (f"GET {uri} HTTP/1.1\r\n\r\n").encode()
     messages, upgrade, tail = parser.feed_data(text)
@@ -1439,7 +1562,7 @@ def test_parse_uri_percent_encoded(
     assert msg.url.fragment == fragment
 
 
-def test_parse_uri_utf8(parser: Any) -> None:
+def test_parse_uri_utf8(parser: HttpRequestParser) -> None:
     if not isinstance(parser, HttpRequestParserPy):
         pytest.xfail("Not valid HTTP. Maybe update py-parser to reject later.")
     text = ("GET /путь?ключ=знач#фраг HTTP/1.1\r\n\r\n").encode()
@@ -1452,7 +1575,7 @@ def test_parse_uri_utf8(parser: Any) -> None:
     assert msg.url.fragment == "фраг"
 
 
-def test_parse_uri_utf8_percent_encoded(parser: Any) -> None:
+def test_parse_uri_utf8_percent_encoded(parser: HttpRequestParser) -> None:
     text = (
         "GET %s HTTP/1.1\r\n\r\n" % quote("/путь?ключ=знач#фраг", safe="/?=#")
     ).encode()
@@ -1470,7 +1593,9 @@ def test_parse_uri_utf8_percent_encoded(parser: Any) -> None:
     "HttpRequestParserC" not in dir(aiohttp.http_parser),
     reason="C based HTTP parser not available",
 )
-def test_parse_bad_method_for_c_parser_raises(loop: Any, protocol: Any) -> None:
+def test_parse_bad_method_for_c_parser_raises(
+    loop: asyncio.AbstractEventLoop, protocol: BaseProtocol
+) -> None:
     payload = b"GET1 /test HTTP/1.1\r\n\r\n"
     parser = HttpRequestParserC(
         protocol,
@@ -1485,8 +1610,8 @@ def test_parse_bad_method_for_c_parser_raises(loop: Any, protocol: Any) -> None:
 
 
 class TestParsePayload:
-    async def test_parse_eof_payload(self, stream: Any) -> None:
-        out = aiohttp.FlowControlDataQueue(stream, 2**16, loop=asyncio.get_event_loop())
+    async def test_parse_eof_payload(self, protocol: BaseProtocol) -> None:
+        out = aiohttp.StreamReader(protocol, 2**16, loop=asyncio.get_running_loop())
         p = HttpPayloadParser(out)
         p.feed_data(b"data")
         p.feed_eof()
@@ -1494,8 +1619,8 @@ class TestParsePayload:
         assert out.is_eof()
         assert [(bytearray(b"data"))] == list(out._buffer)
 
-    async def test_parse_length_payload_eof(self, stream: Any) -> None:
-        out = aiohttp.FlowControlDataQueue(stream, 2**16, loop=asyncio.get_event_loop())
+    async def test_parse_length_payload_eof(self, protocol: BaseProtocol) -> None:
+        out = aiohttp.StreamReader(protocol, 2**16, loop=asyncio.get_running_loop())
 
         p = HttpPayloadParser(out, length=4)
         p.feed_data(b"da")
@@ -1503,15 +1628,19 @@ class TestParsePayload:
         with pytest.raises(http_exceptions.ContentLengthError):
             p.feed_eof()
 
-    async def test_parse_chunked_payload_size_error(self, stream: Any) -> None:
-        out = aiohttp.FlowControlDataQueue(stream, 2**16, loop=asyncio.get_event_loop())
+    async def test_parse_chunked_payload_size_error(
+        self, protocol: BaseProtocol
+    ) -> None:
+        out = aiohttp.StreamReader(protocol, 2**16, loop=asyncio.get_running_loop())
         p = HttpPayloadParser(out, chunked=True)
         with pytest.raises(http_exceptions.TransferEncodingError):
             p.feed_data(b"blah\r\n")
         assert isinstance(out.exception(), http_exceptions.TransferEncodingError)
 
-    async def test_parse_chunked_payload_split_end(self, protocol: Any) -> None:
-        out = aiohttp.StreamReader(protocol, 2**16, loop=None)
+    async def test_parse_chunked_payload_split_end(
+        self, protocol: BaseProtocol
+    ) -> None:
+        out = aiohttp.StreamReader(protocol, 2**16, loop=asyncio.get_running_loop())
         p = HttpPayloadParser(out, chunked=True)
         p.feed_data(b"4\r\nasdf\r\n0\r\n")
         p.feed_data(b"\r\n")
@@ -1519,8 +1648,10 @@ class TestParsePayload:
         assert out.is_eof()
         assert b"asdf" == b"".join(out._buffer)
 
-    async def test_parse_chunked_payload_split_end2(self, protocol: Any) -> None:
-        out = aiohttp.StreamReader(protocol, 2**16, loop=None)
+    async def test_parse_chunked_payload_split_end2(
+        self, protocol: BaseProtocol
+    ) -> None:
+        out = aiohttp.StreamReader(protocol, 2**16, loop=asyncio.get_running_loop())
         p = HttpPayloadParser(out, chunked=True)
         p.feed_data(b"4\r\nasdf\r\n0\r\n\r")
         p.feed_data(b"\n")
@@ -1529,9 +1660,9 @@ class TestParsePayload:
         assert b"asdf" == b"".join(out._buffer)
 
     async def test_parse_chunked_payload_split_end_trailers(
-        self, protocol: Any
+        self, protocol: BaseProtocol
     ) -> None:
-        out = aiohttp.StreamReader(protocol, 2**16, loop=None)
+        out = aiohttp.StreamReader(protocol, 2**16, loop=asyncio.get_running_loop())
         p = HttpPayloadParser(out, chunked=True)
         p.feed_data(b"4\r\nasdf\r\n0\r\n")
         p.feed_data(b"Content-MD5: 912ec803b2ce49e4a541068d495ab570\r\n")
@@ -1541,9 +1672,9 @@ class TestParsePayload:
         assert b"asdf" == b"".join(out._buffer)
 
     async def test_parse_chunked_payload_split_end_trailers2(
-        self, protocol: Any
+        self, protocol: BaseProtocol
     ) -> None:
-        out = aiohttp.StreamReader(protocol, 2**16, loop=None)
+        out = aiohttp.StreamReader(protocol, 2**16, loop=asyncio.get_running_loop())
         p = HttpPayloadParser(out, chunked=True)
         p.feed_data(b"4\r\nasdf\r\n0\r\n")
         p.feed_data(b"Content-MD5: 912ec803b2ce49e4a541068d495ab570\r\n\r")
@@ -1553,9 +1684,9 @@ class TestParsePayload:
         assert b"asdf" == b"".join(out._buffer)
 
     async def test_parse_chunked_payload_split_end_trailers3(
-        self, protocol: Any
+        self, protocol: BaseProtocol
     ) -> None:
-        out = aiohttp.StreamReader(protocol, 2**16, loop=None)
+        out = aiohttp.StreamReader(protocol, 2**16, loop=asyncio.get_running_loop())
         p = HttpPayloadParser(out, chunked=True)
         p.feed_data(b"4\r\nasdf\r\n0\r\nContent-MD5: ")
         p.feed_data(b"912ec803b2ce49e4a541068d495ab570\r\n\r\n")
@@ -1564,18 +1695,18 @@ class TestParsePayload:
         assert b"asdf" == b"".join(out._buffer)
 
     async def test_parse_chunked_payload_split_end_trailers4(
-        self, protocol: Any
+        self, protocol: BaseProtocol
     ) -> None:
-        out = aiohttp.StreamReader(protocol, 2**16, loop=None)
+        out = aiohttp.StreamReader(protocol, 2**16, loop=asyncio.get_running_loop())
         p = HttpPayloadParser(out, chunked=True)
-        p.feed_data(b"4\r\nasdf\r\n0\r\n" b"C")
+        p.feed_data(b"4\r\nasdf\r\n0\r\nC")
         p.feed_data(b"ontent-MD5: 912ec803b2ce49e4a541068d495ab570\r\n\r\n")
 
         assert out.is_eof()
         assert b"asdf" == b"".join(out._buffer)
 
-    async def test_http_payload_parser_length(self, stream: Any) -> None:
-        out = aiohttp.FlowControlDataQueue(stream, 2**16, loop=asyncio.get_event_loop())
+    async def test_http_payload_parser_length(self, protocol: BaseProtocol) -> None:
+        out = aiohttp.StreamReader(protocol, 2**16, loop=asyncio.get_running_loop())
         p = HttpPayloadParser(out, length=2)
         eof, tail = p.feed_data(b"1245")
         assert eof
@@ -1583,43 +1714,49 @@ class TestParsePayload:
         assert b"12" == out._buffer[0]
         assert b"45" == tail
 
-    async def test_http_payload_parser_deflate(self, stream: Any) -> None:
+    async def test_http_payload_parser_deflate(self, protocol: BaseProtocol) -> None:
         # c=compressobj(wbits=15); b''.join([c.compress(b'data'), c.flush()])
         COMPRESSED = b"x\x9cKI,I\x04\x00\x04\x00\x01\x9b"
 
         length = len(COMPRESSED)
-        out = aiohttp.FlowControlDataQueue(stream, 2**16, loop=asyncio.get_event_loop())
+        out = aiohttp.StreamReader(protocol, 2**16, loop=asyncio.get_running_loop())
         p = HttpPayloadParser(out, length=length, compression="deflate")
         p.feed_data(COMPRESSED)
         assert b"data" == out._buffer[0]
         assert out.is_eof()
 
-    async def test_http_payload_parser_deflate_no_hdrs(self, stream: Any) -> None:
+    async def test_http_payload_parser_deflate_no_hdrs(
+        self, protocol: BaseProtocol
+    ) -> None:
         """Tests incorrectly formed data (no zlib headers)."""
         # c=compressobj(wbits=-15); b''.join([c.compress(b'data'), c.flush()])
         COMPRESSED = b"KI,I\x04\x00"
 
         length = len(COMPRESSED)
-        out = aiohttp.FlowControlDataQueue(stream, 2**16, loop=asyncio.get_event_loop())
+        out = aiohttp.StreamReader(protocol, 2**16, loop=asyncio.get_running_loop())
         p = HttpPayloadParser(out, length=length, compression="deflate")
         p.feed_data(COMPRESSED)
         assert b"data" == out._buffer[0]
         assert out.is_eof()
 
-    async def test_http_payload_parser_deflate_light(self, stream: Any) -> None:
+    async def test_http_payload_parser_deflate_light(
+        self, protocol: BaseProtocol
+    ) -> None:
         # c=compressobj(wbits=9); b''.join([c.compress(b'data'), c.flush()])
         COMPRESSED = b"\x18\x95KI,I\x04\x00\x04\x00\x01\x9b"
 
         length = len(COMPRESSED)
-        out = aiohttp.FlowControlDataQueue(stream, 2**16, loop=asyncio.get_event_loop())
+        out = aiohttp.StreamReader(protocol, 2**16, loop=asyncio.get_running_loop())
         p = HttpPayloadParser(out, length=length, compression="deflate")
         p.feed_data(COMPRESSED)
 
         assert b"data" == out._buffer[0]
         assert out.is_eof()
 
-    async def test_http_payload_parser_deflate_split(self, stream: Any) -> None:
-        out = aiohttp.FlowControlDataQueue(stream, 2**16, loop=asyncio.get_event_loop())
+    async def test_http_payload_parser_deflate_split(
+        self, protocol: BaseProtocol
+    ) -> None:
+        out = aiohttp.StreamReader(protocol, 2**16, loop=asyncio.get_running_loop())
         p = HttpPayloadParser(out, compression="deflate")
         # Feeding one correct byte should be enough to choose exact
         # deflate decompressor
@@ -1628,8 +1765,10 @@ class TestParsePayload:
         p.feed_eof()
         assert b"data" == out._buffer[0]
 
-    async def test_http_payload_parser_deflate_split_err(self, stream: Any) -> None:
-        out = aiohttp.FlowControlDataQueue(stream, 2**16, loop=asyncio.get_event_loop())
+    async def test_http_payload_parser_deflate_split_err(
+        self, protocol: BaseProtocol
+    ) -> None:
+        out = aiohttp.StreamReader(protocol, 2**16, loop=asyncio.get_running_loop())
         p = HttpPayloadParser(out, compression="deflate")
         # Feeding one wrong byte should be enough to choose exact
         # deflate decompressor
@@ -1638,16 +1777,18 @@ class TestParsePayload:
         p.feed_eof()
         assert b"data" == out._buffer[0]
 
-    async def test_http_payload_parser_length_zero(self, stream: Any) -> None:
-        out = aiohttp.FlowControlDataQueue(stream, 2**16, loop=asyncio.get_event_loop())
+    async def test_http_payload_parser_length_zero(
+        self, protocol: BaseProtocol
+    ) -> None:
+        out = aiohttp.StreamReader(protocol, 2**16, loop=asyncio.get_running_loop())
         p = HttpPayloadParser(out, length=0)
         assert p.done
         assert out.is_eof()
 
     @pytest.mark.skipif(brotli is None, reason="brotli is not installed")
-    async def test_http_payload_brotli(self, stream: Any) -> None:
+    async def test_http_payload_brotli(self, protocol: BaseProtocol) -> None:
         compressed = brotli.compress(b"brotli data")
-        out = aiohttp.FlowControlDataQueue(stream, 2**16, loop=asyncio.get_event_loop())
+        out = aiohttp.StreamReader(protocol, 2**16, loop=asyncio.get_running_loop())
         p = HttpPayloadParser(out, length=len(compressed), compression="br")
         p.feed_data(compressed)
         assert b"brotli data" == out._buffer[0]
@@ -1655,8 +1796,8 @@ class TestParsePayload:
 
 
 class TestDeflateBuffer:
-    async def test_feed_data(self, stream: Any) -> None:
-        buf = aiohttp.FlowControlDataQueue(stream, 2**16, loop=asyncio.get_event_loop())
+    async def test_feed_data(self, protocol: BaseProtocol) -> None:
+        buf = aiohttp.StreamReader(protocol, 2**16, loop=asyncio.get_running_loop())
         dbuf = DeflateBuffer(buf, "deflate")
 
         dbuf.decompressor = mock.Mock()
@@ -1666,8 +1807,8 @@ class TestDeflateBuffer:
         dbuf.feed_data(b"xxxx")
         assert [b"line"] == list(buf._buffer)
 
-    async def test_feed_data_err(self, stream: Any) -> None:
-        buf = aiohttp.FlowControlDataQueue(stream, 2**16, loop=asyncio.get_event_loop())
+    async def test_feed_data_err(self, protocol: BaseProtocol) -> None:
+        buf = aiohttp.StreamReader(protocol, 2**16, loop=asyncio.get_running_loop())
         dbuf = DeflateBuffer(buf, "deflate")
 
         exc = ValueError()
@@ -1679,8 +1820,8 @@ class TestDeflateBuffer:
             # Should start with b'x', otherwise code switch mocked decoder.
             dbuf.feed_data(b"xsomedata")
 
-    async def test_feed_eof(self, stream: Any) -> None:
-        buf = aiohttp.FlowControlDataQueue(stream, 2**16, loop=asyncio.get_event_loop())
+    async def test_feed_eof(self, protocol: BaseProtocol) -> None:
+        buf = aiohttp.StreamReader(protocol, 2**16, loop=asyncio.get_running_loop())
         dbuf = DeflateBuffer(buf, "deflate")
 
         dbuf.decompressor = mock.Mock()
@@ -1690,8 +1831,8 @@ class TestDeflateBuffer:
         assert [b"line"] == list(buf._buffer)
         assert buf._eof
 
-    async def test_feed_eof_err_deflate(self, stream: Any) -> None:
-        buf = aiohttp.FlowControlDataQueue(stream, 2**16, loop=asyncio.get_event_loop())
+    async def test_feed_eof_err_deflate(self, protocol: BaseProtocol) -> None:
+        buf = aiohttp.StreamReader(protocol, 2**16, loop=asyncio.get_running_loop())
         dbuf = DeflateBuffer(buf, "deflate")
 
         dbuf.decompressor = mock.Mock()
@@ -1701,8 +1842,8 @@ class TestDeflateBuffer:
         with pytest.raises(http_exceptions.ContentEncodingError):
             dbuf.feed_eof()
 
-    async def test_feed_eof_no_err_gzip(self, stream: Any) -> None:
-        buf = aiohttp.FlowControlDataQueue(stream, 2**16, loop=asyncio.get_event_loop())
+    async def test_feed_eof_no_err_gzip(self, protocol: BaseProtocol) -> None:
+        buf = aiohttp.StreamReader(protocol, 2**16, loop=asyncio.get_running_loop())
         dbuf = DeflateBuffer(buf, "gzip")
 
         dbuf.decompressor = mock.Mock()
@@ -1712,8 +1853,8 @@ class TestDeflateBuffer:
         dbuf.feed_eof()
         assert [b"line"] == list(buf._buffer)
 
-    async def test_feed_eof_no_err_brotli(self, stream: Any) -> None:
-        buf = aiohttp.FlowControlDataQueue(stream, 2**16, loop=asyncio.get_event_loop())
+    async def test_feed_eof_no_err_brotli(self, protocol: BaseProtocol) -> None:
+        buf = aiohttp.StreamReader(protocol, 2**16, loop=asyncio.get_running_loop())
         dbuf = DeflateBuffer(buf, "br")
 
         dbuf.decompressor = mock.Mock()
@@ -1723,8 +1864,8 @@ class TestDeflateBuffer:
         dbuf.feed_eof()
         assert [b"line"] == list(buf._buffer)
 
-    async def test_empty_body(self, stream: Any) -> None:
-        buf = aiohttp.FlowControlDataQueue(stream, 2**16, loop=asyncio.get_event_loop())
+    async def test_empty_body(self, protocol: BaseProtocol) -> None:
+        buf = aiohttp.StreamReader(protocol, 2**16, loop=asyncio.get_running_loop())
         dbuf = DeflateBuffer(buf, "deflate")
         dbuf.feed_eof()
 
