@@ -1,6 +1,6 @@
 import asyncio
 from contextlib import suppress
-from typing import Any, Optional, Tuple
+from typing import Any, Optional, Tuple, Type, Union
 
 from .base_protocol import BaseProtocol
 from .client_exceptions import (
@@ -65,9 +65,6 @@ class ResponseHandler(BaseProtocol, DataQueue[Tuple[RawResponseMessage, StreamRe
             or len(self) > 0
             or bool(self._tail)
         )
-
-    def force_close(self) -> None:
-        self._should_close = True
 
     def close(self) -> None:
         transport = self.transport
@@ -162,7 +159,7 @@ class ResponseHandler(BaseProtocol, DataQueue[Tuple[RawResponseMessage, StreamRe
 
     def set_exception(
         self,
-        exc: BaseException,
+        exc: Union[Type[BaseException], BaseException],
         exc_cause: BaseException = _EXC_SENTINEL,
     ) -> None:
         self._should_close = True
