@@ -55,9 +55,7 @@ class FormData:
 
         type_options: MultiDict[str] = MultiDict({"name": name})
         if filename is not None and not isinstance(filename, str):
-            raise TypeError(
-                "filename must be an instance of str. " "Got: %s" % filename
-            )
+            raise TypeError("filename must be an instance of str. Got: %s" % filename)
         if filename is None and isinstance(value, io.IOBase):
             filename = guess_filename(value, name)
         if filename is not None:
@@ -68,7 +66,7 @@ class FormData:
         if content_type is not None:
             if not isinstance(content_type, str):
                 raise TypeError(
-                    "content_type must be an instance of str. " "Got: %s" % content_type
+                    "content_type must be an instance of str. Got: %s" % content_type
                 )
             headers[hdrs.CONTENT_TYPE] = content_type
             self._is_multipart = True
@@ -110,7 +108,7 @@ class FormData:
         if charset == "utf-8":
             content_type = "application/x-www-form-urlencoded"
         else:
-            content_type = "application/x-www-form-urlencoded; " "charset=%s" % charset
+            content_type = "application/x-www-form-urlencoded; charset=%s" % charset
 
         return payload.BytesPayload(
             urlencode(data, doseq=True, encoding=charset).encode(),
