@@ -2,52 +2,96 @@
 
 import asyncio
 import warnings
-from typing import Any, Awaitable, Callable, Dict, List, Optional  # noqa
+from typing import (
+    Any,
+    Awaitable,
+    Callable,
+    Dict,
+    Generic,
+    List,
+    Optional,
+    TypeVar,
+    overload,
+)
 
 from .abc import AbstractStreamWriter
 from .http_parser import RawRequestMessage
 from .streams import StreamReader
-from .web_protocol import RequestHandler, _RequestFactory, _RequestHandler
+from .web_protocol import RequestHandler
 from .web_request import BaseRequest
+from .web_response import StreamResponse
 
 __all__ = ("Server",)
 
+_Request = TypeVar("_Request", bound=BaseRequest)
+_RequestFactory = Callable[
+    [
+        RawRequestMessage,
+        StreamReader,
+        "RequestHandler[_Request]",
+        AbstractStreamWriter,
+        "asyncio.Task[None]",
+    ],
+    _Request,
+]
 
-class Server:
+
+class Server(Generic[_Request]):
+    request_factory: _RequestFactory[_Request]
+
+    @overload
+    def __init__(
+        self: "Server[BaseRequest]",
+        handler: Callable[[_Request], Awaitable[StreamResponse]],
+        *,
+        debug: Optional[bool] = None,
+        handler_cancellation: bool = False,
+        **kwargs: Any,
+    ) -> None: ...
+    @overload
     def __init__(
         self,
-        handler: _RequestHandler,
+        handler: Callable[[_Request], Awaitable[StreamResponse]],
         *,
-        request_factory: Optional[_RequestFactory] = None,
+        request_factory: Optional[_RequestFactory[_Request]],
+        debug: Optional[bool] = None,
+        handler_cancellation: bool = False,
+        **kwargs: Any,
+    ) -> None: ...
+    def __init__(
+        self,
+        handler: Callable[[_Request], Awaitable[StreamResponse]],
+        *,
+        request_factory: Optional[_RequestFactory[_Request]] = None,
         debug: Optional[bool] = None,
         handler_cancellation: bool = False,
         **kwargs: Any,
     ) -> None:
         if debug is not None:
             warnings.warn(
-                "debug argument is no-op since 4.0 " "and scheduled for removal in 5.0",
+                "debug argument is no-op since 4.0 and scheduled for removal in 5.0",
                 DeprecationWarning,
                 stacklevel=2,
             )
         self._loop = asyncio.get_running_loop()
-        self._connections: Dict[RequestHandler, asyncio.Transport] = {}
+        self._connections: Dict[RequestHandler[_Request], asyncio.Transport] = {}
         self._kwargs = kwargs
         self.requests_count = 0
         self.request_handler = handler
-        self.request_factory = request_factory or self._make_request
+        self.request_factory = request_factory or self._make_request  # type: ignore[assignment]
         self.handler_cancellation = handler_cancellation
 
     @property
-    def connections(self) -> List[RequestHandler]:
+    def connections(self) -> List[RequestHandler[_Request]]:
         return list(self._connections.keys())
 
     def connection_made(
-        self, handler: RequestHandler, transport: asyncio.Transport
+        self, handler: RequestHandler[_Request], transport: asyncio.Transport
     ) -> None:
         self._connections[handler] = transport
 
     def connection_lost(
-        self, handler: RequestHandler, exc: Optional[BaseException] = None
+        self, handler: RequestHandler[_Request], exc: Optional[BaseException] = None
     ) -> None:
         if handler in self._connections:
             if handler._task_handler:
@@ -61,7 +105,7 @@ class Server:
         self,
         message: RawRequestMessage,
         payload: StreamReader,
-        protocol: RequestHandler,
+        protocol: RequestHandler[BaseRequest],
         writer: AbstractStreamWriter,
         task: "asyncio.Task[None]",
     ) -> BaseRequest:
@@ -76,7 +120,7 @@ class Server:
         await asyncio.gather(*coros)
         self._connections.clear()
 
-    def __call__(self) -> RequestHandler:
+    def __call__(self) -> RequestHandler[_Request]:
         try:
             return RequestHandler(self, loop=self._loop, **self._kwargs)
         except TypeError:
