@@ -1,40 +1,35 @@
-# type: ignore
 # Tests for aiohttp/client.py
 
+import asyncio
 import gc
 import sys
 from json import JSONDecodeError
-from typing import Any, Callable
 from unittest import mock
 
 import pytest
-from multidict import CIMultiDict
+from multidict import CIMultiDict, CIMultiDictProxy
+from pytest_mock import MockerFixture
 from yarl import URL
 
 import aiohttp
-from aiohttp import http
+from aiohttp import ClientSession, http
 from aiohttp.client_reqrep import ClientResponse, RequestInfo
+from aiohttp.connector import Connection
 from aiohttp.helpers import TimerNoop
 from aiohttp.test_utils import make_mocked_coro
 
 
 class WriterMock(mock.AsyncMock):
-    def __await__(self) -> None:
-        return self().__await__()
-
-    def add_done_callback(self, cb: Callable[[], None]) -> None:
-        cb()
-
     def done(self) -> bool:
         return True
 
 
 @pytest.fixture
-def session():
+def session() -> mock.Mock:
     return mock.Mock()
 
 
-async def test_http_processing_error(session: Any) -> None:
+async def test_http_processing_error(session: ClientSession) -> None:
     loop = mock.Mock()
     request_info = mock.Mock()
     response = ClientResponse(
@@ -53,7 +48,6 @@ async def test_http_processing_error(session: Any) -> None:
 
     connection = mock.Mock()
     connection.protocol = aiohttp.DataQueue(loop)
-    connection.protocol.set_response_params = mock.Mock()
     connection.protocol.set_exception(http.HttpProcessingError())
 
     with pytest.raises(aiohttp.ClientResponseError) as info:
@@ -63,7 +57,7 @@ async def test_http_processing_error(session: Any) -> None:
     response.close()
 
 
-def test_del(session: Any) -> None:
+def test_del(session: ClientSession) -> None:
     loop = mock.Mock()
     response = ClientResponse(
         "get",
@@ -91,7 +85,7 @@ def test_del(session: Any) -> None:
     connection.release.assert_called_with()
 
 
-def test_close(loop: Any, session: Any) -> None:
+def test_close(loop: asyncio.AbstractEventLoop, session: ClientSession) -> None:
     response = ClientResponse(
         "get",
         URL("http://def-cl-resp.org"),
@@ -111,11 +105,13 @@ def test_close(loop: Any, session: Any) -> None:
     response.close()
 
 
-def test_wait_for_100_1(loop: Any, session: Any) -> None:
+def test_wait_for_100_1(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     response = ClientResponse(
         "get",
         URL("http://python.org"),
-        continue100=object(),
+        continue100=loop.create_future(),
         request_info=mock.Mock(),
         writer=WriterMock(),
         timer=TimerNoop(),
@@ -127,7 +123,9 @@ def test_wait_for_100_1(loop: Any, session: Any) -> None:
     response.close()
 
 
-def test_wait_for_100_2(loop: Any, session: Any) -> None:
+def test_wait_for_100_2(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     response = ClientResponse(
         "get",
         URL("http://python.org"),
@@ -143,7 +141,7 @@ def test_wait_for_100_2(loop: Any, session: Any) -> None:
     response.close()
 
 
-def test_repr(loop: Any, session: Any) -> None:
+def test_repr(loop: asyncio.AbstractEventLoop, session: ClientSession) -> None:
     response = ClientResponse(
         "get",
         URL("http://def-cl-resp.org"),
@@ -193,7 +191,9 @@ def test_repr_non_ascii_reason() -> None:
     )
 
 
-async def test_read_and_release_connection(loop: Any, session: Any) -> None:
+async def test_read_and_release_connection(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     response = ClientResponse(
         "get",
         URL("http://def-cl-resp.org"),
@@ -206,7 +206,7 @@ async def test_read_and_release_connection(loop: Any, session: Any) -> None:
         session=session,
     )
 
-    def side_effect(*args, **kwargs):
+    def side_effect(*args: object, **kwargs: object) -> "asyncio.Future[bytes]":
         fut = loop.create_future()
         fut.set_result(b"payload")
         return fut
@@ -219,7 +219,9 @@ async def test_read_and_release_connection(loop: Any, session: Any) -> None:
     assert response._connection is None
 
 
-async def test_read_and_release_connection_with_error(loop: Any, session: Any) -> None:
+async def test_read_and_release_connection_with_error(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     response = ClientResponse(
         "get",
         URL("http://def-cl-resp.org"),
@@ -240,7 +242,7 @@ async def test_read_and_release_connection_with_error(loop: Any, session: Any) -
     assert response._closed
 
 
-async def test_release(loop: Any, session: Any) -> None:
+async def test_release(loop: asyncio.AbstractEventLoop, session: ClientSession) -> None:
     response = ClientResponse(
         "get",
         URL("http://def-cl-resp.org"),
@@ -265,11 +267,13 @@ async def test_release(loop: Any, session: Any) -> None:
     sys.implementation.name != "cpython",
     reason="Other implementations has different GC strategies",
 )
-async def test_release_on_del(loop: Any, session: Any) -> None:
+async def test_release_on_del(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     connection = mock.Mock()
     connection.protocol.upgraded = False
 
-    def run(conn):
+    def run(conn: Connection) -> None:
         response = ClientResponse(
             "get",
             URL("http://def-cl-resp.org"),
@@ -289,12 +293,14 @@ async def test_release_on_del(loop: Any, session: Any) -> None:
     assert connection.release.called
 
 
-async def test_response_eof(loop: Any, session: Any) -> None:
+async def test_response_eof(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     response = ClientResponse(
         "get",
         URL("http://def-cl-resp.org"),
         request_info=mock.Mock(),
-        writer=None,
+        writer=None,  # type: ignore[arg-type]
         continue100=None,
         timer=TimerNoop(),
         traces=[],
@@ -310,7 +316,9 @@ async def test_response_eof(loop: Any, session: Any) -> None:
     assert response._connection is None
 
 
-async def test_response_eof_upgraded(loop: Any, session: Any) -> None:
+async def test_response_eof_upgraded(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     response = ClientResponse(
         "get",
         URL("http://def-cl-resp.org"),
@@ -331,12 +339,14 @@ async def test_response_eof_upgraded(loop: Any, session: Any) -> None:
     assert response._connection is conn
 
 
-async def test_response_eof_after_connection_detach(loop: Any, session: Any) -> None:
+async def test_response_eof_after_connection_detach(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     response = ClientResponse(
         "get",
         URL("http://def-cl-resp.org"),
         request_info=mock.Mock(),
-        writer=None,
+        writer=None,  # type: ignore[arg-type]
         continue100=None,
         timer=TimerNoop(),
         traces=[],
@@ -352,7 +362,7 @@ async def test_response_eof_after_connection_detach(loop: Any, session: Any) -> 
     assert response._connection is None
 
 
-async def test_text(loop: Any, session: Any) -> None:
+async def test_text(loop: asyncio.AbstractEventLoop, session: ClientSession) -> None:
     response = ClientResponse(
         "get",
         URL("http://def-cl-resp.org"),
@@ -365,12 +375,13 @@ async def test_text(loop: Any, session: Any) -> None:
         session=session,
     )
 
-    def side_effect(*args, **kwargs):
+    def side_effect(*args: object, **kwargs: object) -> "asyncio.Future[bytes]":
         fut = loop.create_future()
         fut.set_result('{"тест": "пройден"}'.encode("cp1251"))
         return fut
 
-    response._headers = {"Content-Type": "application/json;charset=cp1251"}
+    h = {"Content-Type": "application/json;charset=cp1251"}
+    response._headers = CIMultiDictProxy(CIMultiDict(h))
     content = response.content = mock.Mock()
     content.read.side_effect = side_effect
 
@@ -379,7 +390,9 @@ async def test_text(loop: Any, session: Any) -> None:
     assert response._connection is None
 
 
-async def test_text_bad_encoding(loop: Any, session: Any) -> None:
+async def test_text_bad_encoding(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     response = ClientResponse(
         "get",
         URL("http://def-cl-resp.org"),
@@ -392,13 +405,14 @@ async def test_text_bad_encoding(loop: Any, session: Any) -> None:
         session=session,
     )
 
-    def side_effect(*args, **kwargs):
+    def side_effect(*args: object, **kwargs: object) -> "asyncio.Future[bytes]":
         fut = loop.create_future()
         fut.set_result('{"тестkey": "пройденvalue"}'.encode("cp1251"))
         return fut
 
     # lie about the encoding
-    response._headers = {"Content-Type": "application/json;charset=utf-8"}
+    h = {"Content-Type": "application/json;charset=utf-8"}
+    response._headers = CIMultiDictProxy(CIMultiDict(h))
     content = response.content = mock.Mock()
     content.read.side_effect = side_effect
     with pytest.raises(UnicodeDecodeError):
@@ -409,7 +423,9 @@ async def test_text_bad_encoding(loop: Any, session: Any) -> None:
     assert response._connection is None
 
 
-async def test_text_custom_encoding(loop: Any, session: Any) -> None:
+async def test_text_custom_encoding(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     response = ClientResponse(
         "get",
         URL("http://def-cl-resp.org"),
@@ -422,25 +438,25 @@ async def test_text_custom_encoding(loop: Any, session: Any) -> None:
         session=session,
     )
 
-    def side_effect(*args, **kwargs):
+    def side_effect(*args: object, **kwargs: object) -> "asyncio.Future[bytes]":
         fut = loop.create_future()
         fut.set_result('{"тест": "пройден"}'.encode("cp1251"))
         return fut
 
-    response._headers = {"Content-Type": "application/json"}
+    h = {"Content-Type": "application/json"}
+    response._headers = CIMultiDictProxy(CIMultiDict(h))
     content = response.content = mock.Mock()
     content.read.side_effect = side_effect
-    response.get_encoding = mock.Mock()
-
-    res = await response.text(encoding="cp1251")
-    assert res == '{"тест": "пройден"}'
-    assert response._connection is None
-    assert not response.get_encoding.called
+    with mock.patch.object(response, "get_encoding") as m:
+        res = await response.text(encoding="cp1251")
+        assert res == '{"тест": "пройден"}'
+        assert response._connection is None
+        assert not m.called
 
 
 @pytest.mark.parametrize("content_type", ("text/plain", "text/plain;charset=invalid"))
 async def test_text_charset_resolver(
-    content_type: str, loop: Any, session: Any
+    content_type: str, loop: asyncio.AbstractEventLoop, session: ClientSession
 ) -> None:
     session._resolve_charset = lambda r, b: "cp1251"
     response = ClientResponse(
@@ -455,12 +471,13 @@ async def test_text_charset_resolver(
         session=session,
     )
 
-    def side_effect(*args, **kwargs):
+    def side_effect(*args: object, **kwargs: object) -> "asyncio.Future[bytes]":
         fut = loop.create_future()
         fut.set_result('{"тест": "пройден"}'.encode("cp1251"))
         return fut
 
-    response._headers = {"Content-Type": content_type}
+    h = {"Content-Type": content_type}
+    response._headers = CIMultiDictProxy(CIMultiDict(h))
     content = response.content = mock.Mock()
     content.read.side_effect = side_effect
 
@@ -471,7 +488,9 @@ async def test_text_charset_resolver(
     assert response.get_encoding() == "cp1251"
 
 
-async def test_get_encoding_body_none(loop: Any, session: Any) -> None:
+async def test_get_encoding_body_none(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     response = ClientResponse(
         "get",
         URL("http://def-cl-resp.org"),
@@ -484,14 +503,10 @@ async def test_get_encoding_body_none(loop: Any, session: Any) -> None:
         session=session,
     )
 
-    def side_effect(*args, **kwargs):
-        fut = loop.create_future()
-        fut.set_result('{"encoding": "test"}')
-        return fut
-
-    response._headers = {"Content-Type": "text/html"}
+    h = {"Content-Type": "text/html"}
+    response._headers = CIMultiDictProxy(CIMultiDict(h))
     content = response.content = mock.Mock()
-    content.read.side_effect = side_effect
+    content.read.side_effect = AssertionError
 
     with pytest.raises(
         RuntimeError,
@@ -501,7 +516,9 @@ async def test_get_encoding_body_none(loop: Any, session: Any) -> None:
     assert response.closed
 
 
-async def test_text_after_read(loop: Any, session: Any) -> None:
+async def test_text_after_read(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     response = ClientResponse(
         "get",
         URL("http://def-cl-resp.org"),
@@ -514,12 +531,13 @@ async def test_text_after_read(loop: Any, session: Any) -> None:
         session=session,
     )
 
-    def side_effect(*args, **kwargs):
+    def side_effect(*args: object, **kwargs: object) -> "asyncio.Future[bytes]":
         fut = loop.create_future()
         fut.set_result('{"тест": "пройден"}'.encode("cp1251"))
         return fut
 
-    response._headers = {"Content-Type": "application/json;charset=cp1251"}
+    h = {"Content-Type": "application/json;charset=cp1251"}
+    response._headers = CIMultiDictProxy(CIMultiDict(h))
     content = response.content = mock.Mock()
     content.read.side_effect = side_effect
 
@@ -528,7 +546,7 @@ async def test_text_after_read(loop: Any, session: Any) -> None:
     assert response._connection is None
 
 
-async def test_json(loop: Any, session: Any) -> None:
+async def test_json(loop: asyncio.AbstractEventLoop, session: ClientSession) -> None:
     response = ClientResponse(
         "get",
         URL("http://def-cl-resp.org"),
@@ -541,12 +559,13 @@ async def test_json(loop: Any, session: Any) -> None:
         session=session,
     )
 
-    def side_effect(*args, **kwargs):
+    def side_effect(*args: object, **kwargs: object) -> "asyncio.Future[bytes]":
         fut = loop.create_future()
         fut.set_result('{"тест": "пройден"}'.encode("cp1251"))
         return fut
 
-    response._headers = {"Content-Type": "application/json;charset=cp1251"}
+    h = {"Content-Type": "application/json;charset=cp1251"}
+    response._headers = CIMultiDictProxy(CIMultiDict(h))
     content = response.content = mock.Mock()
     content.read.side_effect = side_effect
 
@@ -555,7 +574,9 @@ async def test_json(loop: Any, session: Any) -> None:
     assert response._connection is None
 
 
-async def test_json_extended_content_type(loop: Any, session: Any) -> None:
+async def test_json_extended_content_type(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     response = ClientResponse(
         "get",
         URL("http://def-cl-resp.org"),
@@ -568,14 +589,13 @@ async def test_json_extended_content_type(loop: Any, session: Any) -> None:
         session=session,
     )
 
-    def side_effect(*args, **kwargs):
+    def side_effect(*args: object, **kwargs: object) -> "asyncio.Future[bytes]":
         fut = loop.create_future()
         fut.set_result('{"тест": "пройден"}'.encode("cp1251"))
         return fut
 
-    response._headers = {
-        "Content-Type": "application/this.is-1_content+subtype+json;charset=cp1251"
-    }
+    h = {"Content-Type": "application/this.is-1_content+subtype+json;charset=cp1251"}
+    response._headers = CIMultiDictProxy(CIMultiDict(h))
     content = response.content = mock.Mock()
     content.read.side_effect = side_effect
 
@@ -584,7 +604,9 @@ async def test_json_extended_content_type(loop: Any, session: Any) -> None:
     assert response._connection is None
 
 
-async def test_json_custom_content_type(loop: Any, session: Any) -> None:
+async def test_json_custom_content_type(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     response = ClientResponse(
         "get",
         URL("http://def-cl-resp.org"),
@@ -597,12 +619,13 @@ async def test_json_custom_content_type(loop: Any, session: Any) -> None:
         session=session,
     )
 
-    def side_effect(*args, **kwargs):
+    def side_effect(*args: object, **kwargs: object) -> "asyncio.Future[bytes]":
         fut = loop.create_future()
         fut.set_result('{"тест": "пройден"}'.encode("cp1251"))
         return fut
 
-    response._headers = {"Content-Type": "custom/type;charset=cp1251"}
+    h = {"Content-Type": "custom/type;charset=cp1251"}
+    response._headers = CIMultiDictProxy(CIMultiDict(h))
     content = response.content = mock.Mock()
     content.read.side_effect = side_effect
 
@@ -611,7 +634,9 @@ async def test_json_custom_content_type(loop: Any, session: Any) -> None:
     assert response._connection is None
 
 
-async def test_json_custom_loader(loop: Any, session: Any) -> None:
+async def test_json_custom_loader(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     response = ClientResponse(
         "get",
         URL("http://def-cl-resp.org"),
@@ -623,17 +648,20 @@ async def test_json_custom_loader(loop: Any, session: Any) -> None:
         loop=loop,
         session=session,
     )
-    response._headers = {"Content-Type": "application/json;charset=cp1251"}
+    h = {"Content-Type": "application/json;charset=cp1251"}
+    response._headers = CIMultiDictProxy(CIMultiDict(h))
     response._body = b"data"
 
-    def custom(content):
+    def custom(content: str) -> str:
         return content + "-custom"
 
     res = await response.json(loads=custom)
     assert res == "data-custom"
 
 
-async def test_json_invalid_content_type(loop: Any, session: Any) -> None:
+async def test_json_invalid_content_type(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     response = ClientResponse(
         "get",
         URL("http://def-cl-resp.org"),
@@ -645,16 +673,21 @@ async def test_json_invalid_content_type(loop: Any, session: Any) -> None:
         loop=loop,
         session=session,
     )
-    response._headers = {"Content-Type": "data/octet-stream"}
+    h = {"Content-Type": "data/octet-stream"}
+    response._headers = CIMultiDictProxy(CIMultiDict(h))
     response._body = b""
+    response.status = 500
 
     with pytest.raises(aiohttp.ContentTypeError) as info:
         await response.json()
 
     assert info.value.request_info == response.request_info
+    assert info.value.status == 500
 
 
-async def test_json_no_content(loop: Any, session: Any) -> None:
+async def test_json_no_content(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     response = ClientResponse(
         "get",
         URL("http://def-cl-resp.org"),
@@ -666,14 +699,17 @@ async def test_json_no_content(loop: Any, session: Any) -> None:
         loop=loop,
         session=session,
     )
-    response._headers = {"Content-Type": "application/json"}
+    h = {"Content-Type": "application/json"}
+    response._headers = CIMultiDictProxy(CIMultiDict(h))
     response._body = b""
 
     with pytest.raises(JSONDecodeError):
         await response.json(content_type=None)
 
 
-async def test_json_override_encoding(loop: Any, session: Any) -> None:
+async def test_json_override_encoding(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     response = ClientResponse(
         "get",
         URL("http://def-cl-resp.org"),
@@ -686,23 +722,25 @@ async def test_json_override_encoding(loop: Any, session: Any) -> None:
         session=session,
     )
 
-    def side_effect(*args, **kwargs):
+    def side_effect(*args: object, **kwargs: object) -> "asyncio.Future[bytes]":
         fut = loop.create_future()
         fut.set_result('{"тест": "пройден"}'.encode("cp1251"))
         return fut
 
-    response._headers = {"Content-Type": "application/json;charset=utf8"}
+    h = {"Content-Type": "application/json;charset=utf8"}
+    response._headers = CIMultiDictProxy(CIMultiDict(h))
     content = response.content = mock.Mock()
     content.read.side_effect = side_effect
-    response.get_encoding = mock.Mock()
+    with mock.patch.object(response, "get_encoding") as m:
+        res = await response.json(encoding="cp1251")
+        assert res == {"тест": "пройден"}
+        assert response._connection is None
+        assert not m.called
 
-    res = await response.json(encoding="cp1251")
-    assert res == {"тест": "пройден"}
-    assert response._connection is None
-    assert not response.get_encoding.called
 
-
-def test_get_encoding_unknown(loop: Any, session: Any) -> None:
+def test_get_encoding_unknown(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     response = ClientResponse(
         "get",
         URL("http://def-cl-resp.org"),
@@ -715,7 +753,8 @@ def test_get_encoding_unknown(loop: Any, session: Any) -> None:
         session=session,
     )
 
-    response._headers = {"Content-Type": "application/json"}
+    h = {"Content-Type": "application/json"}
+    response._headers = CIMultiDictProxy(CIMultiDict(h))
     assert response.get_encoding() == "utf-8"
 
 
@@ -805,7 +844,8 @@ def test_content_type() -> None:
         loop=mock.Mock(),
         session=mock.Mock(),
     )
-    response._headers = {"Content-Type": "application/json;charset=cp1251"}
+    h = {"Content-Type": "application/json;charset=cp1251"}
+    response._headers = CIMultiDictProxy(CIMultiDict(h))
 
     assert "application/json" == response.content_type
 
@@ -822,7 +862,7 @@ def test_content_type_no_header() -> None:
         loop=mock.Mock(),
         session=mock.Mock(),
     )
-    response._headers = {}
+    response._headers = CIMultiDictProxy(CIMultiDict({}))
 
     assert "application/octet-stream" == response.content_type
 
@@ -839,7 +879,8 @@ def test_charset() -> None:
         loop=mock.Mock(),
         session=mock.Mock(),
     )
-    response._headers = {"Content-Type": "application/json;charset=cp1251"}
+    h = {"Content-Type": "application/json;charset=cp1251"}
+    response._headers = CIMultiDictProxy(CIMultiDict(h))
 
     assert "cp1251" == response.charset
 
@@ -856,7 +897,7 @@ def test_charset_no_header() -> None:
         loop=mock.Mock(),
         session=mock.Mock(),
     )
-    response._headers = {}
+    response._headers = CIMultiDictProxy(CIMultiDict({}))
 
     assert response.charset is None
 
@@ -873,7 +914,8 @@ def test_charset_no_charset() -> None:
         loop=mock.Mock(),
         session=mock.Mock(),
     )
-    response._headers = {"Content-Type": "application/json"}
+    h = {"Content-Type": "application/json"}
+    response._headers = CIMultiDictProxy(CIMultiDict(h))
 
     assert response.charset is None
 
@@ -890,15 +932,15 @@ def test_content_disposition_full() -> None:
         loop=mock.Mock(),
         session=mock.Mock(),
     )
-    response._headers = {
-        "Content-Disposition": 'attachment; filename="archive.tar.gz"; foo=bar'
-    }
+    h = {"Content-Disposition": 'attachment; filename="archive.tar.gz"; foo=bar'}
+    response._headers = CIMultiDictProxy(CIMultiDict(h))
 
+    assert response.content_disposition is not None
     assert "attachment" == response.content_disposition.type
     assert "bar" == response.content_disposition.parameters["foo"]
     assert "archive.tar.gz" == response.content_disposition.filename
     with pytest.raises(TypeError):
-        response.content_disposition.parameters["foo"] = "baz"
+        response.content_disposition.parameters["foo"] = "baz"  # type: ignore[index]
 
 
 def test_content_disposition_no_parameters() -> None:
@@ -913,8 +955,10 @@ def test_content_disposition_no_parameters() -> None:
         loop=mock.Mock(),
         session=mock.Mock(),
     )
-    response._headers = {"Content-Disposition": "attachment"}
+    h = {"Content-Disposition": "attachment"}
+    response._headers = CIMultiDictProxy(CIMultiDict(h))
 
+    assert response.content_disposition is not None
     assert "attachment" == response.content_disposition.type
     assert response.content_disposition.filename is None
     assert {} == response.content_disposition.parameters
@@ -932,17 +976,18 @@ def test_content_disposition_no_header() -> None:
         loop=mock.Mock(),
         session=mock.Mock(),
     )
-    response._headers = {}
+    response._headers = CIMultiDictProxy(CIMultiDict({}))
 
     assert response.content_disposition is None
 
 
 def test_response_request_info() -> None:
-    url = "http://def-cl-resp.org"
-    headers = {"Content-Type": "application/json;charset=cp1251"}
+    url = URL("http://def-cl-resp.org")
+    h = {"Content-Type": "application/json;charset=cp1251"}
+    headers = CIMultiDictProxy(CIMultiDict(h))
     response = ClientResponse(
         "get",
-        URL(url),
+        url,
         request_info=RequestInfo(url, "get", headers, url),
         writer=WriterMock(),
         continue100=None,
@@ -957,11 +1002,12 @@ def test_response_request_info() -> None:
 
 
 def test_request_info_in_exception() -> None:
-    url = "http://def-cl-resp.org"
-    headers = {"Content-Type": "application/json;charset=cp1251"}
+    url = URL("http://def-cl-resp.org")
+    h = {"Content-Type": "application/json;charset=cp1251"}
+    headers = CIMultiDictProxy(CIMultiDict(h))
     response = ClientResponse(
         "get",
-        URL(url),
+        url,
         request_info=RequestInfo(url, "get", headers, url),
         writer=WriterMock(),
         continue100=None,
@@ -978,11 +1024,12 @@ def test_request_info_in_exception() -> None:
 
 
 def test_no_redirect_history_in_exception() -> None:
-    url = "http://def-cl-resp.org"
-    headers = {"Content-Type": "application/json;charset=cp1251"}
+    url = URL("http://def-cl-resp.org")
+    h = {"Content-Type": "application/json;charset=cp1251"}
+    headers = CIMultiDictProxy(CIMultiDict(h))
     response = ClientResponse(
         "get",
-        URL(url),
+        url,
         request_info=RequestInfo(url, "get", headers, url),
         writer=WriterMock(),
         continue100=None,
@@ -999,13 +1046,15 @@ def test_no_redirect_history_in_exception() -> None:
 
 
 def test_redirect_history_in_exception() -> None:
-    hist_url = "http://def-cl-resp.org"
-    url = "http://def-cl-resp.org/index.htm"
-    hist_headers = {"Content-Type": "application/json;charset=cp1251", "Location": url}
-    headers = {"Content-Type": "application/json;charset=cp1251"}
+    hist_url = URL("http://def-cl-resp.org")
+    u = "http://def-cl-resp.org/index.htm"
+    url = URL(u)
+    hist_headers = {"Content-Type": "application/json;charset=cp1251", "Location": u}
+    h = {"Content-Type": "application/json;charset=cp1251"}
+    headers = CIMultiDictProxy(CIMultiDict(h))
     response = ClientResponse(
         "get",
-        URL(url),
+        url,
         request_info=RequestInfo(url, "get", headers, url),
         writer=WriterMock(),
         continue100=None,
@@ -1019,7 +1068,7 @@ def test_redirect_history_in_exception() -> None:
 
     hist_response = ClientResponse(
         "get",
-        URL(hist_url),
+        hist_url,
         request_info=RequestInfo(url, "get", headers, url),
         writer=WriterMock(),
         continue100=None,
@@ -1029,17 +1078,19 @@ def test_redirect_history_in_exception() -> None:
         session=mock.Mock(),
     )
 
-    hist_response._headers = hist_headers
+    hist_response._headers = CIMultiDictProxy(CIMultiDict(hist_headers))
     hist_response.status = 301
     hist_response.reason = "REDIRECT"
 
-    response._history = [hist_response]
+    response._history = (hist_response,)
     with pytest.raises(aiohttp.ClientResponseError) as cm:
         response.raise_for_status()
-    assert [hist_response] == cm.value.history
+    assert (hist_response,) == cm.value.history
 
 
-async def test_response_read_triggers_callback(loop: Any, session: Any) -> None:
+async def test_response_read_triggers_callback(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     trace = mock.Mock()
     trace.send_response_chunk_received = make_mocked_coro()
     response_method = "get"
@@ -1049,7 +1100,7 @@ async def test_response_read_triggers_callback(loop: Any, session: Any) -> None:
     response = ClientResponse(
         response_method,
         response_url,
-        request_info=mock.Mock,
+        request_info=mock.Mock(),
         writer=WriterMock(),
         continue100=None,
         timer=TimerNoop(),
@@ -1058,12 +1109,13 @@ async def test_response_read_triggers_callback(loop: Any, session: Any) -> None:
         traces=[trace],
     )
 
-    def side_effect(*args, **kwargs):
+    def side_effect(*args: object, **kwargs: object) -> "asyncio.Future[bytes]":
         fut = loop.create_future()
         fut.set_result(response_body)
         return fut
 
-    response._headers = {"Content-Type": "application/json;charset=cp1251"}
+    h = {"Content-Type": "application/json;charset=cp1251"}
+    response._headers = CIMultiDictProxy(CIMultiDict(h))
     content = response.content = mock.Mock()
     content.read.side_effect = side_effect
 
@@ -1077,7 +1129,9 @@ async def test_response_read_triggers_callback(loop: Any, session: Any) -> None:
     )
 
 
-def test_response_real_url(loop: Any, session: Any) -> None:
+def test_response_real_url(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     url = URL("http://def-cl-resp.org/#urlfragment")
     response = ClientResponse(
         "get",
@@ -1094,7 +1148,9 @@ def test_response_real_url(loop: Any, session: Any) -> None:
     assert response.real_url == url
 
 
-def test_response_links_comma_separated(loop: Any, session: Any) -> None:
+def test_response_links_comma_separated(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     url = URL("http://def-cl-resp.org/")
     response = ClientResponse(
         "get",
@@ -1107,24 +1163,25 @@ def test_response_links_comma_separated(loop: Any, session: Any) -> None:
         loop=loop,
         session=session,
     )
-    response._headers = CIMultiDict(
-        [
+    h = (
+        (
+            "Link",
             (
-                "Link",
-                (
-                    "<http://example.com/page/1.html>; rel=next, "
-                    "<http://example.com/>; rel=home"
-                ),
-            )
-        ]
+                "<http://example.com/page/1.html>; rel=next, "
+                "<http://example.com/>; rel=home"
+            ),
+        ),
     )
+    response._headers = CIMultiDictProxy(CIMultiDict(h))
     assert response.links == {
         "next": {"url": URL("http://example.com/page/1.html"), "rel": "next"},
         "home": {"url": URL("http://example.com/"), "rel": "home"},
     }
 
 
-def test_response_links_multiple_headers(loop: Any, session: Any) -> None:
+def test_response_links_multiple_headers(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     url = URL("http://def-cl-resp.org/")
     response = ClientResponse(
         "get",
@@ -1137,19 +1194,20 @@ def test_response_links_multiple_headers(loop: Any, session: Any) -> None:
         loop=loop,
         session=session,
     )
-    response._headers = CIMultiDict(
-        [
-            ("Link", "<http://example.com/page/1.html>; rel=next"),
-            ("Link", "<http://example.com/>; rel=home"),
-        ]
+    h = (
+        ("Link", "<http://example.com/page/1.html>; rel=next"),
+        ("Link", "<http://example.com/>; rel=home"),
     )
+    response._headers = CIMultiDictProxy(CIMultiDict(h))
     assert response.links == {
         "next": {"url": URL("http://example.com/page/1.html"), "rel": "next"},
         "home": {"url": URL("http://example.com/"), "rel": "home"},
     }
 
 
-def test_response_links_no_rel(loop: Any, session: Any) -> None:
+def test_response_links_no_rel(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     url = URL("http://def-cl-resp.org/")
     response = ClientResponse(
         "get",
@@ -1162,13 +1220,16 @@ def test_response_links_no_rel(loop: Any, session: Any) -> None:
         loop=loop,
         session=session,
     )
-    response._headers = CIMultiDict([("Link", "<http://example.com/>")])
+    h = (("Link", "<http://example.com/>"),)
+    response._headers = CIMultiDictProxy(CIMultiDict(h))
     assert response.links == {
         "http://example.com/": {"url": URL("http://example.com/")}
     }
 
 
-def test_response_links_quoted(loop: Any, session: Any) -> None:
+def test_response_links_quoted(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     url = URL("http://def-cl-resp.org/")
     response = ClientResponse(
         "get",
@@ -1181,17 +1242,16 @@ def test_response_links_quoted(loop: Any, session: Any) -> None:
         loop=loop,
         session=session,
     )
-    response._headers = CIMultiDict(
-        [
-            ("Link", '<http://example.com/>; rel="home-page"'),
-        ]
-    )
+    h = (("Link", '<http://example.com/>; rel="home-page"'),)
+    response._headers = CIMultiDictProxy(CIMultiDict(h))
     assert response.links == {
         "home-page": {"url": URL("http://example.com/"), "rel": "home-page"}
     }
 
 
-def test_response_links_relative(loop: Any, session: Any) -> None:
+def test_response_links_relative(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     url = URL("http://def-cl-resp.org/")
     response = ClientResponse(
         "get",
@@ -1204,17 +1264,16 @@ def test_response_links_relative(loop: Any, session: Any) -> None:
         loop=loop,
         session=session,
     )
-    response._headers = CIMultiDict(
-        [
-            ("Link", "</relative/path>; rel=rel"),
-        ]
-    )
+    h = (("Link", "</relative/path>; rel=rel"),)
+    response._headers = CIMultiDictProxy(CIMultiDict(h))
     assert response.links == {
         "rel": {"url": URL("http://def-cl-resp.org/relative/path"), "rel": "rel"}
     }
 
 
-def test_response_links_empty(loop: Any, session: Any) -> None:
+def test_response_links_empty(
+    loop: asyncio.AbstractEventLoop, session: ClientSession
+) -> None:
     url = URL("http://def-cl-resp.org/")
     response = ClientResponse(
         "get",
@@ -1227,11 +1286,11 @@ def test_response_links_empty(loop: Any, session: Any) -> None:
         loop=loop,
         session=session,
     )
-    response._headers = CIMultiDict()
+    response._headers = CIMultiDictProxy(CIMultiDict())
     assert response.links == {}
 
 
-def test_response_not_closed_after_get_ok(mocker) -> None:
+def test_response_not_closed_after_get_ok(mocker: MockerFixture) -> None:
     response = ClientResponse(
         "get",
         URL("http://del-cl-resp.org"),
