@@ -1,29 +1,30 @@
-# type: ignore
-from typing import Any
+import io
 from unittest import mock
 
 import pytest
 
 from aiohttp import FormData, web
+from aiohttp.http_writer import StreamWriter
+from aiohttp.pytest_plugin import AiohttpClient
 
 
 @pytest.fixture
-def buf():
+def buf() -> bytearray:
     return bytearray()
 
 
 @pytest.fixture
-def writer(buf: Any):
-    writer = mock.Mock()
+def writer(buf: bytearray) -> StreamWriter:
+    writer = mock.create_autospec(StreamWriter, spec_set=True)
 
-    async def write(chunk):
+    async def write(chunk: bytes) -> None:
         buf.extend(chunk)
 
     writer.write.side_effect = write
-    return writer
+    return writer  # type: ignore[no-any-return]
 
 
-def test_formdata_multipart(buf: Any, writer: Any) -> None:
+def test_formdata_multipart(buf: bytearray) -> None:
     form = FormData()
     assert not form.is_multipart
 
@@ -48,12 +49,22 @@ def test_invalid_formdata_params2() -> None:
         FormData("as")  # 2-char str is not allowed
 
 
+async def test_formdata_textio_charset(buf: bytearray, writer: StreamWriter) -> None:
+    form = FormData()
+    body = io.TextIOWrapper(io.BytesIO(b"\xe6\x97\xa5\xe6\x9c\xac"), encoding="utf-8")
+    form.add_field("foo", body, content_type="text/plain; charset=shift-jis")
+    payload = form()
+    await payload.write(writer)
+    assert b"charset=shift-jis" in buf
+    assert b"\x93\xfa\x96{" in buf
+
+
 def test_invalid_formdata_content_type() -> None:
     form = FormData()
     invalid_vals = [0, 0.1, {}, [], b"foo"]
     for invalid_val in invalid_vals:
         with pytest.raises(TypeError):
-            form.add_field("foo", "bar", content_type=invalid_val)
+            form.add_field("foo", "bar", content_type=invalid_val)  # type: ignore[arg-type]
 
 
 def test_invalid_formdata_filename() -> None:
@@ -61,18 +72,12 @@ def test_invalid_formdata_filename() -> None:
     invalid_vals = [0, 0.1, {}, [], b"foo"]
     for invalid_val in invalid_vals:
         with pytest.raises(TypeError):
-            form.add_field("foo", "bar", filename=invalid_val)
+            form.add_field("foo", "bar", filename=invalid_val)  # type: ignore[arg-type]
 
 
-def test_invalid_formdata_content_transfer_encoding() -> None:
-    form = FormData()
-    invalid_vals = [0, 0.1, {}, [], b"foo"]
-    for invalid_val in invalid_vals:
-        with pytest.raises(TypeError):
-            form.add_field("foo", "bar", content_transfer_encoding=invalid_val)
-
-
-async def test_formdata_field_name_is_quoted(buf: Any, writer: Any) -> None:
+async def test_formdata_field_name_is_quoted(
+    buf: bytearray, writer: StreamWriter
+) -> None:
     form = FormData(charset="ascii")
     form.add_field("email 1", "xxx@x.co", content_type="multipart/form-data")
     payload = form()
@@ -80,7 +85,9 @@ async def test_formdata_field_name_is_quoted(buf: Any, writer: Any) -> None:
     assert b'name="email\\ 1"' in buf
 
 
-async def test_formdata_field_name_is_not_quoted(buf: Any, writer: Any) -> None:
+async def test_formdata_field_name_is_not_quoted(
+    buf: bytearray, writer: StreamWriter
+) -> None:
     form = FormData(quote_fields=False, charset="ascii")
     form.add_field("email 1", "xxx@x.co", content_type="multipart/form-data")
     payload = form()
@@ -88,8 +95,8 @@ async def test_formdata_field_name_is_not_quoted(buf: Any, writer: Any) -> None:
     assert b'name="email 1"' in buf
 
 
-async def test_mark_formdata_as_processed(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_mark_formdata_as_processed(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         return web.Response()
 
     app = web.Application()
