@@ -1,6 +1,6 @@
-# type: ignore
 # Tests for aiohttp/http_writer.py
 import array
+import asyncio
 from typing import Any
 from unittest import mock
 
@@ -8,19 +8,20 @@ import pytest
 from multidict import CIMultiDict
 
 from aiohttp import http
+from aiohttp.base_protocol import BaseProtocol
 from aiohttp.test_utils import make_mocked_coro
 
 
 @pytest.fixture
-def buf():
+def buf() -> bytearray:
     return bytearray()
 
 
 @pytest.fixture
-def transport(buf: Any):
-    transport = mock.Mock()
+def transport(buf: bytearray) -> Any:
+    transport = mock.create_autospec(asyncio.Transport, spec_set=True, instance=True)
 
-    def write(chunk):
+    def write(chunk: bytes) -> None:
         buf.extend(chunk)
 
     transport.write.side_effect = write
@@ -29,32 +30,42 @@ def transport(buf: Any):
 
 
 @pytest.fixture
-def protocol(loop: Any, transport: Any):
-    protocol = mock.Mock(transport=transport)
-    protocol._drain_helper = make_mocked_coro()
-    return protocol
+def protocol(loop: asyncio.AbstractEventLoop, transport: asyncio.Transport) -> Any:
+    return mock.create_autospec(
+        BaseProtocol, spec_set=True, instance=True, transport=transport
+    )
 
 
-def test_payloadwriter_properties(transport: Any, protocol: Any, loop: Any) -> None:
+def test_payloadwriter_properties(
+    transport: asyncio.Transport,
+    protocol: BaseProtocol,
+    loop: asyncio.AbstractEventLoop,
+) -> None:
     writer = http.StreamWriter(protocol, loop)
     assert writer.protocol == protocol
     assert writer.transport == transport
 
 
-async def test_write_payload_eof(transport: Any, protocol: Any, loop: Any) -> None:
-    write = transport.write = mock.Mock()
+async def test_write_payload_eof(
+    transport: asyncio.Transport,
+    protocol: BaseProtocol,
+    loop: asyncio.AbstractEventLoop,
+) -> None:
     msg = http.StreamWriter(protocol, loop)
 
     await msg.write(b"data1")
     await msg.write(b"data2")
     await msg.write_eof()
 
-    content = b"".join([c[1][0] for c in list(write.mock_calls)])
+    content = b"".join([c[1][0] for c in list(transport.write.mock_calls)])  # type: ignore[attr-defined]
     assert b"data1data2" == content.split(b"\r\n\r\n", 1)[-1]
 
 
 async def test_write_payload_chunked(
-    buf: Any, protocol: Any, transport: Any, loop: Any
+    buf: bytearray,
+    protocol: BaseProtocol,
+    transport: asyncio.Transport,
+    loop: asyncio.AbstractEventLoop,
 ) -> None:
     msg = http.StreamWriter(protocol, loop)
     msg.enable_chunking()
@@ -65,7 +76,10 @@ async def test_write_payload_chunked(
 
 
 async def test_write_payload_chunked_multiple(
-    buf: Any, protocol: Any, transport: Any, loop: Any
+    buf: bytearray,
+    protocol: BaseProtocol,
+    transport: asyncio.Transport,
+    loop: asyncio.AbstractEventLoop,
 ) -> None:
     msg = http.StreamWriter(protocol, loop)
     msg.enable_chunking()
@@ -76,38 +90,41 @@ async def test_write_payload_chunked_multiple(
     assert b"5\r\ndata1\r\n5\r\ndata2\r\n0\r\n\r\n" == buf
 
 
-async def test_write_payload_length(protocol: Any, transport: Any, loop: Any) -> None:
-    write = transport.write = mock.Mock()
-
+async def test_write_payload_length(
+    protocol: BaseProtocol,
+    transport: asyncio.Transport,
+    loop: asyncio.AbstractEventLoop,
+) -> None:
     msg = http.StreamWriter(protocol, loop)
     msg.length = 2
     await msg.write(b"d")
     await msg.write(b"ata")
     await msg.write_eof()
 
-    content = b"".join([c[1][0] for c in list(write.mock_calls)])
+    content = b"".join([c[1][0] for c in list(transport.write.mock_calls)])  # type: ignore[attr-defined]
     assert b"da" == content.split(b"\r\n\r\n", 1)[-1]
 
 
 async def test_write_payload_chunked_filter(
-    protocol: Any, transport: Any, loop: Any
+    protocol: BaseProtocol,
+    transport: asyncio.Transport,
+    loop: asyncio.AbstractEventLoop,
 ) -> None:
-    write = transport.write = mock.Mock()
-
     msg = http.StreamWriter(protocol, loop)
     msg.enable_chunking()
     await msg.write(b"da")
     await msg.write(b"ta")
     await msg.write_eof()
 
-    content = b"".join([c[1][0] for c in list(write.mock_calls)])
+    content = b"".join([c[1][0] for c in list(transport.write.mock_calls)])  # type: ignore[attr-defined]
     assert content.endswith(b"2\r\nda\r\n2\r\nta\r\n0\r\n\r\n")
 
 
 async def test_write_payload_chunked_filter_mutiple_chunks(
-    protocol: Any, transport: Any, loop: Any
+    protocol: BaseProtocol,
+    transport: asyncio.Transport,
+    loop: asyncio.AbstractEventLoop,
 ) -> None:
-    write = transport.write = mock.Mock()
     msg = http.StreamWriter(protocol, loop)
     msg.enable_chunking()
     await msg.write(b"da")
@@ -116,30 +133,34 @@ async def test_write_payload_chunked_filter_mutiple_chunks(
     await msg.write(b"at")
     await msg.write(b"a2")
     await msg.write_eof()
-    content = b"".join([c[1][0] for c in list(write.mock_calls)])
+    content = b"".join([c[1][0] for c in list(transport.write.mock_calls)])  # type: ignore[attr-defined]
     assert content.endswith(
-        b"2\r\nda\r\n2\r\nta\r\n2\r\n1d\r\n2\r\nat\r\n" b"2\r\na2\r\n0\r\n\r\n"
+        b"2\r\nda\r\n2\r\nta\r\n2\r\n1d\r\n2\r\nat\r\n2\r\na2\r\n0\r\n\r\n"
     )
 
 
 async def test_write_payload_deflate_compression(
-    protocol: Any, transport: Any, loop: Any
+    protocol: BaseProtocol,
+    transport: asyncio.Transport,
+    loop: asyncio.AbstractEventLoop,
 ) -> None:
     COMPRESSED = b"x\x9cKI,I\x04\x00\x04\x00\x01\x9b"
-    write = transport.write = mock.Mock()
     msg = http.StreamWriter(protocol, loop)
     msg.enable_compression("deflate")
     await msg.write(b"data")
     await msg.write_eof()
 
-    chunks = [c[1][0] for c in list(write.mock_calls)]
+    chunks = [c[1][0] for c in list(transport.write.mock_calls)]  # type: ignore[attr-defined]
     assert all(chunks)
     content = b"".join(chunks)
     assert COMPRESSED == content.split(b"\r\n\r\n", 1)[-1]
 
 
 async def test_write_payload_deflate_and_chunked(
-    buf: Any, protocol: Any, transport: Any, loop: Any
+    buf: bytearray,
+    protocol: BaseProtocol,
+    transport: asyncio.Transport,
+    loop: asyncio.AbstractEventLoop,
 ) -> None:
     msg = http.StreamWriter(protocol, loop)
     msg.enable_compression("deflate")
@@ -149,12 +170,15 @@ async def test_write_payload_deflate_and_chunked(
     await msg.write(b"ta")
     await msg.write_eof()
 
-    thing = b"2\r\nx\x9c\r\n" b"a\r\nKI,I\x04\x00\x04\x00\x01\x9b\r\n" b"0\r\n\r\n"
+    thing = b"2\r\nx\x9c\r\na\r\nKI,I\x04\x00\x04\x00\x01\x9b\r\n0\r\n\r\n"
     assert thing == buf
 
 
 async def test_write_payload_bytes_memoryview(
-    buf: Any, protocol: Any, transport: Any, loop: Any
+    buf: bytearray,
+    protocol: BaseProtocol,
+    transport: asyncio.Transport,
+    loop: asyncio.AbstractEventLoop,
 ) -> None:
     msg = http.StreamWriter(protocol, loop)
 
@@ -168,7 +192,10 @@ async def test_write_payload_bytes_memoryview(
 
 
 async def test_write_payload_short_ints_memoryview(
-    buf: Any, protocol: Any, transport: Any, loop: Any
+    buf: bytearray,
+    protocol: BaseProtocol,
+    transport: asyncio.Transport,
+    loop: asyncio.AbstractEventLoop,
 ) -> None:
     msg = http.StreamWriter(protocol, loop)
     msg.enable_chunking()
@@ -179,14 +206,17 @@ async def test_write_payload_short_ints_memoryview(
     await msg.write_eof()
 
     endians = (
-        (b"6\r\n" b"\x00A\x00B\x00C\r\n" b"0\r\n\r\n"),
-        (b"6\r\n" b"A\x00B\x00C\x00\r\n" b"0\r\n\r\n"),
+        (b"6\r\n\x00A\x00B\x00C\r\n0\r\n\r\n"),
+        (b"6\r\nA\x00B\x00C\x00\r\n0\r\n\r\n"),
     )
     assert buf in endians
 
 
 async def test_write_payload_2d_shape_memoryview(
-    buf: Any, protocol: Any, transport: Any, loop: Any
+    buf: bytearray,
+    protocol: BaseProtocol,
+    transport: asyncio.Transport,
+    loop: asyncio.AbstractEventLoop,
 ) -> None:
     msg = http.StreamWriter(protocol, loop)
     msg.enable_chunking()
@@ -197,12 +227,15 @@ async def test_write_payload_2d_shape_memoryview(
     await msg.write(payload)
     await msg.write_eof()
 
-    thing = b"6\r\n" b"ABCDEF\r\n" b"0\r\n\r\n"
+    thing = b"6\r\nABCDEF\r\n0\r\n\r\n"
     assert thing == buf
 
 
 async def test_write_payload_slicing_long_memoryview(
-    buf: Any, protocol: Any, transport: Any, loop: Any
+    buf: bytearray,
+    protocol: BaseProtocol,
+    transport: asyncio.Transport,
+    loop: asyncio.AbstractEventLoop,
 ) -> None:
     msg = http.StreamWriter(protocol, loop)
     msg.length = 4
@@ -217,18 +250,26 @@ async def test_write_payload_slicing_long_memoryview(
     assert thing == buf
 
 
-async def test_write_drain(protocol: Any, transport: Any, loop: Any) -> None:
+async def test_write_drain(
+    protocol: BaseProtocol,
+    transport: asyncio.Transport,
+    loop: asyncio.AbstractEventLoop,
+) -> None:
     msg = http.StreamWriter(protocol, loop)
-    msg.drain = make_mocked_coro()
-    await msg.write(b"1" * (64 * 1024 * 2), drain=False)
-    assert not msg.drain.called
+    with mock.patch.object(msg, "drain", autospec=True, spec_set=True) as m:
+        await msg.write(b"1" * (64 * 1024 * 2), drain=False)
+        assert not m.called
 
-    await msg.write(b"1", drain=True)
-    assert msg.drain.called
-    assert msg.buffer_size == 0
+        await msg.write(b"1", drain=True)
+        assert m.called
+        assert msg.buffer_size == 0
 
 
-async def test_write_calls_callback(protocol: Any, transport: Any, loop: Any) -> None:
+async def test_write_calls_callback(
+    protocol: BaseProtocol,
+    transport: asyncio.Transport,
+    loop: asyncio.AbstractEventLoop,
+) -> None:
     on_chunk_sent = make_mocked_coro()
     msg = http.StreamWriter(protocol, loop, on_chunk_sent=on_chunk_sent)
     chunk = b"1"
@@ -238,7 +279,9 @@ async def test_write_calls_callback(protocol: Any, transport: Any, loop: Any) ->
 
 
 async def test_write_eof_calls_callback(
-    protocol: Any, transport: Any, loop: Any
+    protocol: BaseProtocol,
+    transport: asyncio.Transport,
+    loop: asyncio.AbstractEventLoop,
 ) -> None:
     on_chunk_sent = make_mocked_coro()
     msg = http.StreamWriter(protocol, loop, on_chunk_sent=on_chunk_sent)
@@ -249,19 +292,23 @@ async def test_write_eof_calls_callback(
 
 
 async def test_write_to_closing_transport(
-    protocol: Any, transport: Any, loop: Any
+    protocol: BaseProtocol,
+    transport: asyncio.Transport,
+    loop: asyncio.AbstractEventLoop,
 ) -> None:
     msg = http.StreamWriter(protocol, loop)
 
     await msg.write(b"Before closing")
-    transport.is_closing.return_value = True
+    transport.is_closing.return_value = True  # type: ignore[attr-defined]
 
     with pytest.raises(ConnectionResetError):
         await msg.write(b"After closing")
 
 
 async def test_write_to_closed_transport(
-    protocol: Any, transport: Any, loop: Any
+    protocol: BaseProtocol,
+    transport: asyncio.Transport,
+    loop: asyncio.AbstractEventLoop,
 ) -> None:
     """Test that writing to a closed transport raises ConnectionResetError.
 
@@ -277,21 +324,31 @@ async def test_write_to_closed_transport(
         await msg.write(b"After transport closed")
 
 
-async def test_drain(protocol: Any, transport: Any, loop: Any) -> None:
+async def test_drain(
+    protocol: BaseProtocol,
+    transport: asyncio.Transport,
+    loop: asyncio.AbstractEventLoop,
+) -> None:
     msg = http.StreamWriter(protocol, loop)
     await msg.drain()
-    assert protocol._drain_helper.called
+    assert protocol._drain_helper.called  # type: ignore[attr-defined]
 
 
-async def test_drain_no_transport(protocol: Any, transport: Any, loop: Any) -> None:
+async def test_drain_no_transport(
+    protocol: BaseProtocol,
+    transport: asyncio.Transport,
+    loop: asyncio.AbstractEventLoop,
+) -> None:
     msg = http.StreamWriter(protocol, loop)
     msg._protocol.transport = None
     await msg.drain()
-    assert not protocol._drain_helper.called
+    assert not protocol._drain_helper.called  # type: ignore[attr-defined]
 
 
 async def test_write_headers_prevents_injection(
-    protocol: Any, transport: Any, loop: Any
+    protocol: BaseProtocol,
+    transport: asyncio.Transport,
+    loop: asyncio.AbstractEventLoop,
 ) -> None:
     msg = http.StreamWriter(protocol, loop)
     status_line = "HTTP/1.1 200 OK"
