@@ -1,13 +1,15 @@
-# type: ignore
 import asyncio
 import io
 import json
 import pathlib
+import sys
 import zlib
-from typing import Any, Optional
+from types import TracebackType
+from typing import Dict, Optional, Tuple, Type
 from unittest import mock
 
 import pytest
+from multidict import CIMultiDict, CIMultiDictProxy
 
 import aiohttp
 from aiohttp import payload
@@ -18,23 +20,29 @@ from aiohttp.hdrs import (
     CONTENT_TYPE,
 )
 from aiohttp.helpers import parse_mimetype
-from aiohttp.multipart import MultipartResponseWrapper
+from aiohttp.multipart import BodyPartReader, MultipartReader, MultipartResponseWrapper
 from aiohttp.streams import StreamReader
-from aiohttp.test_utils import make_mocked_coro
+
+if sys.version_info >= (3, 11):
+    from typing import Self
+else:
+    from typing import TypeVar
+
+    Self = TypeVar("Self", bound="Stream")
 
 BOUNDARY: bytes = b"--:"
 
 
 @pytest.fixture
-def buf():
+def buf() -> bytearray:
     return bytearray()
 
 
 @pytest.fixture
-def stream(buf: Any):
+def stream(buf: bytearray) -> mock.Mock:
     writer = mock.Mock()
 
-    async def write(chunk):
+    async def write(chunk: bytes) -> None:
         buf.extend(chunk)
 
     writer.write.side_effect = write
@@ -42,50 +50,50 @@ def stream(buf: Any):
 
 
 @pytest.fixture
-def writer():
+def writer() -> aiohttp.MultipartWriter:
     return aiohttp.MultipartWriter(boundary=":")
 
 
-class Response:
-    headers: Any
-    content: Any
+class Stream(StreamReader):
+    def __init__(self, content: bytes) -> None:
+        self.content = io.BytesIO(content)
 
-    def __init__(self, headers: Any, content: Any) -> None:
+    async def read(self, size: Optional[int] = None) -> bytes:
+        return self.content.read(size)
+
+    def at_eof(self) -> bool:
+        return self.content.tell() == len(self.content.getbuffer())
+
+    async def readline(self) -> bytes:
+        return self.content.readline()
+
+    def unread_data(self, data: bytes) -> None:
+        self.content = io.BytesIO(data + self.content.read())
+
+    def __enter__(self) -> Self:
+        return self
+
+    def __exit__(
+        self,
+        exc_type: Optional[Type[BaseException]],
+        exc_value: Optional[BaseException],
+        traceback: Optional[TracebackType],
+    ) -> None:
+        self.content.close()
+
+
+class Response:
+    def __init__(self, headers: CIMultiDictProxy[str], content: Stream) -> None:
         self.headers = headers
         self.content = content
 
 
-class Stream:
-    content: Any
-
-    def __init__(self, content: Any) -> None:
-        self.content = io.BytesIO(content)
-
-    async def read(self, size: Optional[Any] = None):
-        return self.content.read(size)
-
-    def at_eof(self):
-        return self.content.tell() == len(self.content.getbuffer())
-
-    async def readline(self):
-        return self.content.readline()
-
-    def unread_data(self, data: Any) -> None:
-        self.content = io.BytesIO(data + self.content.read())
-
-    def __enter__(self):
-        return self
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        self.content.close()
-
-
 class StreamWithShortenRead(Stream):
-    def __init__(self, content: Any) -> None:
+    def __init__(self, content: bytes) -> None:
         self._first = True
         super().__init__(content)
 
-    async def read(self, size: Optional[Any] = None):
+    async def read(self, size: Optional[int] = None) -> bytes:
         if size is not None and self._first:
             self._first = False
             size = size // 2
@@ -94,44 +102,52 @@ class StreamWithShortenRead(Stream):
 
 class TestMultipartResponseWrapper:
     def test_at_eof(self) -> None:
-        wrapper = MultipartResponseWrapper(mock.Mock(), mock.Mock())
+        m_resp = mock.create_autospec(aiohttp.ClientResponse, spec_set=True)
+        m_stream = mock.create_autospec(MultipartReader, spec_set=True)
+        wrapper = MultipartResponseWrapper(m_resp, m_stream)
         wrapper.at_eof()
-        assert wrapper.resp.content.at_eof.called
+        assert m_resp.content.at_eof.called
 
     async def test_next(self) -> None:
-        wrapper = MultipartResponseWrapper(mock.Mock(), mock.Mock())
-        wrapper.stream.next = make_mocked_coro(b"")
-        wrapper.stream.at_eof.return_value = False
+        m_resp = mock.create_autospec(aiohttp.ClientResponse, spec_set=True)
+        m_stream = mock.create_autospec(MultipartReader, spec_set=True)
+        wrapper = MultipartResponseWrapper(m_resp, m_stream)
+        m_stream.next.return_value = b""
+        m_stream.at_eof.return_value = False
         await wrapper.next()
-        assert wrapper.stream.next.called
+        assert m_stream.next.called
 
     async def test_release(self) -> None:
-        wrapper = MultipartResponseWrapper(mock.Mock(), mock.Mock())
-        wrapper.resp.release = make_mocked_coro(None)
+        m_resp = mock.create_autospec(aiohttp.ClientResponse, spec_set=True)
+        m_stream = mock.create_autospec(MultipartReader, spec_set=True)
+        wrapper = MultipartResponseWrapper(m_resp, m_stream)
         await wrapper.release()
-        assert wrapper.resp.release.called
+        assert m_resp.release.called
 
     async def test_release_when_stream_at_eof(self) -> None:
-        wrapper = MultipartResponseWrapper(mock.Mock(), mock.Mock())
-        wrapper.resp.release = make_mocked_coro(None)
-        wrapper.stream.next = make_mocked_coro(b"")
-        wrapper.stream.at_eof.return_value = True
+        m_resp = mock.create_autospec(aiohttp.ClientResponse, spec_set=True)
+        m_stream = mock.create_autospec(MultipartReader, spec_set=True)
+        wrapper = MultipartResponseWrapper(m_resp, m_stream)
+        m_stream.next.return_value = b""
+        m_stream.at_eof.return_value = True
         await wrapper.next()
-        assert wrapper.stream.next.called
-        assert wrapper.resp.release.called
+        assert m_stream.next.called
+        assert m_resp.release.called
 
 
 class TestPartReader:
     async def test_next(self) -> None:
         with Stream(b"Hello, world!\r\n--:") as stream:
-            obj = aiohttp.BodyPartReader(BOUNDARY, {}, stream)
+            d = CIMultiDictProxy[str](CIMultiDict())
+            obj = aiohttp.BodyPartReader(BOUNDARY, d, stream)
             result = await obj.next()
             assert b"Hello, world!" == result
             assert obj.at_eof()
 
     async def test_next_next(self) -> None:
         with Stream(b"Hello, world!\r\n--:") as stream:
-            obj = aiohttp.BodyPartReader(BOUNDARY, {}, stream)
+            d = CIMultiDictProxy[str](CIMultiDict())
+            obj = aiohttp.BodyPartReader(BOUNDARY, d, stream)
             result = await obj.next()
             assert b"Hello, world!" == result
             assert obj.at_eof()
@@ -140,21 +156,24 @@ class TestPartReader:
 
     async def test_read(self) -> None:
         with Stream(b"Hello, world!\r\n--:") as stream:
-            obj = aiohttp.BodyPartReader(BOUNDARY, {}, stream)
+            d = CIMultiDictProxy[str](CIMultiDict())
+            obj = aiohttp.BodyPartReader(BOUNDARY, d, stream)
             result = await obj.read()
             assert b"Hello, world!" == result
             assert obj.at_eof()
 
     async def test_read_chunk_at_eof(self) -> None:
         with Stream(b"--:") as stream:
-            obj = aiohttp.BodyPartReader(BOUNDARY, {}, stream)
+            d = CIMultiDictProxy[str](CIMultiDict())
+            obj = aiohttp.BodyPartReader(BOUNDARY, d, stream)
             obj._at_eof = True
             result = await obj.read_chunk()
         assert b"" == result
 
     async def test_read_chunk_without_content_length(self) -> None:
         with Stream(b"Hello, world!\r\n--:") as stream:
-            obj = aiohttp.BodyPartReader(BOUNDARY, {}, stream)
+            d = CIMultiDictProxy[str](CIMultiDict())
+            obj = aiohttp.BodyPartReader(BOUNDARY, d, stream)
             c1 = await obj.read_chunk(8)
             c2 = await obj.read_chunk(8)
             c3 = await obj.read_chunk(8)
@@ -164,7 +183,7 @@ class TestPartReader:
     async def test_read_incomplete_chunk(self) -> None:
         with Stream(b"") as stream:
 
-            def prepare(data):
+            def prepare(data: bytes) -> bytes:
                 return data
 
             with mock.patch.object(
@@ -177,7 +196,8 @@ class TestPartReader:
                     prepare(b""),
                 ],
             ):
-                obj = aiohttp.BodyPartReader(BOUNDARY, {}, stream)
+                d = CIMultiDictProxy[str](CIMultiDict())
+                obj = aiohttp.BodyPartReader(BOUNDARY, d, stream)
                 c1 = await obj.read_chunk(8)
                 assert c1 == b"Hello, "
                 c2 = await obj.read_chunk(8)
@@ -187,7 +207,8 @@ class TestPartReader:
 
     async def test_read_all_at_once(self) -> None:
         with Stream(b"Hello, World!\r\n--:--\r\n") as stream:
-            obj = aiohttp.BodyPartReader(BOUNDARY, {}, stream)
+            d = CIMultiDictProxy[str](CIMultiDict())
+            obj = aiohttp.BodyPartReader(BOUNDARY, d, stream)
             result = await obj.read_chunk()
             assert b"Hello, World!" == result
             result = await obj.read_chunk()
@@ -196,7 +217,8 @@ class TestPartReader:
 
     async def test_read_incomplete_body_chunked(self) -> None:
         with Stream(b"Hello, World!\r\n-") as stream:
-            obj = aiohttp.BodyPartReader(BOUNDARY, {}, stream)
+            d = CIMultiDictProxy[str](CIMultiDict())
+            obj = aiohttp.BodyPartReader(BOUNDARY, d, stream)
             result = b""
             with pytest.raises(AssertionError):
                 for _ in range(4):
@@ -206,7 +228,7 @@ class TestPartReader:
     async def test_read_boundary_with_incomplete_chunk(self) -> None:
         with Stream(b"") as stream:
 
-            def prepare(data):
+            def prepare(data: bytes) -> bytes:
                 return data
 
             with mock.patch.object(
@@ -219,7 +241,8 @@ class TestPartReader:
                     prepare(b""),
                 ],
             ):
-                obj = aiohttp.BodyPartReader(BOUNDARY, {}, stream)
+                d = CIMultiDictProxy[str](CIMultiDict())
+                obj = aiohttp.BodyPartReader(BOUNDARY, d, stream)
                 c1 = await obj.read_chunk(12)
                 assert c1 == b"Hello, World"
                 c2 = await obj.read_chunk(8)
@@ -229,7 +252,8 @@ class TestPartReader:
 
     async def test_multi_read_chunk(self) -> None:
         with Stream(b"Hello,\r\n--:\r\n\r\nworld!\r\n--:--") as stream:
-            obj = aiohttp.BodyPartReader(BOUNDARY, {}, stream)
+            d = CIMultiDictProxy[str](CIMultiDict())
+            obj = aiohttp.BodyPartReader(BOUNDARY, d, stream)
             result = await obj.read_chunk(8)
             assert b"Hello," == result
             result = await obj.read_chunk(8)
@@ -239,8 +263,9 @@ class TestPartReader:
     async def test_read_chunk_properly_counts_read_bytes(self) -> None:
         expected = b"." * 10
         size = len(expected)
+        h = CIMultiDictProxy(CIMultiDict({"CONTENT-LENGTH": str(size)}))
         with StreamWithShortenRead(expected + b"\r\n--:--") as stream:
-            obj = aiohttp.BodyPartReader(BOUNDARY, {"CONTENT-LENGTH": size}, stream)
+            obj = aiohttp.BodyPartReader(BOUNDARY, h, stream)
             result = bytearray()
             while True:
                 chunk = await obj.read_chunk()
@@ -253,14 +278,16 @@ class TestPartReader:
 
     async def test_read_does_not_read_boundary(self) -> None:
         with Stream(b"Hello, world!\r\n--:") as stream:
-            obj = aiohttp.BodyPartReader(BOUNDARY, {}, stream)
+            d = CIMultiDictProxy[str](CIMultiDict())
+            obj = aiohttp.BodyPartReader(BOUNDARY, d, stream)
             result = await obj.read()
             assert b"Hello, world!" == result
             assert b"--:" == (await stream.read())
 
     async def test_multiread(self) -> None:
         with Stream(b"Hello,\r\n--:\r\n\r\nworld!\r\n--:--") as stream:
-            obj = aiohttp.BodyPartReader(BOUNDARY, {}, stream)
+            d = CIMultiDictProxy[str](CIMultiDict())
+            obj = aiohttp.BodyPartReader(BOUNDARY, d, stream)
             result = await obj.read()
             assert b"Hello," == result
             result = await obj.read()
@@ -269,7 +296,8 @@ class TestPartReader:
 
     async def test_read_multiline(self) -> None:
         with Stream(b"Hello\n,\r\nworld!\r\n--:--") as stream:
-            obj = aiohttp.BodyPartReader(BOUNDARY, {}, stream)
+            d = CIMultiDictProxy[str](CIMultiDict())
+            obj = aiohttp.BodyPartReader(BOUNDARY, d, stream)
             result = await obj.read()
             assert b"Hello\n,\r\nworld!" == result
             result = await obj.read()
@@ -277,27 +305,28 @@ class TestPartReader:
             assert obj.at_eof()
 
     async def test_read_respects_content_length(self) -> None:
+        h = CIMultiDictProxy(CIMultiDict({"CONTENT-LENGTH": "100500"}))
         with Stream(b"." * 100500 + b"\r\n--:--") as stream:
-            obj = aiohttp.BodyPartReader(BOUNDARY, {"CONTENT-LENGTH": 100500}, stream)
+            obj = aiohttp.BodyPartReader(BOUNDARY, h, stream)
             result = await obj.read()
             assert b"." * 100500 == result
             assert obj.at_eof()
 
     async def test_read_with_content_encoding_gzip(self) -> None:
+        h = CIMultiDictProxy(CIMultiDict({CONTENT_ENCODING: "gzip"}))
         with Stream(
             b"\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\x03\x0b\xc9\xccMU"
             b"(\xc9W\x08J\xcdI\xacP\x04\x00$\xfb\x9eV\x0e\x00\x00\x00"
             b"\r\n--:--"
         ) as stream:
-            obj = aiohttp.BodyPartReader(BOUNDARY, {CONTENT_ENCODING: "gzip"}, stream)
+            obj = aiohttp.BodyPartReader(BOUNDARY, h, stream)
             result = await obj.read(decode=True)
         assert b"Time to Relax!" == result
 
     async def test_read_with_content_encoding_deflate(self) -> None:
+        h = CIMultiDictProxy(CIMultiDict({CONTENT_ENCODING: "deflate"}))
         with Stream(b"\x0b\xc9\xccMU(\xc9W\x08J\xcdI\xacP\x04\x00\r\n--:--") as stream:
-            obj = aiohttp.BodyPartReader(
-                BOUNDARY, {CONTENT_ENCODING: "deflate"}, stream
-            )
+            obj = aiohttp.BodyPartReader(BOUNDARY, h, stream)
             result = await obj.read(decode=True)
         assert b"Time to Relax!" == result
 
@@ -307,32 +336,30 @@ class TestPartReader:
             b"(\xc9W\x08J\xcdI\xacP\x04\x00$\xfb\x9eV\x0e\x00\x00\x00"
             b"\r\n"
         )
+        h = CIMultiDictProxy(CIMultiDict({CONTENT_ENCODING: "identity"}))
         with Stream(thing + b"--:--") as stream:
-            obj = aiohttp.BodyPartReader(
-                BOUNDARY, {CONTENT_ENCODING: "identity"}, stream
-            )
+            obj = aiohttp.BodyPartReader(BOUNDARY, h, stream)
             result = await obj.read(decode=True)
         assert thing[:-2] == result
 
     async def test_read_with_content_encoding_unknown(self) -> None:
+        h = CIMultiDictProxy(CIMultiDict({CONTENT_ENCODING: "snappy"}))
         with Stream(b"\x0e4Time to Relax!\r\n--:--") as stream:
-            obj = aiohttp.BodyPartReader(BOUNDARY, {CONTENT_ENCODING: "snappy"}, stream)
+            obj = aiohttp.BodyPartReader(BOUNDARY, h, stream)
             with pytest.raises(RuntimeError):
                 await obj.read(decode=True)
 
     async def test_read_with_content_transfer_encoding_base64(self) -> None:
+        h = CIMultiDictProxy(CIMultiDict({CONTENT_TRANSFER_ENCODING: "base64"}))
         with Stream(b"VGltZSB0byBSZWxheCE=\r\n--:--") as stream:
-            obj = aiohttp.BodyPartReader(
-                BOUNDARY, {CONTENT_TRANSFER_ENCODING: "base64"}, stream
-            )
+            obj = aiohttp.BodyPartReader(BOUNDARY, h, stream)
             result = await obj.read(decode=True)
         assert b"Time to Relax!" == result
 
     async def test_decode_with_content_transfer_encoding_base64(self) -> None:
+        h = CIMultiDictProxy(CIMultiDict({CONTENT_TRANSFER_ENCODING: "base64"}))
         with Stream(b"VG\r\r\nltZSB0byBSZ\r\nWxheCE=\r\n--:--") as stream:
-            obj = aiohttp.BodyPartReader(
-                BOUNDARY, {CONTENT_TRANSFER_ENCODING: "base64"}, stream
-            )
+            obj = aiohttp.BodyPartReader(BOUNDARY, h, stream)
             result = b""
             while not obj.at_eof():
                 chunk = await obj.read_chunk(size=6)
@@ -340,12 +367,13 @@ class TestPartReader:
         assert b"Time to Relax!" == result
 
     async def test_read_with_content_transfer_encoding_quoted_printable(self) -> None:
+        h = CIMultiDictProxy(
+            CIMultiDict({CONTENT_TRANSFER_ENCODING: "quoted-printable"})
+        )
         with Stream(
-            b"=D0=9F=D1=80=D0=B8=D0=B2=D0=B5=D1=82," b" =D0=BC=D0=B8=D1=80!\r\n--:--"
+            b"=D0=9F=D1=80=D0=B8=D0=B2=D0=B5=D1=82, =D0=BC=D0=B8=D1=80!\r\n--:--"
         ) as stream:
-            obj = aiohttp.BodyPartReader(
-                BOUNDARY, {CONTENT_TRANSFER_ENCODING: "quoted-printable"}, stream
-            )
+            obj = aiohttp.BodyPartReader(BOUNDARY, h, stream)
             result = await obj.read(decode=True)
         expected = (
             b"\xd0\x9f\xd1\x80\xd0\xb8\xd0\xb2\xd0\xb5\xd1\x82,"
@@ -355,163 +383,164 @@ class TestPartReader:
 
     @pytest.mark.parametrize("encoding", ("binary", "8bit", "7bit"))
     async def test_read_with_content_transfer_encoding_binary(
-        self, encoding: Any
+        self, encoding: str
     ) -> None:
         data = (
             b"\xd0\x9f\xd1\x80\xd0\xb8\xd0\xb2\xd0\xb5\xd1\x82,"
             b" \xd0\xbc\xd0\xb8\xd1\x80!"
         )
+        h = CIMultiDictProxy(CIMultiDict({CONTENT_TRANSFER_ENCODING: encoding}))
         with Stream(data + b"\r\n--:--") as stream:
-            obj = aiohttp.BodyPartReader(
-                BOUNDARY, {CONTENT_TRANSFER_ENCODING: encoding}, stream
-            )
+            obj = aiohttp.BodyPartReader(BOUNDARY, h, stream)
             result = await obj.read(decode=True)
         assert data == result
 
     async def test_read_with_content_transfer_encoding_unknown(self) -> None:
+        h = CIMultiDictProxy(CIMultiDict({CONTENT_TRANSFER_ENCODING: "unknown"}))
         with Stream(b"\x0e4Time to Relax!\r\n--:--") as stream:
-            obj = aiohttp.BodyPartReader(
-                BOUNDARY, {CONTENT_TRANSFER_ENCODING: "unknown"}, stream
-            )
+            obj = aiohttp.BodyPartReader(BOUNDARY, h, stream)
             with pytest.raises(RuntimeError):
                 await obj.read(decode=True)
 
     async def test_read_text(self) -> None:
         with Stream(b"Hello, world!\r\n--:--") as stream:
-            obj = aiohttp.BodyPartReader(BOUNDARY, {}, stream)
+            d = CIMultiDictProxy[str](CIMultiDict())
+            obj = aiohttp.BodyPartReader(BOUNDARY, d, stream)
             result = await obj.text()
         assert "Hello, world!" == result
 
     async def test_read_text_default_encoding(self) -> None:
         with Stream("Привет, Мир!\r\n--:--".encode()) as stream:
-            obj = aiohttp.BodyPartReader(BOUNDARY, {}, stream)
+            d = CIMultiDictProxy[str](CIMultiDict())
+            obj = aiohttp.BodyPartReader(BOUNDARY, d, stream)
             result = await obj.text()
         assert "Привет, Мир!" == result
 
     async def test_read_text_encoding(self) -> None:
         with Stream("Привет, Мир!\r\n--:--".encode("cp1251")) as stream:
-            obj = aiohttp.BodyPartReader(BOUNDARY, {}, stream)
+            d = CIMultiDictProxy[str](CIMultiDict())
+            obj = aiohttp.BodyPartReader(BOUNDARY, d, stream)
             result = await obj.text(encoding="cp1251")
         assert "Привет, Мир!" == result
 
     async def test_read_text_guess_encoding(self) -> None:
+        h = CIMultiDictProxy(CIMultiDict({CONTENT_TYPE: "text/plain;charset=cp1251"}))
         with Stream("Привет, Мир!\r\n--:--".encode("cp1251")) as stream:
-            obj = aiohttp.BodyPartReader(
-                BOUNDARY, {CONTENT_TYPE: "text/plain;charset=cp1251"}, stream
-            )
+            obj = aiohttp.BodyPartReader(BOUNDARY, h, stream)
             result = await obj.text()
         assert "Привет, Мир!" == result
 
     async def test_read_text_compressed(self) -> None:
+        h = CIMultiDictProxy(
+            CIMultiDict({CONTENT_ENCODING: "deflate", CONTENT_TYPE: "text/plain"})
+        )
         with Stream(b"\x0b\xc9\xccMU(\xc9W\x08J\xcdI\xacP\x04\x00\r\n--:--") as stream:
-            obj = aiohttp.BodyPartReader(
-                BOUNDARY,
-                {CONTENT_ENCODING: "deflate", CONTENT_TYPE: "text/plain"},
-                stream,
-            )
+            obj = aiohttp.BodyPartReader(BOUNDARY, h, stream)
             result = await obj.text()
         assert "Time to Relax!" == result
 
     async def test_read_text_while_closed(self) -> None:
+        h = CIMultiDictProxy(CIMultiDict({CONTENT_TYPE: "text/plain"}))
         with Stream(b"") as stream:
-            obj = aiohttp.BodyPartReader(BOUNDARY, {CONTENT_TYPE: "text/plain"}, stream)
+            obj = aiohttp.BodyPartReader(BOUNDARY, h, stream)
             obj._at_eof = True
             result = await obj.text()
         assert "" == result
 
     async def test_read_json(self) -> None:
+        h = CIMultiDictProxy(CIMultiDict({CONTENT_TYPE: "application/json"}))
         with Stream(b'{"test": "passed"}\r\n--:--') as stream:
-            obj = aiohttp.BodyPartReader(
-                BOUNDARY, {CONTENT_TYPE: "application/json"}, stream
-            )
+            obj = aiohttp.BodyPartReader(BOUNDARY, h, stream)
             result = await obj.json()
         assert {"test": "passed"} == result
 
     async def test_read_json_encoding(self) -> None:
+        h = CIMultiDictProxy(CIMultiDict({CONTENT_TYPE: "application/json"}))
         with Stream('{"тест": "пассед"}\r\n--:--'.encode("cp1251")) as stream:
-            obj = aiohttp.BodyPartReader(
-                BOUNDARY, {CONTENT_TYPE: "application/json"}, stream
-            )
+            obj = aiohttp.BodyPartReader(BOUNDARY, h, stream)
             result = await obj.json(encoding="cp1251")
         assert {"тест": "пассед"} == result
 
     async def test_read_json_guess_encoding(self) -> None:
+        h = CIMultiDictProxy(
+            CIMultiDict({CONTENT_TYPE: "application/json; charset=cp1251"})
+        )
         with Stream('{"тест": "пассед"}\r\n--:--'.encode("cp1251")) as stream:
-            obj = aiohttp.BodyPartReader(
-                BOUNDARY, {CONTENT_TYPE: "application/json; charset=cp1251"}, stream
-            )
+            obj = aiohttp.BodyPartReader(BOUNDARY, h, stream)
             result = await obj.json()
         assert {"тест": "пассед"} == result
 
     async def test_read_json_compressed(self) -> None:
+        h = CIMultiDictProxy(
+            CIMultiDict({CONTENT_ENCODING: "deflate", CONTENT_TYPE: "application/json"})
+        )
         with Stream(b"\xabV*I-.Q\xb2RP*H,.NMQ\xaa\x05\x00\r\n--:--") as stream:
-            obj = aiohttp.BodyPartReader(
-                BOUNDARY,
-                {CONTENT_ENCODING: "deflate", CONTENT_TYPE: "application/json"},
-                stream,
-            )
+            obj = aiohttp.BodyPartReader(BOUNDARY, h, stream)
             result = await obj.json()
         assert {"test": "passed"} == result
 
     async def test_read_json_while_closed(self) -> None:
+        h = CIMultiDictProxy(CIMultiDict({CONTENT_TYPE: "application/json"}))
         with Stream(b"") as stream:
-            obj = aiohttp.BodyPartReader(
-                BOUNDARY, {CONTENT_TYPE: "application/json"}, stream
-            )
+            obj = aiohttp.BodyPartReader(BOUNDARY, h, stream)
             obj._at_eof = True
             result = await obj.json()
         assert result is None
 
     async def test_read_form(self) -> None:
+        h = CIMultiDictProxy(
+            CIMultiDict({CONTENT_TYPE: "application/x-www-form-urlencoded"})
+        )
         with Stream(b"foo=bar&foo=baz&boo=\r\n--:--") as stream:
-            obj = aiohttp.BodyPartReader(
-                BOUNDARY, {CONTENT_TYPE: "application/x-www-form-urlencoded"}, stream
-            )
+            obj = aiohttp.BodyPartReader(BOUNDARY, h, stream)
             result = await obj.form()
         assert [("foo", "bar"), ("foo", "baz"), ("boo", "")] == result
 
     async def test_read_form_invalid_utf8(self) -> None:
+        h = CIMultiDictProxy(
+            CIMultiDict({CONTENT_TYPE: "application/x-www-form-urlencoded"})
+        )
         with Stream(b"\xff\r\n--:--") as stream:
-            obj = aiohttp.BodyPartReader(
-                BOUNDARY, {CONTENT_TYPE: "application/x-www-form-urlencoded"}, stream
-            )
+            obj = aiohttp.BodyPartReader(BOUNDARY, h, stream)
             with pytest.raises(
                 ValueError, match="data cannot be decoded with utf-8 encoding"
             ):
                 await obj.form()
 
     async def test_read_form_encoding(self) -> None:
+        h = CIMultiDictProxy(
+            CIMultiDict({CONTENT_TYPE: "application/x-www-form-urlencoded"})
+        )
         with Stream("foo=bar&foo=baz&boo=\r\n--:--".encode("cp1251")) as stream:
-            obj = aiohttp.BodyPartReader(
-                BOUNDARY, {CONTENT_TYPE: "application/x-www-form-urlencoded"}, stream
-            )
+            obj = aiohttp.BodyPartReader(BOUNDARY, h, stream)
             result = await obj.form(encoding="cp1251")
         assert [("foo", "bar"), ("foo", "baz"), ("boo", "")] == result
 
     async def test_read_form_guess_encoding(self) -> None:
-        with Stream(b"foo=bar&foo=baz&boo=\r\n--:--") as stream:
-            obj = aiohttp.BodyPartReader(
-                BOUNDARY,
-                {CONTENT_TYPE: "application/x-www-form-urlencoded; charset=utf-8"},
-                stream,
+        h = CIMultiDictProxy(
+            CIMultiDict(
+                {CONTENT_TYPE: "application/x-www-form-urlencoded; charset=utf-8"}
             )
+        )
+        with Stream(b"foo=bar&foo=baz&boo=\r\n--:--") as stream:
+            obj = aiohttp.BodyPartReader(BOUNDARY, h, stream)
             result = await obj.form()
         assert [("foo", "bar"), ("foo", "baz"), ("boo", "")] == result
 
     async def test_read_form_while_closed(self) -> None:
+        h = CIMultiDictProxy(
+            CIMultiDict({CONTENT_TYPE: "application/x-www-form-urlencoded"})
+        )
         with Stream(b"") as stream:
-            obj = aiohttp.BodyPartReader(
-                BOUNDARY,
-                {CONTENT_TYPE: "application/x-www-form-urlencoded"},
-                stream,
-            )
+            obj = aiohttp.BodyPartReader(BOUNDARY, h, stream)
             obj._at_eof = True
             result = await obj.form()
         assert not result
 
     async def test_readline(self) -> None:
         with Stream(b"Hello\n,\r\nworld!\r\n--:--") as stream:
-            obj = aiohttp.BodyPartReader(BOUNDARY, {}, stream)
+            d = CIMultiDictProxy[str](CIMultiDict())
+            obj = aiohttp.BodyPartReader(BOUNDARY, d, stream)
             result = await obj.readline()
             assert b"Hello\n" == result
             result = await obj.readline()
@@ -524,29 +553,32 @@ class TestPartReader:
 
     async def test_release(self) -> None:
         with Stream(b"Hello,\r\n--:\r\n\r\nworld!\r\n--:--") as stream:
-            obj = aiohttp.BodyPartReader(BOUNDARY, {}, stream)
+            d = CIMultiDictProxy[str](CIMultiDict())
+            obj = aiohttp.BodyPartReader(BOUNDARY, d, stream)
             await obj.release()
             assert obj.at_eof()
             assert b"--:\r\n\r\nworld!\r\n--:--" == stream.content.read()
 
     async def test_release_respects_content_length(self) -> None:
+        h = CIMultiDictProxy(CIMultiDict({"CONTENT-LENGTH": "100500"}))
         with Stream(b"." * 100500 + b"\r\n--:--") as stream:
-            obj = aiohttp.BodyPartReader(BOUNDARY, {"CONTENT-LENGTH": 100500}, stream)
-            result = await obj.release()
-            assert result is None
+            obj = aiohttp.BodyPartReader(BOUNDARY, h, stream)
+            await obj.release()
             assert obj.at_eof()
 
     async def test_release_release(self) -> None:
         with Stream(b"Hello,\r\n--:\r\n\r\nworld!\r\n--:--") as stream:
-            obj = aiohttp.BodyPartReader(BOUNDARY, {}, stream)
+            d = CIMultiDictProxy[str](CIMultiDict())
+            obj = aiohttp.BodyPartReader(BOUNDARY, d, stream)
             await obj.release()
             await obj.release()
             assert b"--:\r\n\r\nworld!\r\n--:--" == stream.content.read()
 
     async def test_filename(self) -> None:
-        part = aiohttp.BodyPartReader(
-            BOUNDARY, {CONTENT_DISPOSITION: "attachment; filename=foo.html"}, None
+        h = CIMultiDictProxy(
+            CIMultiDict({CONTENT_DISPOSITION: "attachment; filename=foo.html"})
         )
+        part = aiohttp.BodyPartReader(BOUNDARY, h, mock.Mock())
         assert "foo.html" == part.filename
 
     async def test_reading_long_part(self) -> None:
@@ -555,49 +587,56 @@ class TestPartReader:
         stream = StreamReader(protocol, 2**16, loop=asyncio.get_event_loop())
         stream.feed_data(b"0" * size + b"\r\n--:--")
         stream.feed_eof()
-        obj = aiohttp.BodyPartReader(BOUNDARY, {}, stream)
+        d = CIMultiDictProxy[str](CIMultiDict())
+        obj = aiohttp.BodyPartReader(BOUNDARY, d, stream)
         data = await obj.read()
         assert len(data) == size
 
 
 class TestMultipartReader:
     def test_from_response(self) -> None:
+        h = CIMultiDictProxy(
+            CIMultiDict({CONTENT_TYPE: 'multipart/related;boundary=":"'})
+        )
         with Stream(b"--:\r\n\r\nhello\r\n--:--") as stream:
-            resp = Response(
-                {CONTENT_TYPE: 'multipart/related;boundary=":"'},
-                stream,
-            )
-            res = aiohttp.MultipartReader.from_response(resp)
+            resp = Response(h, stream)
+            res = aiohttp.MultipartReader.from_response(resp)  # type: ignore[arg-type]
         assert isinstance(res, MultipartResponseWrapper)
         assert isinstance(res.stream, aiohttp.MultipartReader)
 
     def test_bad_boundary(self) -> None:
+        h = CIMultiDictProxy(
+            CIMultiDict({CONTENT_TYPE: "multipart/related;boundary=" + "a" * 80})
+        )
         with Stream(b"") as stream:
-            resp = Response(
-                {CONTENT_TYPE: "multipart/related;boundary=" + "a" * 80}, stream
-            )
+            resp = Response(h, stream)
             with pytest.raises(ValueError):
-                aiohttp.MultipartReader.from_response(resp)
+                aiohttp.MultipartReader.from_response(resp)  # type: ignore[arg-type]
 
     def test_dispatch(self) -> None:
+        h = CIMultiDictProxy(CIMultiDict({CONTENT_TYPE: "text/plain"}))
         with Stream(b"--:\r\n\r\necho\r\n--:--") as stream:
             reader = aiohttp.MultipartReader(
                 {CONTENT_TYPE: 'multipart/related;boundary=":"'},
                 stream,
             )
-            res = reader._get_part_reader({CONTENT_TYPE: "text/plain"})
+            res = reader._get_part_reader(h)
         assert isinstance(res, reader.part_reader_cls)
 
     def test_dispatch_bodypart(self) -> None:
+        h = CIMultiDictProxy(CIMultiDict({CONTENT_TYPE: "text/plain"}))
         with Stream(b"--:\r\n\r\necho\r\n--:--") as stream:
             reader = aiohttp.MultipartReader(
                 {CONTENT_TYPE: 'multipart/related;boundary=":"'},
                 stream,
             )
-            res = reader._get_part_reader({CONTENT_TYPE: "text/plain"})
+            res = reader._get_part_reader(h)
         assert isinstance(res, reader.part_reader_cls)
 
     def test_dispatch_multipart(self) -> None:
+        h = CIMultiDictProxy(
+            CIMultiDict({CONTENT_TYPE: "multipart/related;boundary=--:--"})
+        )
         with Stream(
             b"----:--\r\n"
             b"\r\n"
@@ -612,15 +651,16 @@ class TestMultipartReader:
                 {CONTENT_TYPE: 'multipart/related;boundary=":"'},
                 stream,
             )
-            res = reader._get_part_reader(
-                {CONTENT_TYPE: "multipart/related;boundary=--:--"}
-            )
+            res = reader._get_part_reader(h)
         assert isinstance(res, reader.__class__)
 
     def test_dispatch_custom_multipart_reader(self) -> None:
         class CustomReader(aiohttp.MultipartReader):
             pass
 
+        h = CIMultiDictProxy(
+            CIMultiDict({CONTENT_TYPE: "multipart/related;boundary=--:--"})
+        )
         with Stream(
             b"----:--\r\n"
             b"\r\n"
@@ -636,17 +676,15 @@ class TestMultipartReader:
                 stream,
             )
             reader.multipart_reader_cls = CustomReader
-            res = reader._get_part_reader(
-                {CONTENT_TYPE: "multipart/related;boundary=--:--"}
-            )
+            res = reader._get_part_reader(h)
         assert isinstance(res, CustomReader)
 
     async def test_emit_next(self) -> None:
+        h = CIMultiDictProxy(
+            CIMultiDict({CONTENT_TYPE: 'multipart/related;boundary=":"'})
+        )
         with Stream(b"--:\r\n\r\necho\r\n--:--") as stream:
-            reader = aiohttp.MultipartReader(
-                {CONTENT_TYPE: 'multipart/related;boundary=":"'},
-                stream,
-            )
+            reader = aiohttp.MultipartReader(h, stream)
             res = await reader.next()
         assert isinstance(res, reader.part_reader_cls)
 
@@ -658,6 +696,68 @@ class TestMultipartReader:
             )
             with pytest.raises(ValueError):
                 await reader.next()
+
+    @pytest.mark.skipif(sys.version_info < (3, 10), reason="Needs anext()")
+    async def test_read_boundary_across_chunks(self) -> None:
+        class SplitBoundaryStream(StreamReader):
+            def __init__(self) -> None:
+                self.content = [
+                    b"--foobar\r\n\r\n",
+                    b"Hello,\r\n-",
+                    b"-fo",
+                    b"ob",
+                    b"ar\r\n",
+                    b"\r\nwor",
+                    b"ld!",
+                    b"\r\n--f",
+                    b"oobar--",
+                ]
+
+            async def read(self, size: Optional[int] = None) -> bytes:
+                chunk = self.content.pop(0)
+                assert size is not None and len(chunk) <= size
+                return chunk
+
+            def at_eof(self) -> bool:
+                return not self.content
+
+            async def readline(self) -> bytes:
+                line = b""
+                while self.content and b"\n" not in line:
+                    line += self.content.pop(0)
+                line, *extra = line.split(b"\n", maxsplit=1)
+                if extra and extra[0]:
+                    self.content.insert(0, extra[0])
+                return line + b"\n"
+
+            def unread_data(self, data: bytes) -> None:
+                if self.content:
+                    self.content[0] = data + self.content[0]
+                else:
+                    self.content.append(data)
+
+        stream = SplitBoundaryStream()
+        reader = aiohttp.MultipartReader(
+            {CONTENT_TYPE: 'multipart/related;boundary="foobar"'}, stream
+        )
+        part = await anext(reader)
+        assert isinstance(part, BodyPartReader)
+        result = await part.read_chunk(10)
+        assert result == b"Hello,"
+        result = await part.read_chunk(10)
+        assert result == b""
+        assert part.at_eof()
+
+        part = await anext(reader)
+        assert isinstance(part, BodyPartReader)
+        result = await part.read_chunk(10)
+        assert result == b"world!"
+        result = await part.read_chunk(10)
+        assert result == b""
+        assert part.at_eof()
+
+        with pytest.raises(StopAsyncIteration):
+            await anext(reader)
 
     async def test_release(self) -> None:
         with Stream(
@@ -704,9 +804,7 @@ class TestMultipartReader:
             assert res is None
 
     async def test_second_next_releases_previous_object(self) -> None:
-        with Stream(
-            b"--:\r\n" b"\r\n" b"test\r\n" b"--:\r\n" b"\r\n" b"passed\r\n" b"--:--"
-        ) as stream:
+        with Stream(b"--:\r\n\r\ntest\r\n--:\r\n\r\npassed\r\n--:--") as stream:
             reader = aiohttp.MultipartReader(
                 {CONTENT_TYPE: 'multipart/related;boundary=":"'},
                 stream,
@@ -714,13 +812,12 @@ class TestMultipartReader:
             first = await reader.next()
             assert isinstance(first, aiohttp.BodyPartReader)
             second = await reader.next()
+            assert second is not None
             assert first.at_eof()
             assert not second.at_eof()
 
     async def test_release_without_read_the_last_object(self) -> None:
-        with Stream(
-            b"--:\r\n" b"\r\n" b"test\r\n" b"--:\r\n" b"\r\n" b"passed\r\n" b"--:--"
-        ) as stream:
+        with Stream(b"--:\r\n\r\ntest\r\n--:\r\n\r\npassed\r\n--:--") as stream:
             reader = aiohttp.MultipartReader(
                 {CONTENT_TYPE: 'multipart/related;boundary=":"'},
                 stream,
@@ -729,6 +826,8 @@ class TestMultipartReader:
             second = await reader.next()
             third = await reader.next()
 
+            assert first is not None
+            assert second is not None
             assert first.at_eof()
             assert second.at_eof()
             assert second.at_eof()
@@ -754,6 +853,7 @@ class TestMultipartReader:
                 part = await reader.next()
                 if part is None:
                     break
+                assert isinstance(part, BodyPartReader)
                 while not part.at_eof():
                     read_part += await part.read_chunk(3)
                 body_parts.append(read_part)
@@ -780,6 +880,7 @@ class TestMultipartReader:
                 part = await reader.next()
                 if part is None:
                     break
+                assert isinstance(part, BodyPartReader)
                 while not part.at_eof():
                     chunk = await part.read_chunk(5)
                     assert chunk
@@ -808,6 +909,7 @@ class TestMultipartReader:
             assert isinstance(first, aiohttp.BodyPartReader)
             second = await reader.next()
 
+            assert isinstance(second, BodyPartReader)
             assert first.at_eof()
             assert not second.at_eof()
 
@@ -836,12 +938,15 @@ class TestMultipartReader:
                 stream,
             )
             field1 = await reader.next()
+            assert isinstance(field1, BodyPartReader)
             assert field1.name == "field1"
             assert field1.get_charset("default") == "ascii"
             field2 = await reader.next()
+            assert isinstance(field2, BodyPartReader)
             assert field2.name == "field2"
             assert field2.get_charset("default") == "UTF-8"
             field3 = await reader.next()
+            assert isinstance(field3, BodyPartReader)
             assert field3.name == "field3"
             assert field3.get_charset("default") == "ascii"
 
@@ -864,12 +969,14 @@ class TestMultipartReader:
                 await reader.next()
 
 
-async def test_writer(writer: Any) -> None:
+async def test_writer(writer: aiohttp.MultipartWriter) -> None:
     assert writer.size == 7
     assert writer.boundary == ":"
 
 
-async def test_writer_serialize_io_chunk(buf: Any, stream: Any, writer: Any) -> None:
+async def test_writer_serialize_io_chunk(
+    buf: bytearray, stream: Stream, writer: aiohttp.MultipartWriter
+) -> None:
     with io.BytesIO(b"foobarbaz") as file_handle:
         writer.append(file_handle)
         await writer.write(stream)
@@ -879,7 +986,9 @@ async def test_writer_serialize_io_chunk(buf: Any, stream: Any, writer: Any) -> 
     )
 
 
-async def test_writer_serialize_json(buf: Any, stream: Any, writer: Any) -> None:
+async def test_writer_serialize_json(
+    buf: bytearray, stream: Stream, writer: aiohttp.MultipartWriter
+) -> None:
     writer.append_json({"привет": "мир"})
     await writer.write(stream)
     assert (
@@ -888,7 +997,9 @@ async def test_writer_serialize_json(buf: Any, stream: Any, writer: Any) -> None
     )
 
 
-async def test_writer_serialize_form(buf: Any, stream: Any, writer: Any) -> None:
+async def test_writer_serialize_form(
+    buf: bytearray, stream: Stream, writer: aiohttp.MultipartWriter
+) -> None:
     data = [("foo", "bar"), ("foo", "baz"), ("boo", "zoo")]
     writer.append_form(data)
     await writer.write(stream)
@@ -896,7 +1007,9 @@ async def test_writer_serialize_form(buf: Any, stream: Any, writer: Any) -> None
     assert b"foo=bar&foo=baz&boo=zoo" in buf
 
 
-async def test_writer_serialize_form_dict(buf: Any, stream: Any, writer: Any) -> None:
+async def test_writer_serialize_form_dict(
+    buf: bytearray, stream: Stream, writer: aiohttp.MultipartWriter
+) -> None:
     data = {"hello": "мир"}
     writer.append_form(data)
     await writer.write(stream)
@@ -904,11 +1017,13 @@ async def test_writer_serialize_form_dict(buf: Any, stream: Any, writer: Any) ->
     assert b"hello=%D0%BC%D0%B8%D1%80" in buf
 
 
-async def test_writer_write(buf: Any, stream: Any, writer: Any) -> None:
+async def test_writer_write(
+    buf: bytearray, stream: Stream, writer: aiohttp.MultipartWriter
+) -> None:
     writer.append("foo-bar-baz")
     writer.append_json({"test": "passed"})
     writer.append_form({"test": "passed"})
-    writer.append_form([("one", 1), ("two", 2)])
+    writer.append_form([("one", "1"), ("two", "2")])
 
     sub_multipart = aiohttp.MultipartWriter(boundary="::")
     sub_multipart.append("nested content")
@@ -950,12 +1065,12 @@ async def test_writer_write(buf: Any, stream: Any, writer: Any) -> None:
     ) == bytes(buf)
 
 
-async def test_writer_write_no_close_boundary(buf: Any, stream: Any) -> None:
+async def test_writer_write_no_close_boundary(buf: bytearray, stream: Stream) -> None:
     writer = aiohttp.MultipartWriter(boundary=":")
     writer.append("foo-bar-baz")
     writer.append_json({"test": "passed"})
     writer.append_form({"test": "passed"})
-    writer.append_form([("one", 1), ("two", 2)])
+    writer.append_form([("one", "1"), ("two", "2")])
     await writer.write(stream, close_boundary=False)
 
     assert (
@@ -982,13 +1097,15 @@ async def test_writer_write_no_close_boundary(buf: Any, stream: Any) -> None:
     ) == bytes(buf)
 
 
-async def test_writer_write_no_parts(buf: Any, stream: Any, writer: Any) -> None:
+async def test_writer_write_no_parts(
+    buf: bytearray, stream: Stream, writer: aiohttp.MultipartWriter
+) -> None:
     await writer.write(stream)
     assert b"--:--\r\n" == bytes(buf)
 
 
 async def test_writer_serialize_with_content_encoding_gzip(
-    buf: Any, stream: Any, writer: Any
+    buf: bytearray, stream: Stream, writer: aiohttp.MultipartWriter
 ) -> None:
     writer.append("Time to Relax!", {CONTENT_ENCODING: "gzip"})
     await writer.write(stream)
@@ -1006,7 +1123,7 @@ async def test_writer_serialize_with_content_encoding_gzip(
 
 
 async def test_writer_serialize_with_content_encoding_deflate(
-    buf: Any, stream: Any, writer: Any
+    buf: bytearray, stream: Stream, writer: aiohttp.MultipartWriter
 ) -> None:
     writer.append("Time to Relax!", {CONTENT_ENCODING: "deflate"})
     await writer.write(stream)
@@ -1022,7 +1139,7 @@ async def test_writer_serialize_with_content_encoding_deflate(
 
 
 async def test_writer_serialize_with_content_encoding_identity(
-    buf: Any, stream: Any, writer: Any
+    buf: bytearray, stream: Stream, writer: aiohttp.MultipartWriter
 ) -> None:
     thing = b"\x0b\xc9\xccMU(\xc9W\x08J\xcdI\xacP\x04\x00"
     writer.append(thing, {CONTENT_ENCODING: "identity"})
@@ -1039,14 +1156,14 @@ async def test_writer_serialize_with_content_encoding_identity(
 
 
 def test_writer_serialize_with_content_encoding_unknown(
-    buf: Any, stream: Any, writer: Any
+    buf: bytearray, stream: Stream, writer: aiohttp.MultipartWriter
 ) -> None:
     with pytest.raises(RuntimeError):
         writer.append("Time to Relax!", {CONTENT_ENCODING: "snappy"})
 
 
 async def test_writer_with_content_transfer_encoding_base64(
-    buf: Any, stream: Any, writer: Any
+    buf: bytearray, stream: Stream, writer: aiohttp.MultipartWriter
 ) -> None:
     writer.append("Time to Relax!", {CONTENT_TRANSFER_ENCODING: "base64"})
     await writer.write(stream)
@@ -1061,7 +1178,7 @@ async def test_writer_with_content_transfer_encoding_base64(
 
 
 async def test_writer_content_transfer_encoding_quote_printable(
-    buf: Any, stream: Any, writer: Any
+    buf: bytearray, stream: Stream, writer: aiohttp.MultipartWriter
 ) -> None:
     writer.append("Привет, мир!", {CONTENT_TRANSFER_ENCODING: "quoted-printable"})
     await writer.write(stream)
@@ -1079,14 +1196,14 @@ async def test_writer_content_transfer_encoding_quote_printable(
 
 
 def test_writer_content_transfer_encoding_unknown(
-    buf: Any, stream: Any, writer: Any
+    buf: bytearray, stream: Stream, writer: aiohttp.MultipartWriter
 ) -> None:
     with pytest.raises(RuntimeError):
         writer.append("Time to Relax!", {CONTENT_TRANSFER_ENCODING: "unknown"})
 
 
 class TestMultipartWriter:
-    def test_default_subtype(self, writer: Any) -> None:
+    def test_default_subtype(self, writer: aiohttp.MultipartWriter) -> None:
         mimetype = parse_mimetype(writer.headers.get(CONTENT_TYPE))
 
         assert "multipart" == mimetype.type
@@ -1110,54 +1227,56 @@ class TestMultipartWriter:
         with pytest.raises(ValueError):
             aiohttp.MultipartWriter(boundary="X" * 71)
 
-    def test_default_headers(self, writer: Any) -> None:
+    def test_default_headers(self, writer: aiohttp.MultipartWriter) -> None:
         expected = {CONTENT_TYPE: 'multipart/mixed; boundary=":"'}
         assert expected == writer.headers
 
-    def test_iter_parts(self, writer: Any) -> None:
+    def test_iter_parts(self, writer: aiohttp.MultipartWriter) -> None:
         writer.append("foo")
         writer.append("bar")
         writer.append("baz")
         assert 3 == len(list(writer))
 
-    def test_append(self, writer: Any) -> None:
+    def test_append(self, writer: aiohttp.MultipartWriter) -> None:
         assert 0 == len(writer)
         writer.append("hello, world!")
         assert 1 == len(writer)
         assert isinstance(writer._parts[0][0], payload.Payload)
 
-    def test_append_with_headers(self, writer: Any) -> None:
+    def test_append_with_headers(self, writer: aiohttp.MultipartWriter) -> None:
         writer.append("hello, world!", {"x-foo": "bar"})
         assert 1 == len(writer)
         assert "x-foo" in writer._parts[0][0].headers
         assert writer._parts[0][0].headers["x-foo"] == "bar"
 
-    def test_append_json(self, writer: Any) -> None:
+    def test_append_json(self, writer: aiohttp.MultipartWriter) -> None:
         writer.append_json({"foo": "bar"})
         assert 1 == len(writer)
         part = writer._parts[0][0]
         assert part.headers[CONTENT_TYPE] == "application/json"
 
-    def test_append_part(self, writer: Any) -> None:
+    def test_append_part(self, writer: aiohttp.MultipartWriter) -> None:
         part = payload.get_payload("test", headers={CONTENT_TYPE: "text/plain"})
         writer.append(part, {CONTENT_TYPE: "test/passed"})
         assert 1 == len(writer)
         part = writer._parts[0][0]
         assert part.headers[CONTENT_TYPE] == "test/passed"
 
-    def test_append_json_overrides_content_type(self, writer: Any) -> None:
+    def test_append_json_overrides_content_type(
+        self, writer: aiohttp.MultipartWriter
+    ) -> None:
         writer.append_json({"foo": "bar"}, {CONTENT_TYPE: "test/passed"})
         assert 1 == len(writer)
         part = writer._parts[0][0]
         assert part.headers[CONTENT_TYPE] == "test/passed"
 
-    def test_append_form(self, writer: Any) -> None:
+    def test_append_form(self, writer: aiohttp.MultipartWriter) -> None:
         writer.append_form({"foo": "bar"}, {CONTENT_TYPE: "test/passed"})
         assert 1 == len(writer)
         part = writer._parts[0][0]
         assert part.headers[CONTENT_TYPE] == "test/passed"
 
-    def test_append_multipart(self, writer: Any) -> None:
+    def test_append_multipart(self, writer: aiohttp.MultipartWriter) -> None:
         subwriter = aiohttp.MultipartWriter(boundary=":")
         subwriter.append_json({"foo": "bar"})
         writer.append(subwriter, {CONTENT_TYPE: "test/passed"})
@@ -1165,13 +1284,13 @@ class TestMultipartWriter:
         part = writer._parts[0][0]
         assert part.headers[CONTENT_TYPE] == "test/passed"
 
-    def test_set_content_disposition_after_append(self):
+    def test_set_content_disposition_after_append(self) -> None:
         writer = aiohttp.MultipartWriter("form-data")
         part = writer.append("some-data")
         part.set_content_disposition("form-data", name="method")
         assert 'name="method"' in part.headers[CONTENT_DISPOSITION]
 
-    def test_automatic_content_disposition(self):
+    def test_automatic_content_disposition(self) -> None:
         writer = aiohttp.MultipartWriter("form-data")
         writer.append_json(())
         part = payload.StringPayload("foo")
@@ -1207,7 +1326,7 @@ class TestMultipartWriter:
                 writer.append(None)
 
     async def test_write_preserves_content_disposition(
-        self, buf: Any, stream: Any
+        self, buf: bytearray, stream: Stream
     ) -> None:
         with aiohttp.MultipartWriter(boundary=":") as writer:
             part = writer.append(b"foo", headers={CONTENT_TYPE: "test/passed"})
@@ -1226,7 +1345,7 @@ class TestMultipartWriter:
         assert message == b"foo\r\n--:--\r\n"
 
     async def test_preserve_content_disposition_header(
-        self, buf: Any, stream: Any
+        self, buf: bytearray, stream: Stream
     ) -> None:
         # https://github.com/aio-libs/aiohttp/pull/3475#issuecomment-451072381
         with pathlib.Path(__file__).open("rb") as fobj:
@@ -1252,7 +1371,7 @@ class TestMultipartWriter:
         )
 
     async def test_set_content_disposition_override(
-        self, buf: Any, stream: Any
+        self, buf: bytearray, stream: Stream
     ) -> None:
         # https://github.com/aio-libs/aiohttp/pull/3475#issuecomment-451072381
         with pathlib.Path(__file__).open("rb") as fobj:
@@ -1278,7 +1397,7 @@ class TestMultipartWriter:
         )
 
     async def test_reset_content_disposition_header(
-        self, buf: Any, stream: Any
+        self, buf: bytearray, stream: Stream
     ) -> None:
         # https://github.com/aio-libs/aiohttp/pull/3475#issuecomment-451072381
         with pathlib.Path(__file__).open("rb") as fobj:
@@ -1305,7 +1424,13 @@ class TestMultipartWriter:
 
 
 async def test_async_for_reader() -> None:
-    data = [{"test": "passed"}, 42, b"plain text", b"aiohttp\n", b"no epilogue"]
+    data: Tuple[Dict[str, str], int, bytes, bytes, bytes] = (
+        {"test": "passed"},
+        42,
+        b"plain text",
+        b"aiohttp\n",
+        b"no epilogue",
+    )
     with Stream(
         b"\r\n".join(
             [
@@ -1353,8 +1478,9 @@ async def test_async_for_reader() -> None:
         )
         idata = iter(data)
 
-        async def check(reader):
+        async def check(reader: aiohttp.MultipartReader) -> None:
             async for part in reader:
+                assert part is not None
                 if isinstance(part, aiohttp.BodyPartReader):
                     if part.headers[CONTENT_TYPE] == "application/json":
                         assert next(idata) == (await part.json())
@@ -1367,7 +1493,8 @@ async def test_async_for_reader() -> None:
 
 
 async def test_async_for_bodypart() -> None:
+    h = CIMultiDictProxy[str](CIMultiDict())
     with Stream(b"foobarbaz\r\n--:--") as stream:
-        part = aiohttp.BodyPartReader(boundary=b"--:", headers={}, content=stream)
+        part = aiohttp.BodyPartReader(boundary=b"--:", headers=h, content=stream)
         async for data in part:
             assert data == b"foobarbaz"
