@@ -1,27 +1,25 @@
-# type: ignore
 import asyncio
 import gc
 import socket
 import ssl
 import unittest
-from typing import Any
 from unittest import mock
 
 from yarl import URL
 
 import aiohttp
 from aiohttp.client_reqrep import ClientRequest, ClientResponse
+from aiohttp.connector import _SSL_CONTEXT_VERIFIED
 from aiohttp.helpers import TimerNoop
 from aiohttp.test_utils import make_mocked_coro
 
 
 class TestProxy(unittest.TestCase):
-    loop: Any
-    response_mock_attrs: Any = {
+    response_mock_attrs = {
         "status": 200,
     }
-    mocked_response: Any = mock.Mock(**response_mock_attrs)
-    clientrequest_mock_attrs: Any = {
+    mocked_response = mock.Mock(**response_mock_attrs)
+    clientrequest_mock_attrs = {
         "return_value.send.return_value.start": make_mocked_coro(mocked_response),
     }
 
@@ -42,7 +40,9 @@ class TestProxy(unittest.TestCase):
         autospec=True,
         spec_set=True,
     )
-    def test_connect(self, start_connection: Any, ClientRequestMock: Any) -> None:
+    def test_connect(
+        self, start_connection: mock.Mock, ClientRequestMock: mock.Mock
+    ) -> None:
         req = ClientRequest(
             "GET",
             URL("http://www.python.org"),
@@ -52,46 +52,50 @@ class TestProxy(unittest.TestCase):
         self.assertEqual(str(req.proxy), "http://proxy.example.com")
 
         # mock all the things!
-        async def make_conn():
+        async def make_conn() -> aiohttp.TCPConnector:
             return aiohttp.TCPConnector()
 
         connector = self.loop.run_until_complete(make_conn())
-        connector._resolve_host = make_mocked_coro(
-            [
-                {
-                    "hostname": "hostname",
-                    "host": "127.0.0.1",
-                    "port": 80,
-                    "family": socket.AF_INET,
-                    "proto": 0,
-                    "flags": 0,
+        r = {
+            "hostname": "hostname",
+            "host": "127.0.0.1",
+            "port": 80,
+            "family": socket.AF_INET,
+            "proto": 0,
+            "flags": 0,
+        }
+        with mock.patch.object(
+            connector, "_resolve_host", autospec=True, return_value=[r]
+        ):
+            proto = mock.Mock(
+                **{
+                    "transport.get_extra_info.return_value": False,
                 }
-            ]
-        )
+            )
+            with mock.patch.object(
+                self.loop,
+                "create_connection",
+                autospec=True,
+                return_value=(proto.transport, proto),
+            ):
+                conn = self.loop.run_until_complete(
+                    connector.connect(req, [], aiohttp.ClientTimeout())
+                )
+                self.assertEqual(req.url, URL("http://www.python.org"))
+                self.assertIs(conn._protocol, proto)
+                self.assertIs(conn.transport, proto.transport)
 
-        proto = mock.Mock(
-            **{
-                "transport.get_extra_info.return_value": False,
-            }
-        )
-        self.loop.create_connection = make_mocked_coro((proto.transport, proto))
-        conn = self.loop.run_until_complete(
-            connector.connect(req, None, aiohttp.ClientTimeout())
-        )
-        self.assertEqual(req.url, URL("http://www.python.org"))
-        self.assertIs(conn._protocol, proto)
-        self.assertIs(conn.transport, proto.transport)
+                ClientRequestMock.assert_called_with(
+                    "GET",
+                    URL("http://proxy.example.com"),
+                    auth=None,
+                    headers={"Host": "www.python.org"},
+                    loop=self.loop,
+                    ssl=True,
+                )
 
-        ClientRequestMock.assert_called_with(
-            "GET",
-            URL("http://proxy.example.com"),
-            auth=None,
-            headers={"Host": "www.python.org"},
-            loop=self.loop,
-            ssl=True,
-        )
-
-        conn.close()
+                conn.close()
+        self.loop.run_until_complete(connector.close())
 
     @mock.patch("aiohttp.connector.ClientRequest")
     @mock.patch(
@@ -99,7 +103,9 @@ class TestProxy(unittest.TestCase):
         autospec=True,
         spec_set=True,
     )
-    def test_proxy_headers(self, start_connection: Any, ClientRequestMock: Any) -> None:
+    def test_proxy_headers(
+        self, start_connection: mock.Mock, ClientRequestMock: mock.Mock
+    ) -> None:
         req = ClientRequest(
             "GET",
             URL("http://www.python.org"),
@@ -110,59 +116,63 @@ class TestProxy(unittest.TestCase):
         self.assertEqual(str(req.proxy), "http://proxy.example.com")
 
         # mock all the things!
-        async def make_conn():
+        async def make_conn() -> aiohttp.TCPConnector:
             return aiohttp.TCPConnector()
 
         connector = self.loop.run_until_complete(make_conn())
-        connector._resolve_host = make_mocked_coro(
-            [
-                {
-                    "hostname": "hostname",
-                    "host": "127.0.0.1",
-                    "port": 80,
-                    "family": socket.AF_INET,
-                    "proto": 0,
-                    "flags": 0,
+        r = {
+            "hostname": "hostname",
+            "host": "127.0.0.1",
+            "port": 80,
+            "family": socket.AF_INET,
+            "proto": 0,
+            "flags": 0,
+        }
+        with mock.patch.object(
+            connector, "_resolve_host", autospec=True, return_value=[r]
+        ):
+            proto = mock.Mock(
+                **{
+                    "transport.get_extra_info.return_value": False,
                 }
-            ]
-        )
+            )
+            with mock.patch.object(
+                self.loop,
+                "create_connection",
+                autospec=True,
+                return_value=(proto.transport, proto),
+            ):
+                conn = self.loop.run_until_complete(
+                    connector.connect(req, [], aiohttp.ClientTimeout())
+                )
+                self.assertEqual(req.url, URL("http://www.python.org"))
+                self.assertIs(conn._protocol, proto)
+                self.assertIs(conn.transport, proto.transport)
 
-        proto = mock.Mock(
-            **{
-                "transport.get_extra_info.return_value": False,
-            }
-        )
-        self.loop.create_connection = make_mocked_coro((proto.transport, proto))
-        conn = self.loop.run_until_complete(
-            connector.connect(req, None, aiohttp.ClientTimeout())
-        )
-        self.assertEqual(req.url, URL("http://www.python.org"))
-        self.assertIs(conn._protocol, proto)
-        self.assertIs(conn.transport, proto.transport)
+                ClientRequestMock.assert_called_with(
+                    "GET",
+                    URL("http://proxy.example.com"),
+                    auth=None,
+                    headers={"Host": "www.python.org", "Foo": "Bar"},
+                    loop=self.loop,
+                    ssl=True,
+                )
 
-        ClientRequestMock.assert_called_with(
-            "GET",
-            URL("http://proxy.example.com"),
-            auth=None,
-            headers={"Host": "www.python.org", "Foo": "Bar"},
-            loop=self.loop,
-            ssl=True,
-        )
-
-        conn.close()
+                conn.close()
+        self.loop.run_until_complete(connector.close())
 
     @mock.patch(
         "aiohttp.connector.aiohappyeyeballs.start_connection",
         autospec=True,
         spec_set=True,
     )
-    def test_proxy_auth(self, start_connection: Any) -> None:
+    def test_proxy_auth(self, start_connection: mock.Mock) -> None:
         with self.assertRaises(ValueError) as ctx:
             ClientRequest(
                 "GET",
                 URL("http://python.org"),
                 proxy=URL("http://proxy.example.com"),
-                proxy_auth=("user", "pass"),
+                proxy_auth=("user", "pass"),  # type: ignore[arg-type]
                 loop=mock.Mock(),
             )
         self.assertEqual(
@@ -175,65 +185,70 @@ class TestProxy(unittest.TestCase):
         autospec=True,
         spec_set=True,
     )
-    def test_proxy_dns_error(self, start_connection: Any) -> None:
-        async def make_conn():
+    def test_proxy_dns_error(self, start_connection: mock.Mock) -> None:
+        async def make_conn() -> aiohttp.TCPConnector:
             return aiohttp.TCPConnector()
 
-        connector: aiohttp.TCPConnector = self.loop.run_until_complete(make_conn())
-        connector._resolve_host = make_mocked_coro(
-            raise_exception=OSError("dont take it serious")
-        )
-
-        req = ClientRequest(
-            "GET",
-            URL("http://www.python.org"),
-            proxy=URL("http://proxy.example.com"),
-            loop=self.loop,
-        )
-        expected_headers = dict(req.headers)
-        with self.assertRaises(aiohttp.ClientConnectorError):
-            self.loop.run_until_complete(
-                connector.connect(req, None, aiohttp.ClientTimeout())
+        connector = self.loop.run_until_complete(make_conn())
+        with mock.patch.object(
+            connector,
+            "_resolve_host",
+            autospec=True,
+            side_effect=OSError("dont take it serious"),
+        ):
+            req = ClientRequest(
+                "GET",
+                URL("http://www.python.org"),
+                proxy=URL("http://proxy.example.com"),
+                loop=self.loop,
             )
-        self.assertEqual(req.url.path, "/")
-        self.assertEqual(dict(req.headers), expected_headers)
+            expected_headers = dict(req.headers)
+            with self.assertRaises(aiohttp.ClientConnectorError):
+                self.loop.run_until_complete(
+                    connector.connect(req, [], aiohttp.ClientTimeout())
+                )
+            self.assertEqual(req.url.path, "/")
+            self.assertEqual(dict(req.headers), expected_headers)
+        self.loop.run_until_complete(connector.close())
 
     @mock.patch(
         "aiohttp.connector.aiohappyeyeballs.start_connection",
         autospec=True,
         spec_set=True,
     )
-    def test_proxy_connection_error(self, start_connection: Any) -> None:
-        async def make_conn():
+    def test_proxy_connection_error(self, start_connection: mock.Mock) -> None:
+        async def make_conn() -> aiohttp.TCPConnector:
             return aiohttp.TCPConnector()
 
         connector = self.loop.run_until_complete(make_conn())
-        connector._resolve_host = make_mocked_coro(
-            [
-                {
-                    "hostname": "www.python.org",
-                    "host": "127.0.0.1",
-                    "port": 80,
-                    "family": socket.AF_INET,
-                    "proto": 0,
-                    "flags": socket.AI_NUMERICHOST,
-                }
-            ]
-        )
-        connector._loop.create_connection = make_mocked_coro(
-            raise_exception=OSError("dont take it serious")
-        )
-
-        req = ClientRequest(
-            "GET",
-            URL("http://www.python.org"),
-            proxy=URL("http://proxy.example.com"),
-            loop=self.loop,
-        )
-        with self.assertRaises(aiohttp.ClientProxyConnectionError):
-            self.loop.run_until_complete(
-                connector.connect(req, None, aiohttp.ClientTimeout())
-            )
+        r = {
+            "hostname": "www.python.org",
+            "host": "127.0.0.1",
+            "port": 80,
+            "family": socket.AF_INET,
+            "proto": 0,
+            "flags": socket.AI_NUMERICHOST,
+        }
+        with mock.patch.object(
+            connector, "_resolve_host", autospec=True, return_value=[r]
+        ):
+            with mock.patch.object(
+                connector._loop,
+                "create_connection",
+                autospec=True,
+                side_effect=OSError("dont take it serious"),
+            ):
+                req = ClientRequest(
+                    "GET",
+                    URL("http://www.python.org"),
+                    proxy=URL("http://proxy.example.com"),
+                    loop=self.loop,
+                )
+                with self.assertRaises(aiohttp.ClientProxyConnectionError):
+                    self.loop.run_until_complete(
+                        connector.connect(req, [], aiohttp.ClientTimeout())
+                    )
+        self.loop.run_until_complete(connector.close())
 
     @mock.patch("aiohttp.connector.ClientRequest")
     @mock.patch(
@@ -242,7 +257,7 @@ class TestProxy(unittest.TestCase):
         spec_set=True,
     )
     def test_proxy_server_hostname_default(
-        self, start_connection: Any, ClientRequestMock: Any
+        self, start_connection: mock.Mock, ClientRequestMock: mock.Mock
     ) -> None:
         proxy_req = ClientRequest(
             "GET", URL("http://proxy.example.com"), loop=self.loop
@@ -253,23 +268,24 @@ class TestProxy(unittest.TestCase):
             "get",
             URL("http://proxy.example.com"),
             request_info=mock.Mock(),
-            writer=None,
+            writer=None,  # type: ignore[arg-type]
             continue100=None,
             timer=TimerNoop(),
             traces=[],
             loop=self.loop,
             session=mock.Mock(),
         )
-        proxy_req.send = make_mocked_coro(proxy_resp)
-        proxy_resp.start = make_mocked_coro(mock.Mock(status=200))
+        with mock.patch.object(
+            proxy_req, "send", autospec=True, return_value=proxy_resp
+        ):
+            with mock.patch.object(proxy_resp, "start", autospec=True) as m:
+                m.return_value.status = 200
 
-        async def make_conn():
-            return aiohttp.TCPConnector()
+                async def make_conn() -> aiohttp.TCPConnector:
+                    return aiohttp.TCPConnector()
 
-        connector = self.loop.run_until_complete(make_conn())
-        connector._resolve_host = make_mocked_coro(
-            [
-                {
+                connector = self.loop.run_until_complete(make_conn())
+                r = {
                     "hostname": "hostname",
                     "host": "127.0.0.1",
                     "port": 80,
@@ -277,30 +293,43 @@ class TestProxy(unittest.TestCase):
                     "proto": 0,
                     "flags": 0,
                 }
-            ]
-        )
+                with mock.patch.object(
+                    connector, "_resolve_host", autospec=True, return_value=[r]
+                ):
+                    tr, proto = mock.Mock(), mock.Mock()
+                    with mock.patch.object(
+                        self.loop,
+                        "create_connection",
+                        autospec=True,
+                        return_value=(tr, proto),
+                    ):
+                        with mock.patch.object(
+                            self.loop,
+                            "start_tls",
+                            autospec=True,
+                            return_value=mock.Mock(),
+                        ) as tls_m:
+                            req = ClientRequest(
+                                "GET",
+                                URL("https://www.python.org"),
+                                proxy=URL("http://proxy.example.com"),
+                                loop=self.loop,
+                            )
+                            self.loop.run_until_complete(
+                                connector._create_connection(
+                                    req, [], aiohttp.ClientTimeout()
+                                )
+                            )
 
-        tr, proto = mock.Mock(), mock.Mock()
-        self.loop.create_connection = make_mocked_coro((tr, proto))
-        self.loop.start_tls = make_mocked_coro(mock.Mock())
+                            self.assertEqual(
+                                tls_m.call_args.kwargs["server_hostname"],
+                                "www.python.org",
+                            )
 
-        req = ClientRequest(
-            "GET",
-            URL("https://www.python.org"),
-            proxy=URL("http://proxy.example.com"),
-            loop=self.loop,
-        )
-        self.loop.run_until_complete(
-            connector._create_connection(req, None, aiohttp.ClientTimeout())
-        )
-
-        self.assertEqual(
-            self.loop.start_tls.call_args.kwargs["server_hostname"], "www.python.org"
-        )
-
-        self.loop.run_until_complete(proxy_req.close())
-        proxy_resp.close()
-        self.loop.run_until_complete(req.close())
+                            self.loop.run_until_complete(proxy_req.close())
+                            proxy_resp.close()
+                            self.loop.run_until_complete(req.close())
+                self.loop.run_until_complete(connector.close())
 
     @mock.patch("aiohttp.connector.ClientRequest")
     @mock.patch(
@@ -309,7 +338,7 @@ class TestProxy(unittest.TestCase):
         spec_set=True,
     )
     def test_proxy_server_hostname_override(
-        self, start_connection: Any, ClientRequestMock: Any
+        self, start_connection: mock.Mock, ClientRequestMock: mock.Mock
     ) -> None:
         proxy_req = ClientRequest(
             "GET",
@@ -322,23 +351,24 @@ class TestProxy(unittest.TestCase):
             "get",
             URL("http://proxy.example.com"),
             request_info=mock.Mock(),
-            writer=None,
+            writer=None,  # type: ignore[arg-type]
             continue100=None,
             timer=TimerNoop(),
             traces=[],
             loop=self.loop,
             session=mock.Mock(),
         )
-        proxy_req.send = make_mocked_coro(proxy_resp)
-        proxy_resp.start = make_mocked_coro(mock.Mock(status=200))
+        with mock.patch.object(
+            proxy_req, "send", autospec=True, return_value=proxy_resp
+        ):
+            with mock.patch.object(proxy_resp, "start", autospec=True) as m:
+                m.return_value.status = 200
 
-        async def make_conn():
-            return aiohttp.TCPConnector()
+                async def make_conn() -> aiohttp.TCPConnector:
+                    return aiohttp.TCPConnector()
 
-        connector = self.loop.run_until_complete(make_conn())
-        connector._resolve_host = make_mocked_coro(
-            [
-                {
+                connector = self.loop.run_until_complete(make_conn())
+                r = {
                     "hostname": "hostname",
                     "host": "127.0.0.1",
                     "port": 80,
@@ -346,32 +376,44 @@ class TestProxy(unittest.TestCase):
                     "proto": 0,
                     "flags": 0,
                 }
-            ]
-        )
+                with mock.patch.object(
+                    connector, "_resolve_host", autospec=True, return_value=[r]
+                ):
+                    tr, proto = mock.Mock(), mock.Mock()
+                    with mock.patch.object(
+                        self.loop,
+                        "create_connection",
+                        autospec=True,
+                        return_value=(tr, proto),
+                    ):
+                        with mock.patch.object(
+                            self.loop,
+                            "start_tls",
+                            autospec=True,
+                            return_value=mock.Mock(),
+                        ) as tls_m:
+                            req = ClientRequest(
+                                "GET",
+                                URL("https://www.python.org"),
+                                proxy=URL("http://proxy.example.com"),
+                                server_hostname="server-hostname.example.com",
+                                loop=self.loop,
+                            )
+                            self.loop.run_until_complete(
+                                connector._create_connection(
+                                    req, [], aiohttp.ClientTimeout()
+                                )
+                            )
 
-        tr, proto = mock.Mock(), mock.Mock()
-        self.loop.create_connection = make_mocked_coro((tr, proto))
-        self.loop.start_tls = make_mocked_coro(mock.Mock())
+                            self.assertEqual(
+                                tls_m.call_args.kwargs["server_hostname"],
+                                "server-hostname.example.com",
+                            )
 
-        req = ClientRequest(
-            "GET",
-            URL("https://www.python.org"),
-            proxy=URL("http://proxy.example.com"),
-            server_hostname="server-hostname.example.com",
-            loop=self.loop,
-        )
-        self.loop.run_until_complete(
-            connector._create_connection(req, None, aiohttp.ClientTimeout())
-        )
-
-        self.assertEqual(
-            self.loop.start_tls.call_args.kwargs["server_hostname"],
-            "server-hostname.example.com",
-        )
-
-        self.loop.run_until_complete(proxy_req.close())
-        proxy_resp.close()
-        self.loop.run_until_complete(req.close())
+                            self.loop.run_until_complete(proxy_req.close())
+                            proxy_resp.close()
+                            self.loop.run_until_complete(req.close())
+                self.loop.run_until_complete(connector.close())
 
     @mock.patch("aiohttp.connector.ClientRequest")
     @mock.patch(
@@ -379,7 +421,9 @@ class TestProxy(unittest.TestCase):
         autospec=True,
         spec_set=True,
     )
-    def test_https_connect(self, start_connection: Any, ClientRequestMock: Any) -> None:
+    def test_https_connect(
+        self, start_connection: mock.Mock, ClientRequestMock: mock.Mock
+    ) -> None:
         proxy_req = ClientRequest(
             "GET", URL("http://proxy.example.com"), loop=self.loop
         )
@@ -389,23 +433,24 @@ class TestProxy(unittest.TestCase):
             "get",
             URL("http://proxy.example.com"),
             request_info=mock.Mock(),
-            writer=None,
+            writer=None,  # type: ignore[arg-type]
             continue100=None,
             timer=TimerNoop(),
             traces=[],
             loop=self.loop,
             session=mock.Mock(),
         )
-        proxy_req.send = make_mocked_coro(proxy_resp)
-        proxy_resp.start = make_mocked_coro(mock.Mock(status=200))
+        with mock.patch.object(
+            proxy_req, "send", autospec=True, return_value=proxy_resp
+        ):
+            with mock.patch.object(proxy_resp, "start", autospec=True) as m:
+                m.return_value.status = 200
 
-        async def make_conn():
-            return aiohttp.TCPConnector()
+                async def make_conn() -> aiohttp.TCPConnector:
+                    return aiohttp.TCPConnector()
 
-        connector = self.loop.run_until_complete(make_conn())
-        connector._resolve_host = make_mocked_coro(
-            [
-                {
+                connector = self.loop.run_until_complete(make_conn())
+                r = {
                     "hostname": "hostname",
                     "host": "127.0.0.1",
                     "port": 80,
@@ -413,30 +458,44 @@ class TestProxy(unittest.TestCase):
                     "proto": 0,
                     "flags": 0,
                 }
-            ]
-        )
+                with mock.patch.object(
+                    connector, "_resolve_host", autospec=True, return_value=[r]
+                ):
+                    tr, proto = mock.Mock(), mock.Mock()
+                    with mock.patch.object(
+                        self.loop,
+                        "create_connection",
+                        autospec=True,
+                        return_value=(tr, proto),
+                    ):
+                        with mock.patch.object(
+                            self.loop,
+                            "start_tls",
+                            autospec=True,
+                            return_value=mock.Mock(),
+                        ):
+                            req = ClientRequest(
+                                "GET",
+                                URL("https://www.python.org"),
+                                proxy=URL("http://proxy.example.com"),
+                                loop=self.loop,
+                            )
+                            self.loop.run_until_complete(
+                                connector._create_connection(
+                                    req, [], aiohttp.ClientTimeout()
+                                )
+                            )
 
-        tr, proto = mock.Mock(), mock.Mock()
-        self.loop.create_connection = make_mocked_coro((tr, proto))
-        self.loop.start_tls = make_mocked_coro(mock.Mock())
+                            self.assertEqual(req.url.path, "/")
+                            self.assertEqual(proxy_req.method, "CONNECT")
+                            self.assertEqual(
+                                proxy_req.url, URL("https://www.python.org")
+                            )
 
-        req = ClientRequest(
-            "GET",
-            URL("https://www.python.org"),
-            proxy=URL("http://proxy.example.com"),
-            loop=self.loop,
-        )
-        self.loop.run_until_complete(
-            connector._create_connection(req, None, aiohttp.ClientTimeout())
-        )
-
-        self.assertEqual(req.url.path, "/")
-        self.assertEqual(proxy_req.method, "CONNECT")
-        self.assertEqual(proxy_req.url, URL("https://www.python.org"))
-
-        self.loop.run_until_complete(proxy_req.close())
-        proxy_resp.close()
-        self.loop.run_until_complete(req.close())
+                            self.loop.run_until_complete(proxy_req.close())
+                            proxy_resp.close()
+                            self.loop.run_until_complete(req.close())
+                self.loop.run_until_complete(connector.close())
 
     @mock.patch("aiohttp.connector.ClientRequest")
     @mock.patch(
@@ -445,7 +504,7 @@ class TestProxy(unittest.TestCase):
         spec_set=True,
     )
     def test_https_connect_certificate_error(
-        self, start_connection: Any, ClientRequestMock: Any
+        self, start_connection: mock.Mock, ClientRequestMock: mock.Mock
     ) -> None:
         proxy_req = ClientRequest(
             "GET", URL("http://proxy.example.com"), loop=self.loop
@@ -456,23 +515,24 @@ class TestProxy(unittest.TestCase):
             "get",
             URL("http://proxy.example.com"),
             request_info=mock.Mock(),
-            writer=None,
+            writer=None,  # type: ignore[arg-type]
             continue100=None,
             timer=TimerNoop(),
             traces=[],
             loop=self.loop,
             session=mock.Mock(),
         )
-        proxy_req.send = make_mocked_coro(proxy_resp)
-        proxy_resp.start = make_mocked_coro(mock.Mock(status=200))
+        with mock.patch.object(
+            proxy_req, "send", autospec=True, return_value=proxy_resp
+        ):
+            with mock.patch.object(proxy_resp, "start", autospec=True) as m:
+                m.return_value.status = 200
 
-        async def make_conn():
-            return aiohttp.TCPConnector()
+                async def make_conn() -> aiohttp.TCPConnector:
+                    return aiohttp.TCPConnector()
 
-        connector = self.loop.run_until_complete(make_conn())
-        connector._resolve_host = make_mocked_coro(
-            [
-                {
+                connector = self.loop.run_until_complete(make_conn())
+                r = {
                     "hostname": "hostname",
                     "host": "127.0.0.1",
                     "port": 80,
@@ -480,24 +540,39 @@ class TestProxy(unittest.TestCase):
                     "proto": 0,
                     "flags": 0,
                 }
-            ]
-        )
-
-        # Called on connection to http://proxy.example.com
-        self.loop.create_connection = make_mocked_coro((mock.Mock(), mock.Mock()))
-        # Called on connection to https://www.python.org
-        self.loop.start_tls = make_mocked_coro(raise_exception=ssl.CertificateError)
-
-        req = ClientRequest(
-            "GET",
-            URL("https://www.python.org"),
-            proxy=URL("http://proxy.example.com"),
-            loop=self.loop,
-        )
-        with self.assertRaises(aiohttp.ClientConnectorCertificateError):
-            self.loop.run_until_complete(
-                connector._create_connection(req, None, aiohttp.ClientTimeout())
-            )
+                with mock.patch.object(
+                    connector, "_resolve_host", autospec=True, return_value=[r]
+                ):
+                    tr, proto = mock.Mock(), mock.Mock()
+                    # Called on connection to http://proxy.example.com
+                    with mock.patch.object(
+                        self.loop,
+                        "create_connection",
+                        autospec=True,
+                        return_value=(tr, proto),
+                    ):
+                        # Called on connection to https://www.python.org
+                        with mock.patch.object(
+                            self.loop,
+                            "start_tls",
+                            autospec=True,
+                            side_effect=ssl.CertificateError,
+                        ):
+                            req = ClientRequest(
+                                "GET",
+                                URL("https://www.python.org"),
+                                proxy=URL("http://proxy.example.com"),
+                                loop=self.loop,
+                            )
+                            with self.assertRaises(
+                                aiohttp.ClientConnectorCertificateError
+                            ):
+                                self.loop.run_until_complete(
+                                    connector._create_connection(
+                                        req, [], aiohttp.ClientTimeout()
+                                    )
+                                )
+                self.loop.run_until_complete(connector.close())
 
     @mock.patch("aiohttp.connector.ClientRequest")
     @mock.patch(
@@ -506,7 +581,7 @@ class TestProxy(unittest.TestCase):
         spec_set=True,
     )
     def test_https_connect_ssl_error(
-        self, start_connection: Any, ClientRequestMock: Any
+        self, start_connection: mock.Mock, ClientRequestMock: mock.Mock
     ) -> None:
         proxy_req = ClientRequest(
             "GET", URL("http://proxy.example.com"), loop=self.loop
@@ -517,23 +592,24 @@ class TestProxy(unittest.TestCase):
             "get",
             URL("http://proxy.example.com"),
             request_info=mock.Mock(),
-            writer=None,
+            writer=None,  # type: ignore[arg-type]
             continue100=None,
             timer=TimerNoop(),
             traces=[],
             loop=self.loop,
             session=mock.Mock(),
         )
-        proxy_req.send = make_mocked_coro(proxy_resp)
-        proxy_resp.start = make_mocked_coro(mock.Mock(status=200))
+        with mock.patch.object(
+            proxy_req, "send", autospec=True, return_value=proxy_resp
+        ):
+            with mock.patch.object(proxy_resp, "start", autospec=True) as m:
+                m.return_value.status = 200
 
-        async def make_conn():
-            return aiohttp.TCPConnector()
+                async def make_conn() -> aiohttp.TCPConnector:
+                    return aiohttp.TCPConnector()
 
-        connector = self.loop.run_until_complete(make_conn())
-        connector._resolve_host = make_mocked_coro(
-            [
-                {
+                connector = self.loop.run_until_complete(make_conn())
+                r = {
                     "hostname": "hostname",
                     "host": "127.0.0.1",
                     "port": 80,
@@ -541,26 +617,37 @@ class TestProxy(unittest.TestCase):
                     "proto": 0,
                     "flags": 0,
                 }
-            ]
-        )
-
-        # Called on connection to http://proxy.example.com
-        self.loop.create_connection = make_mocked_coro(
-            (mock.Mock(), mock.Mock()),
-        )
-        # Called on connection to https://www.python.org
-        self.loop.start_tls = make_mocked_coro(raise_exception=ssl.SSLError)
-
-        req = ClientRequest(
-            "GET",
-            URL("https://www.python.org"),
-            proxy=URL("http://proxy.example.com"),
-            loop=self.loop,
-        )
-        with self.assertRaises(aiohttp.ClientConnectorSSLError):
-            self.loop.run_until_complete(
-                connector._create_connection(req, None, aiohttp.ClientTimeout())
-            )
+                with mock.patch.object(
+                    connector, "_resolve_host", autospec=True, return_value=[r]
+                ):
+                    tr, proto = mock.Mock(), mock.Mock()
+                    # Called on connection to http://proxy.example.com
+                    with mock.patch.object(
+                        self.loop,
+                        "create_connection",
+                        autospec=True,
+                        return_value=(tr, proto),
+                    ):
+                        # Called on connection to https://www.python.org
+                        with mock.patch.object(
+                            self.loop,
+                            "start_tls",
+                            autospec=True,
+                            side_effect=ssl.SSLError,
+                        ):
+                            req = ClientRequest(
+                                "GET",
+                                URL("https://www.python.org"),
+                                proxy=URL("http://proxy.example.com"),
+                                loop=self.loop,
+                            )
+                            with self.assertRaises(aiohttp.ClientConnectorSSLError):
+                                self.loop.run_until_complete(
+                                    connector._create_connection(
+                                        req, [], aiohttp.ClientTimeout()
+                                    )
+                                )
+                self.loop.run_until_complete(connector.close())
 
     @mock.patch("aiohttp.connector.ClientRequest")
     @mock.patch(
@@ -569,7 +656,7 @@ class TestProxy(unittest.TestCase):
         spec_set=True,
     )
     def test_https_connect_http_proxy_error(
-        self, start_connection: Any, ClientRequestMock: Any
+        self, start_connection: mock.Mock, ClientRequestMock: mock.Mock
     ) -> None:
         proxy_req = ClientRequest(
             "GET", URL("http://proxy.example.com"), loop=self.loop
@@ -580,23 +667,25 @@ class TestProxy(unittest.TestCase):
             "get",
             URL("http://proxy.example.com"),
             request_info=mock.Mock(),
-            writer=None,
+            writer=None,  # type: ignore[arg-type]
             continue100=None,
             timer=TimerNoop(),
             traces=[],
             loop=self.loop,
             session=mock.Mock(),
         )
-        proxy_req.send = make_mocked_coro(proxy_resp)
-        proxy_resp.start = make_mocked_coro(mock.Mock(status=400, reason="bad request"))
+        with mock.patch.object(
+            proxy_req, "send", autospec=True, return_value=proxy_resp
+        ):
+            with mock.patch.object(proxy_resp, "start", autospec=True) as m:
+                m.return_value.status = 400
+                m.return_value.reason = "bad request"
 
-        async def make_conn():
-            return aiohttp.TCPConnector()
+                async def make_conn() -> aiohttp.TCPConnector:
+                    return aiohttp.TCPConnector()
 
-        connector = self.loop.run_until_complete(make_conn())
-        connector._resolve_host = make_mocked_coro(
-            [
-                {
+                connector = self.loop.run_until_complete(make_conn())
+                r = {
                     "hostname": "hostname",
                     "host": "127.0.0.1",
                     "port": 80,
@@ -604,29 +693,37 @@ class TestProxy(unittest.TestCase):
                     "proto": 0,
                     "flags": 0,
                 }
-            ]
-        )
+                with mock.patch.object(
+                    connector, "_resolve_host", autospec=True, return_value=[r]
+                ):
+                    tr, proto = mock.Mock(), mock.Mock()
+                    tr.get_extra_info.return_value = None
+                    # Called on connection to http://proxy.example.com
+                    with mock.patch.object(
+                        self.loop,
+                        "create_connection",
+                        autospec=True,
+                        return_value=(tr, proto),
+                    ):
+                        req = ClientRequest(
+                            "GET",
+                            URL("https://www.python.org"),
+                            proxy=URL("http://proxy.example.com"),
+                            loop=self.loop,
+                        )
+                        with self.assertRaisesRegex(
+                            aiohttp.ClientHttpProxyError, "400, message='bad request'"
+                        ):
+                            self.loop.run_until_complete(
+                                connector._create_connection(
+                                    req, [], aiohttp.ClientTimeout()
+                                )
+                            )
 
-        tr, proto = mock.Mock(), mock.Mock()
-        tr.get_extra_info.return_value = None
-        self.loop.create_connection = make_mocked_coro((tr, proto))
-
-        req = ClientRequest(
-            "GET",
-            URL("https://www.python.org"),
-            proxy=URL("http://proxy.example.com"),
-            loop=self.loop,
-        )
-        with self.assertRaisesRegex(
-            aiohttp.ClientHttpProxyError, "400, message='bad request'"
-        ):
-            self.loop.run_until_complete(
-                connector._create_connection(req, None, aiohttp.ClientTimeout())
-            )
-
-        self.loop.run_until_complete(proxy_req.close())
-        proxy_resp.close()
-        self.loop.run_until_complete(req.close())
+                        self.loop.run_until_complete(proxy_req.close())
+                        proxy_resp.close()
+                        self.loop.run_until_complete(req.close())
+                self.loop.run_until_complete(connector.close())
 
     @mock.patch("aiohttp.connector.ClientRequest")
     @mock.patch(
@@ -635,7 +732,7 @@ class TestProxy(unittest.TestCase):
         spec_set=True,
     )
     def test_https_connect_resp_start_error(
-        self, start_connection: Any, ClientRequestMock: Any
+        self, start_connection: mock.Mock, ClientRequestMock: mock.Mock
     ) -> None:
         proxy_req = ClientRequest(
             "GET", URL("http://proxy.example.com"), loop=self.loop
@@ -646,23 +743,25 @@ class TestProxy(unittest.TestCase):
             "get",
             URL("http://proxy.example.com"),
             request_info=mock.Mock(),
-            writer=None,
+            writer=None,  # type: ignore[arg-type]
             continue100=None,
             timer=TimerNoop(),
             traces=[],
             loop=self.loop,
             session=mock.Mock(),
         )
-        proxy_req.send = make_mocked_coro(proxy_resp)
-        proxy_resp.start = make_mocked_coro(raise_exception=OSError("error message"))
+        with mock.patch.object(
+            proxy_req, "send", autospec=True, return_value=proxy_resp
+        ):
+            with mock.patch.object(
+                proxy_resp, "start", autospec=True, side_effect=OSError("error message")
+            ):
 
-        async def make_conn():
-            return aiohttp.TCPConnector()
+                async def make_conn() -> aiohttp.TCPConnector:
+                    return aiohttp.TCPConnector()
 
-        connector = self.loop.run_until_complete(make_conn())
-        connector._resolve_host = make_mocked_coro(
-            [
-                {
+                connector = self.loop.run_until_complete(make_conn())
+                r = {
                     "hostname": "hostname",
                     "host": "127.0.0.1",
                     "port": 80,
@@ -670,23 +769,31 @@ class TestProxy(unittest.TestCase):
                     "proto": 0,
                     "flags": 0,
                 }
-            ]
-        )
-
-        tr, proto = mock.Mock(), mock.Mock()
-        tr.get_extra_info.return_value = None
-        self.loop.create_connection = make_mocked_coro((tr, proto))
-
-        req = ClientRequest(
-            "GET",
-            URL("https://www.python.org"),
-            proxy=URL("http://proxy.example.com"),
-            loop=self.loop,
-        )
-        with self.assertRaisesRegex(OSError, "error message"):
-            self.loop.run_until_complete(
-                connector._create_connection(req, None, aiohttp.ClientTimeout())
-            )
+                with mock.patch.object(
+                    connector, "_resolve_host", autospec=True, return_value=[r]
+                ):
+                    tr, proto = mock.Mock(), mock.Mock()
+                    tr.get_extra_info.return_value = None
+                    # Called on connection to http://proxy.example.com
+                    with mock.patch.object(
+                        self.loop,
+                        "create_connection",
+                        autospec=True,
+                        return_value=(tr, proto),
+                    ):
+                        req = ClientRequest(
+                            "GET",
+                            URL("https://www.python.org"),
+                            proxy=URL("http://proxy.example.com"),
+                            loop=self.loop,
+                        )
+                        with self.assertRaisesRegex(OSError, "error message"):
+                            self.loop.run_until_complete(
+                                connector._create_connection(
+                                    req, [], aiohttp.ClientTimeout()
+                                )
+                            )
+                self.loop.run_until_complete(connector.close())
 
     @mock.patch("aiohttp.connector.ClientRequest")
     @mock.patch(
@@ -694,43 +801,46 @@ class TestProxy(unittest.TestCase):
         autospec=True,
         spec_set=True,
     )
-    def test_request_port(self, start_connection: Any, ClientRequestMock: Any) -> None:
+    def test_request_port(
+        self, start_connection: mock.Mock, ClientRequestMock: mock.Mock
+    ) -> None:
         proxy_req = ClientRequest(
             "GET", URL("http://proxy.example.com"), loop=self.loop
         )
         ClientRequestMock.return_value = proxy_req
 
-        async def make_conn():
+        async def make_conn() -> aiohttp.TCPConnector:
             return aiohttp.TCPConnector()
 
         connector = self.loop.run_until_complete(make_conn())
-        connector._resolve_host = make_mocked_coro(
-            [
-                {
-                    "hostname": "hostname",
-                    "host": "127.0.0.1",
-                    "port": 80,
-                    "family": socket.AF_INET,
-                    "proto": 0,
-                    "flags": 0,
-                }
-            ]
-        )
-
-        tr, proto = mock.Mock(), mock.Mock()
-        tr.get_extra_info.return_value = None
-        self.loop.create_connection = make_mocked_coro((tr, proto))
-
-        req = ClientRequest(
-            "GET",
-            URL("http://localhost:1234/path"),
-            proxy=URL("http://proxy.example.com"),
-            loop=self.loop,
-        )
-        self.loop.run_until_complete(
-            connector._create_connection(req, None, aiohttp.ClientTimeout())
-        )
-        self.assertEqual(req.url, URL("http://localhost:1234/path"))
+        r = {
+            "hostname": "hostname",
+            "host": "127.0.0.1",
+            "port": 80,
+            "family": socket.AF_INET,
+            "proto": 0,
+            "flags": 0,
+        }
+        with mock.patch.object(
+            connector, "_resolve_host", autospec=True, return_value=[r]
+        ):
+            tr, proto = mock.Mock(), mock.Mock()
+            tr.get_extra_info.return_value = None
+            # Called on connection to http://proxy.example.com
+            with mock.patch.object(
+                self.loop, "create_connection", autospec=True, return_value=(tr, proto)
+            ):
+                req = ClientRequest(
+                    "GET",
+                    URL("http://localhost:1234/path"),
+                    proxy=URL("http://proxy.example.com"),
+                    loop=self.loop,
+                )
+                self.loop.run_until_complete(
+                    connector._create_connection(req, [], aiohttp.ClientTimeout())
+                )
+                self.assertEqual(req.url, URL("http://localhost:1234/path"))
+        self.loop.run_until_complete(connector.close())
 
     def test_proxy_auth_property(self) -> None:
         req = aiohttp.ClientRequest(
@@ -758,7 +868,7 @@ class TestProxy(unittest.TestCase):
         spec_set=True,
     )
     def test_https_connect_pass_ssl_context(
-        self, start_connection: Any, ClientRequestMock: Any
+        self, start_connection: mock.Mock, ClientRequestMock: mock.Mock
     ) -> None:
         proxy_req = ClientRequest(
             "GET", URL("http://proxy.example.com"), loop=self.loop
@@ -769,23 +879,24 @@ class TestProxy(unittest.TestCase):
             "get",
             URL("http://proxy.example.com"),
             request_info=mock.Mock(),
-            writer=None,
+            writer=None,  # type: ignore[arg-type]
             continue100=None,
             timer=TimerNoop(),
             traces=[],
             loop=self.loop,
             session=mock.Mock(),
         )
-        proxy_req.send = make_mocked_coro(proxy_resp)
-        proxy_resp.start = make_mocked_coro(mock.Mock(status=200))
+        with mock.patch.object(
+            proxy_req, "send", autospec=True, return_value=proxy_resp
+        ):
+            with mock.patch.object(proxy_resp, "start", autospec=True) as m:
+                m.return_value.status = 200
 
-        async def make_conn():
-            return aiohttp.TCPConnector()
+                async def make_conn() -> aiohttp.TCPConnector:
+                    return aiohttp.TCPConnector()
 
-        connector = self.loop.run_until_complete(make_conn())
-        connector._resolve_host = make_mocked_coro(
-            [
-                {
+                connector = self.loop.run_until_complete(make_conn())
+                r = {
                     "hostname": "hostname",
                     "host": "127.0.0.1",
                     "port": 80,
@@ -793,38 +904,52 @@ class TestProxy(unittest.TestCase):
                     "proto": 0,
                     "flags": 0,
                 }
-            ]
-        )
+                with mock.patch.object(
+                    connector, "_resolve_host", autospec=True, return_value=[r]
+                ):
+                    tr, proto = mock.Mock(), mock.Mock()
+                    with mock.patch.object(
+                        self.loop,
+                        "create_connection",
+                        autospec=True,
+                        return_value=(tr, proto),
+                    ):
+                        with mock.patch.object(
+                            self.loop,
+                            "start_tls",
+                            autospec=True,
+                            return_value=mock.Mock(),
+                        ) as tls_m:
+                            req = ClientRequest(
+                                "GET",
+                                URL("https://www.python.org"),
+                                proxy=URL("http://proxy.example.com"),
+                                loop=self.loop,
+                            )
+                            self.loop.run_until_complete(
+                                connector._create_connection(
+                                    req, [], aiohttp.ClientTimeout()
+                                )
+                            )
 
-        tr, proto = mock.Mock(), mock.Mock()
-        self.loop.create_connection = make_mocked_coro((tr, proto))
-        self.loop.start_tls = make_mocked_coro(mock.Mock())
+                            tls_m.assert_called_with(
+                                mock.ANY,
+                                mock.ANY,
+                                _SSL_CONTEXT_VERIFIED,
+                                server_hostname="www.python.org",
+                                ssl_handshake_timeout=mock.ANY,
+                            )
 
-        req = ClientRequest(
-            "GET",
-            URL("https://www.python.org"),
-            proxy=URL("http://proxy.example.com"),
-            loop=self.loop,
-        )
-        self.loop.run_until_complete(
-            connector._create_connection(req, None, aiohttp.ClientTimeout())
-        )
+                            self.assertEqual(req.url.path, "/")
+                            self.assertEqual(proxy_req.method, "CONNECT")
+                            self.assertEqual(
+                                proxy_req.url, URL("https://www.python.org")
+                            )
 
-        self.loop.start_tls.assert_called_with(
-            mock.ANY,
-            mock.ANY,
-            connector._make_ssl_context(True),
-            server_hostname="www.python.org",
-            ssl_handshake_timeout=mock.ANY,
-        )
-
-        self.assertEqual(req.url.path, "/")
-        self.assertEqual(proxy_req.method, "CONNECT")
-        self.assertEqual(proxy_req.url, URL("https://www.python.org"))
-
-        self.loop.run_until_complete(proxy_req.close())
-        proxy_resp.close()
-        self.loop.run_until_complete(req.close())
+                            self.loop.run_until_complete(proxy_req.close())
+                            proxy_resp.close()
+                            self.loop.run_until_complete(req.close())
+                self.loop.run_until_complete(connector.close())
 
     @mock.patch("aiohttp.connector.ClientRequest")
     @mock.patch(
@@ -832,7 +957,9 @@ class TestProxy(unittest.TestCase):
         autospec=True,
         spec_set=True,
     )
-    def test_https_auth(self, start_connection: Any, ClientRequestMock: Any) -> None:
+    def test_https_auth(
+        self, start_connection: mock.Mock, ClientRequestMock: mock.Mock
+    ) -> None:
         proxy_req = ClientRequest(
             "GET",
             URL("http://proxy.example.com"),
@@ -845,23 +972,24 @@ class TestProxy(unittest.TestCase):
             "get",
             URL("http://proxy.example.com"),
             request_info=mock.Mock(),
-            writer=None,
+            writer=None,  # type: ignore[arg-type]
             continue100=None,
             timer=TimerNoop(),
             traces=[],
             loop=self.loop,
             session=mock.Mock(),
         )
-        proxy_req.send = make_mocked_coro(proxy_resp)
-        proxy_resp.start = make_mocked_coro(mock.Mock(status=200))
+        with mock.patch.object(
+            proxy_req, "send", autospec=True, return_value=proxy_resp
+        ):
+            with mock.patch.object(proxy_resp, "start", autospec=True) as m:
+                m.return_value.status = 200
 
-        async def make_conn():
-            return aiohttp.TCPConnector()
+                async def make_conn() -> aiohttp.TCPConnector:
+                    return aiohttp.TCPConnector()
 
-        connector = self.loop.run_until_complete(make_conn())
-        connector._resolve_host = make_mocked_coro(
-            [
-                {
+                connector = self.loop.run_until_complete(make_conn())
+                r = {
                     "hostname": "hostname",
                     "host": "127.0.0.1",
                     "port": 80,
@@ -869,38 +997,50 @@ class TestProxy(unittest.TestCase):
                     "proto": 0,
                     "flags": 0,
                 }
-            ]
-        )
+                with mock.patch.object(
+                    connector, "_resolve_host", autospec=True, return_value=[r]
+                ) as host_m:
+                    tr, proto = mock.Mock(), mock.Mock()
+                    with mock.patch.object(
+                        self.loop,
+                        "create_connection",
+                        autospec=True,
+                        return_value=(tr, proto),
+                    ):
+                        with mock.patch.object(
+                            self.loop,
+                            "start_tls",
+                            autospec=True,
+                            return_value=mock.Mock(),
+                        ):
+                            self.assertIn("AUTHORIZATION", proxy_req.headers)
+                            self.assertNotIn("PROXY-AUTHORIZATION", proxy_req.headers)
 
-        tr, proto = mock.Mock(), mock.Mock()
-        self.loop.create_connection = make_mocked_coro((tr, proto))
-        self.loop.start_tls = make_mocked_coro(mock.Mock())
+                            req = ClientRequest(
+                                "GET",
+                                URL("https://www.python.org"),
+                                proxy=URL("http://proxy.example.com"),
+                                loop=self.loop,
+                            )
+                            self.assertNotIn("AUTHORIZATION", req.headers)
+                            self.assertNotIn("PROXY-AUTHORIZATION", req.headers)
+                            self.loop.run_until_complete(
+                                connector._create_connection(
+                                    req, [], aiohttp.ClientTimeout()
+                                )
+                            )
 
-        self.assertIn("AUTHORIZATION", proxy_req.headers)
-        self.assertNotIn("PROXY-AUTHORIZATION", proxy_req.headers)
+                            self.assertEqual(req.url.path, "/")
+                            self.assertNotIn("AUTHORIZATION", req.headers)
+                            self.assertNotIn("PROXY-AUTHORIZATION", req.headers)
+                            self.assertNotIn("AUTHORIZATION", proxy_req.headers)
+                            self.assertIn("PROXY-AUTHORIZATION", proxy_req.headers)
 
-        req = ClientRequest(
-            "GET",
-            URL("https://www.python.org"),
-            proxy=URL("http://proxy.example.com"),
-            loop=self.loop,
-        )
-        self.assertNotIn("AUTHORIZATION", req.headers)
-        self.assertNotIn("PROXY-AUTHORIZATION", req.headers)
-        self.loop.run_until_complete(
-            connector._create_connection(req, None, aiohttp.ClientTimeout())
-        )
+                            host_m.assert_called_with(
+                                "proxy.example.com", 80, traces=mock.ANY
+                            )
 
-        self.assertEqual(req.url.path, "/")
-        self.assertNotIn("AUTHORIZATION", req.headers)
-        self.assertNotIn("PROXY-AUTHORIZATION", req.headers)
-        self.assertNotIn("AUTHORIZATION", proxy_req.headers)
-        self.assertIn("PROXY-AUTHORIZATION", proxy_req.headers)
-
-        connector._resolve_host.assert_called_with(
-            "proxy.example.com", 80, traces=mock.ANY
-        )
-
-        self.loop.run_until_complete(proxy_req.close())
-        proxy_resp.close()
-        self.loop.run_until_complete(req.close())
+                            self.loop.run_until_complete(proxy_req.close())
+                            proxy_resp.close()
+                            self.loop.run_until_complete(req.close())
+                self.loop.run_until_complete(connector.close())
