@@ -1,7 +1,7 @@
-# type: ignore
+import asyncio
 import gzip
-from socket import socket
-from typing import Any
+import socket
+from typing import Callable, Iterator, List, Mapping, NoReturn
 from unittest import mock
 
 import pytest
@@ -12,24 +12,26 @@ import aiohttp
 from aiohttp import web
 from aiohttp.test_utils import (
     AioHTTPTestCase,
-    RawTestServer as _RawTestServer,
-    TestClient as _TestClient,
-    TestServer as _TestServer,
+    RawTestServer,
+    TestClient,
+    TestServer,
     get_port_socket,
     loop_context,
     make_mocked_request,
 )
+
+_TestClient = TestClient[web.Request]
 
 _hello_world_str = "Hello, world"
 _hello_world_bytes = _hello_world_str.encode("utf-8")
 _hello_world_gz = gzip.compress(_hello_world_bytes)
 
 
-def _create_example_app():
-    async def hello(request):
+def _create_example_app() -> web.Application:
+    async def hello(request: web.Request) -> web.Response:
         return web.Response(body=_hello_world_bytes)
 
-    async def websocket_handler(request):
+    async def websocket_handler(request: web.Request) -> web.WebSocketResponse:
         ws = web.WebSocketResponse()
         await ws.prepare(request)
         msg = await ws.receive()
@@ -41,7 +43,7 @@ def _create_example_app():
 
         return ws
 
-    async def cookie_handler(request):
+    async def cookie_handler(request: web.Request) -> web.Response:
         resp = web.Response(body=_hello_world_bytes)
         resp.set_cookie("cookie", "val")
         return resp
@@ -55,20 +57,22 @@ def _create_example_app():
 
 # these exist to test the pytest scenario
 @pytest.fixture
-def loop() -> None:
+def loop() -> Iterator[asyncio.AbstractEventLoop]:
     with loop_context() as loop:
         yield loop
 
 
 @pytest.fixture
-def app():
+def app() -> web.Application:
     return _create_example_app()
 
 
 @pytest.fixture
-def test_client(loop: Any, app: Any) -> None:
-    async def make_client():
-        return _TestClient(_TestServer(app))
+def test_client(
+    loop: asyncio.AbstractEventLoop, app: web.Application
+) -> Iterator[_TestClient]:
+    async def make_client() -> TestClient[web.Request]:
+        return TestClient(TestServer(app))
 
     client = loop.run_until_complete(make_client())
 
@@ -81,13 +85,13 @@ async def test_aiohttp_client_close_is_idempotent() -> None:
     # a test client, called multiple times, should
     # not attempt to close the server again.
     app = _create_example_app()
-    client = _TestClient(_TestServer(app))
+    client = TestClient(TestServer(app))
     await client.close()
     await client.close()
 
 
 class TestAioHTTPTestCase(AioHTTPTestCase):
-    async def get_application(self):
+    async def get_application(self) -> web.Application:
         return _create_example_app()
 
     async def test_example_with_loop(self) -> None:
@@ -112,7 +116,7 @@ class TestAioHTTPTestCase(AioHTTPTestCase):
         await test_get_route()
 
 
-def test_get_route(loop: Any, test_client: Any) -> None:
+def test_get_route(loop: asyncio.AbstractEventLoop, test_client: _TestClient) -> None:
     async def test_get_route() -> None:
         resp = await test_client.request("GET", "/")
         assert resp.status == 200
@@ -122,7 +126,9 @@ def test_get_route(loop: Any, test_client: Any) -> None:
     loop.run_until_complete(test_get_route())
 
 
-async def test_client_websocket(loop: Any, test_client: Any) -> None:
+async def test_client_websocket(
+    loop: asyncio.AbstractEventLoop, test_client: _TestClient
+) -> None:
     resp = await test_client.ws_connect("/websocket")
     await resp.send_str("foo")
     msg = await resp.receive()
@@ -133,7 +139,9 @@ async def test_client_websocket(loop: Any, test_client: Any) -> None:
     assert msg.type == aiohttp.WSMsgType.CLOSE
 
 
-async def test_client_cookie(loop: Any, test_client: Any) -> None:
+async def test_client_cookie(
+    loop: asyncio.AbstractEventLoop, test_client: _TestClient
+) -> None:
     assert not test_client.session.cookie_jar
     await test_client.get("/cookie")
     cookies = list(test_client.session.cookie_jar)
@@ -144,20 +152,24 @@ async def test_client_cookie(loop: Any, test_client: Any) -> None:
 @pytest.mark.parametrize(
     "method", ["get", "post", "options", "post", "put", "patch", "delete"]
 )
-async def test_test_client_methods(method: Any, loop: Any, test_client: Any) -> None:
+async def test_test_client_methods(
+    method: str, loop: asyncio.AbstractEventLoop, test_client: _TestClient
+) -> None:
     resp = await getattr(test_client, method)("/")
     assert resp.status == 200
     text = await resp.text()
     assert _hello_world_str == text
 
 
-async def test_test_client_head(loop: Any, test_client: Any) -> None:
+async def test_test_client_head(
+    loop: asyncio.AbstractEventLoop, test_client: _TestClient
+) -> None:
     resp = await test_client.head("/")
     assert resp.status == 200
 
 
 @pytest.mark.parametrize("headers", [{"token": "x"}, CIMultiDict({"token": "x"}), {}])
-def test_make_mocked_request(headers: Any) -> None:
+def test_make_mocked_request(headers: List[Mapping[str, str]]) -> None:
     req = make_mocked_request("GET", "/", headers=headers)
     assert req.method == "GET"
     assert req.path == "/"
@@ -167,11 +179,13 @@ def test_make_mocked_request(headers: Any) -> None:
 
 def test_make_mocked_request_sslcontext() -> None:
     req = make_mocked_request("GET", "/")
+    assert req.transport is not None
     assert req.transport.get_extra_info("sslcontext") is None
 
 
 def test_make_mocked_request_unknown_extra_info() -> None:
     req = make_mocked_request("GET", "/")
+    assert req.transport is not None
     assert req.transport.get_extra_info("unknown_extra_info") is None
 
 
@@ -190,7 +204,7 @@ def test_make_mocked_request_app_can_store_values() -> None:
 def test_make_mocked_request_app_access_non_existing() -> None:
     req = make_mocked_request("GET", "/")
     with pytest.raises(AttributeError):
-        req.app.foo
+        req.app.foo  # type: ignore[attr-defined]
 
 
 def test_make_mocked_request_match_info() -> None:
@@ -204,6 +218,11 @@ def test_make_mocked_request_content() -> None:
     assert req.content is payload
 
 
+async def test_make_mocked_request_empty_payload() -> None:
+    req = make_mocked_request("GET", "/")
+    assert await req.read() == b""
+
+
 def test_make_mocked_request_transport() -> None:
     transport = mock.Mock()
     req = make_mocked_request("GET", "/", transport=transport)
@@ -212,8 +231,8 @@ def test_make_mocked_request_transport() -> None:
 
 async def test_test_client_props() -> None:
     app = _create_example_app()
-    server = _TestServer(app, scheme="http", host="127.0.0.1")
-    client = _TestClient(server)
+    server = TestServer(app, scheme="http", host="127.0.0.1")
+    client = TestClient(server)
     assert client.scheme == "http"
     assert client.host == "127.0.0.1"
     assert client.port is None
@@ -225,11 +244,11 @@ async def test_test_client_props() -> None:
 
 
 async def test_test_client_raw_server_props() -> None:
-    async def hello(request):
-        return web.Response()  # pragma: no cover
+    async def hello(request: web.BaseRequest) -> NoReturn:
+        assert False
 
-    server = _RawTestServer(hello, scheme="http", host="127.0.0.1")
-    client = _TestClient(server)
+    server = RawTestServer(hello, scheme="http", host="127.0.0.1")
+    client = TestClient(server)
     assert client.scheme == "http"
     assert client.host == "127.0.0.1"
     assert client.port is None
@@ -240,9 +259,9 @@ async def test_test_client_raw_server_props() -> None:
     assert client.port is None
 
 
-async def test_test_server_context_manager(loop: Any) -> None:
+async def test_test_server_context_manager(loop: asyncio.AbstractEventLoop) -> None:
     app = _create_example_app()
-    async with _TestServer(app) as server:
+    async with TestServer(app) as server:
         client = aiohttp.ClientSession()
         resp = await client.head(server.make_url("/"))
         assert resp.status == 200
@@ -252,16 +271,18 @@ async def test_test_server_context_manager(loop: Any) -> None:
 
 def test_client_unsupported_arg() -> None:
     with pytest.raises(TypeError) as e:
-        _TestClient("string")
+        TestClient("string")  # type: ignore[arg-type]
 
     assert (
         str(e.value) == "server must be TestServer instance, found type: <class 'str'>"
     )
 
 
-async def test_server_make_url_yarl_compatibility(loop: Any) -> None:
+async def test_server_make_url_yarl_compatibility(
+    loop: asyncio.AbstractEventLoop,
+) -> None:
     app = _create_example_app()
-    async with _TestServer(app) as server:
+    async with TestServer(app) as server:
         make_url = server.make_url
         assert make_url(URL("/foo")) == make_url("/foo")
         with pytest.raises(AssertionError):
@@ -270,7 +291,9 @@ async def test_server_make_url_yarl_compatibility(loop: Any) -> None:
             make_url(URL("http://foo.com"))
 
 
-def test_testcase_no_app(testdir: Any, loop: Any) -> None:
+def test_testcase_no_app(
+    testdir: pytest.Testdir, loop: asyncio.AbstractEventLoop
+) -> None:
     testdir.makepyfile(
         """
         from aiohttp.test_utils import AioHTTPTestCase
@@ -285,8 +308,10 @@ def test_testcase_no_app(testdir: Any, loop: Any) -> None:
     result.stdout.fnmatch_lines(["*TypeError*"])
 
 
-async def test_server_context_manager(app: Any, loop: Any) -> None:
-    async with _TestServer(app) as server:
+async def test_server_context_manager(
+    app: web.Application, loop: asyncio.AbstractEventLoop
+) -> None:
+    async with TestServer(app) as server:
         async with aiohttp.ClientSession() as client:
             async with client.head(server.make_url("/")) as resp:
                 assert resp.status == 200
@@ -296,9 +321,9 @@ async def test_server_context_manager(app: Any, loop: Any) -> None:
     "method", ["head", "get", "post", "options", "post", "put", "patch", "delete"]
 )
 async def test_client_context_manager_response(
-    method: Any, app: Any, loop: Any
+    method: str, app: web.Application, loop: asyncio.AbstractEventLoop
 ) -> None:
-    async with _TestClient(_TestServer(app)) as client:
+    async with TestClient(TestServer(app)) as client:
         async with getattr(client, method)("/") as resp:
             assert resp.status == 200
             if method != "head":
@@ -306,9 +331,13 @@ async def test_client_context_manager_response(
                 assert "Hello, world" in text
 
 
-async def test_custom_port(loop: Any, app: Any, aiohttp_unused_port: Any) -> None:
+async def test_custom_port(
+    loop: asyncio.AbstractEventLoop,
+    app: web.Application,
+    aiohttp_unused_port: Callable[[], int],
+) -> None:
     port = aiohttp_unused_port()
-    client = _TestClient(_TestServer(app, port=port))
+    client = TestClient(TestServer(app, port=port))
     await client.start_server()
 
     assert client.server.port == port
@@ -326,25 +355,25 @@ async def test_custom_port(loop: Any, app: Any, aiohttp_unused_port: Any) -> Non
     [("127.0.0.1", "127.0.0.1"), ("localhost", "127.0.0.1"), ("::1", "::1")],
 )
 async def test_test_server_hostnames(
-    hostname: Any, expected_host: Any, loop: Any
+    hostname: str, expected_host: str, loop: asyncio.AbstractEventLoop
 ) -> None:
     app = _create_example_app()
-    server = _TestServer(app, host=hostname, loop=loop)
+    server = TestServer(app, host=hostname, loop=loop)
     async with server:
         pass
     assert server.host == expected_host
 
 
-@pytest.mark.parametrize("test_server_cls", [_TestServer, _RawTestServer])
+@pytest.mark.parametrize("test_server_cls", [TestServer, RawTestServer])
 async def test_base_test_server_socket_factory(
-    test_server_cls: type, app: Any, loop: Any
+    test_server_cls: type, app: web.Application, loop: asyncio.AbstractEventLoop
 ) -> None:
     factory_called = False
 
-    def factory(*args, **kwargs) -> socket:
+    def factory(host: str, port: int, family: socket.AddressFamily) -> socket.socket:
         nonlocal factory_called
         factory_called = True
-        return get_port_socket(*args, **kwargs)
+        return get_port_socket(host, port, family)
 
     server = test_server_cls(app, loop=loop, socket_factory=factory)
     async with server:
