@@ -1,6 +1,5 @@
-# type: ignore
 from types import SimpleNamespace
-from typing import Any
+from typing import Any, Tuple
 from unittest.mock import Mock
 
 import pytest
@@ -103,7 +102,9 @@ class TestTrace:
             ("dns_cache_miss", (Mock(),), TraceDnsCacheMissParams),
         ],
     )
-    async def test_send(self, signal: Any, params: Any, param_obj: Any) -> None:
+    async def test_send(
+        self, signal: str, params: Tuple[Mock, ...], param_obj: Any
+    ) -> None:
         session = Mock()
         trace_request_ctx = Mock()
         callback = Mock(side_effect=make_mocked_coro(Mock()))
