@@ -1,61 +1,53 @@
-# type: ignore
+import asyncio
 import pathlib
 import re
 from collections.abc import Container, Iterable, Mapping, MutableMapping, Sized
 from functools import partial
-from typing import Any
+from typing import Awaitable, Callable, Dict, List, NoReturn, Optional, Type
 from urllib.parse import unquote
 
 import pytest
-from re_assert import Matches
 from yarl import URL
 
 import aiohttp
 from aiohttp import hdrs, web
 from aiohttp.test_utils import make_mocked_request
-from aiohttp.web import HTTPMethodNotAllowed, HTTPNotFound, Response
 from aiohttp.web_urldispatcher import (
     PATH_SEP,
-    AbstractResource,
     Domain,
-    DynamicResource,
     MaskDomain,
-    PlainResource,
-    ResourceRoute,
-    StaticResource,
     SystemRoute,
-    View,
     _default_expect_handler,
 )
 
 
-def make_handler():
-    async def handler(request):
-        return Response(request)  # pragma: no cover
+def make_handler() -> Callable[[web.Request], Awaitable[NoReturn]]:
+    async def handler(request: web.Request) -> NoReturn:
+        assert False
 
     return handler
 
 
-def make_partial_handler():
-    async def handler(a, request):
-        return Response(request)  # pragma: no cover
+def make_partial_handler() -> Callable[[web.Request], Awaitable[NoReturn]]:
+    async def handler(a: int, request: web.Request) -> NoReturn:
+        assert False
 
     return partial(handler, 5)
 
 
 @pytest.fixture
-def app():
+def app() -> web.Application:
     return web.Application()
 
 
 @pytest.fixture
-def router(app: Any):
+def router(app: web.Application) -> web.UrlDispatcher:
     return app.router
 
 
 @pytest.fixture
-def fill_routes(router: Any):
-    def go():
+def fill_routes(router: web.UrlDispatcher) -> Callable[[], List[web.AbstractRoute]]:
+    def go() -> List[web.AbstractRoute]:
         route1 = router.add_route("GET", "/plain", make_handler())
         route2 = router.add_route("GET", "/variable/{name}", make_handler())
         resource = router.add_static("/static", pathlib.Path(aiohttp.__file__).parent)
@@ -64,7 +56,7 @@ def fill_routes(router: Any):
     return go
 
 
-def test_register_uncommon_http_methods(router: Any) -> None:
+def test_register_uncommon_http_methods(router: web.UrlDispatcher) -> None:
     uncommon_http_methods = {
         "PROPFIND",
         "PROPPATCH",
@@ -81,20 +73,20 @@ def test_register_uncommon_http_methods(router: Any) -> None:
         router.add_route(method, "/handler/to/path", make_handler())
 
 
-async def test_add_partial_handler(router: Any) -> None:
+async def test_add_partial_handler(router: web.UrlDispatcher) -> None:
     handler = make_partial_handler()
     router.add_get("/handler/to/path", handler)
 
 
-async def test_add_sync_handler(router: Any) -> None:
-    def handler(request):
-        pass
+async def test_add_sync_handler(router: web.UrlDispatcher) -> None:
+    def handler(request: web.Request) -> NoReturn:
+        assert False
 
     with pytest.raises(TypeError):
         router.add_get("/handler/to/path", handler)
 
 
-async def test_add_route_root(router: Any) -> None:
+async def test_add_route_root(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/", handler)
     req = make_mocked_request("GET", "/")
@@ -105,7 +97,7 @@ async def test_add_route_root(router: Any) -> None:
     assert info.route.name is None
 
 
-async def test_add_route_simple(router: Any) -> None:
+async def test_add_route_simple(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/handler/to/path", handler)
     req = make_mocked_request("GET", "/handler/to/path")
@@ -116,7 +108,7 @@ async def test_add_route_simple(router: Any) -> None:
     assert info.route.name is None
 
 
-async def test_add_with_matchdict(router: Any) -> None:
+async def test_add_with_matchdict(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/handler/{to}", handler)
     req = make_mocked_request("GET", "/handler/tail")
@@ -127,7 +119,7 @@ async def test_add_with_matchdict(router: Any) -> None:
     assert info.route.name is None
 
 
-async def test_add_with_matchdict_with_colon(router: Any) -> None:
+async def test_add_with_matchdict_with_colon(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/handler/{to}", handler)
     req = make_mocked_request("GET", "/handler/1:2:3")
@@ -138,7 +130,7 @@ async def test_add_with_matchdict_with_colon(router: Any) -> None:
     assert info.route.name is None
 
 
-async def test_add_route_with_add_get_shortcut(router: Any) -> None:
+async def test_add_route_with_add_get_shortcut(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_get("/handler/to/path", handler)
     req = make_mocked_request("GET", "/handler/to/path")
@@ -149,7 +141,7 @@ async def test_add_route_with_add_get_shortcut(router: Any) -> None:
     assert info.route.name is None
 
 
-async def test_add_route_with_add_post_shortcut(router: Any) -> None:
+async def test_add_route_with_add_post_shortcut(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_post("/handler/to/path", handler)
     req = make_mocked_request("POST", "/handler/to/path")
@@ -160,7 +152,7 @@ async def test_add_route_with_add_post_shortcut(router: Any) -> None:
     assert info.route.name is None
 
 
-async def test_add_route_with_add_put_shortcut(router: Any) -> None:
+async def test_add_route_with_add_put_shortcut(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_put("/handler/to/path", handler)
     req = make_mocked_request("PUT", "/handler/to/path")
@@ -171,7 +163,7 @@ async def test_add_route_with_add_put_shortcut(router: Any) -> None:
     assert info.route.name is None
 
 
-async def test_add_route_with_add_patch_shortcut(router: Any) -> None:
+async def test_add_route_with_add_patch_shortcut(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_patch("/handler/to/path", handler)
     req = make_mocked_request("PATCH", "/handler/to/path")
@@ -182,7 +174,7 @@ async def test_add_route_with_add_patch_shortcut(router: Any) -> None:
     assert info.route.name is None
 
 
-async def test_add_route_with_add_delete_shortcut(router: Any) -> None:
+async def test_add_route_with_add_delete_shortcut(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_delete("/handler/to/path", handler)
     req = make_mocked_request("DELETE", "/handler/to/path")
@@ -193,7 +185,7 @@ async def test_add_route_with_add_delete_shortcut(router: Any) -> None:
     assert info.route.name is None
 
 
-async def test_add_route_with_add_head_shortcut(router: Any) -> None:
+async def test_add_route_with_add_head_shortcut(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_head("/handler/to/path", handler)
     req = make_mocked_request("HEAD", "/handler/to/path")
@@ -204,7 +196,7 @@ async def test_add_route_with_add_head_shortcut(router: Any) -> None:
     assert info.route.name is None
 
 
-async def test_add_with_name(router: Any) -> None:
+async def test_add_with_name(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/handler/to/path", handler, name="name")
     req = make_mocked_request("GET", "/handler/to/path")
@@ -213,7 +205,7 @@ async def test_add_with_name(router: Any) -> None:
     assert "name" == info.route.name
 
 
-async def test_add_with_tailing_slash(router: Any) -> None:
+async def test_add_with_tailing_slash(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/handler/to/path/", handler)
     req = make_mocked_request("GET", "/handler/to/path/")
@@ -223,37 +215,37 @@ async def test_add_with_tailing_slash(router: Any) -> None:
     assert handler is info.handler
 
 
-def test_add_invalid_path(router: Any) -> None:
+def test_add_invalid_path(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     with pytest.raises(ValueError):
         router.add_route("GET", "/{/", handler)
 
 
-def test_add_url_invalid1(router: Any) -> None:
+def test_add_url_invalid1(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     with pytest.raises(ValueError):
         router.add_route("post", "/post/{id", handler)
 
 
-def test_add_url_invalid2(router: Any) -> None:
+def test_add_url_invalid2(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     with pytest.raises(ValueError):
         router.add_route("post", "/post/{id{}}", handler)
 
 
-def test_add_url_invalid3(router: Any) -> None:
+def test_add_url_invalid3(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     with pytest.raises(ValueError):
         router.add_route("post", "/post/{id{}", handler)
 
 
-def test_add_url_invalid4(router: Any) -> None:
+def test_add_url_invalid4(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     with pytest.raises(ValueError):
         router.add_route("post", '/post/{id"}', handler)
 
 
-async def test_add_url_escaping(router: Any) -> None:
+async def test_add_url_escaping(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/+$", handler)
 
@@ -263,7 +255,7 @@ async def test_add_url_escaping(router: Any) -> None:
     assert handler is info.handler
 
 
-async def test_any_method(router: Any) -> None:
+async def test_any_method(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     route = router.add_route(hdrs.METH_ANY, "/", handler)
 
@@ -279,7 +271,7 @@ async def test_any_method(router: Any) -> None:
     assert info1.route is info2.route
 
 
-async def test_match_second_result_in_table(router: Any) -> None:
+async def test_match_second_result_in_table(router: web.UrlDispatcher) -> None:
     handler1 = make_handler()
     handler2 = make_handler()
     router.add_route("GET", "/h1", handler1)
@@ -291,7 +283,7 @@ async def test_match_second_result_in_table(router: Any) -> None:
     assert handler2 is info.handler
 
 
-async def test_raise_method_not_allowed(router: Any) -> None:
+async def test_raise_method_not_allowed(router: web.UrlDispatcher) -> None:
     handler1 = make_handler()
     handler2 = make_handler()
     router.add_route("GET", "/", handler1)
@@ -302,7 +294,7 @@ async def test_raise_method_not_allowed(router: Any) -> None:
     assert isinstance(match_info.route, SystemRoute)
     assert {} == match_info
 
-    with pytest.raises(HTTPMethodNotAllowed) as ctx:
+    with pytest.raises(web.HTTPMethodNotAllowed) as ctx:
         await match_info.handler(req)
 
     exc = ctx.value
@@ -311,7 +303,7 @@ async def test_raise_method_not_allowed(router: Any) -> None:
     assert {"POST", "GET"} == exc.allowed_methods
 
 
-async def test_raise_method_not_found(router: Any) -> None:
+async def test_raise_method_not_found(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/a", handler)
     req = make_mocked_request("GET", "/b")
@@ -320,25 +312,24 @@ async def test_raise_method_not_found(router: Any) -> None:
     assert isinstance(match_info.route, SystemRoute)
     assert {} == match_info
 
-    with pytest.raises(HTTPNotFound) as ctx:
+    with pytest.raises(web.HTTPNotFound) as ctx:
         await match_info.handler(req)
 
     exc = ctx.value
     assert 404 == exc.status
 
 
-def test_double_add_url_with_the_same_name(router: Any) -> None:
+def test_double_add_url_with_the_same_name(router: web.UrlDispatcher) -> None:
     handler1 = make_handler()
     handler2 = make_handler()
     router.add_route("GET", "/get", handler1, name="name")
 
-    regexp = "Duplicate 'name', already handled by"
     with pytest.raises(ValueError) as ctx:
         router.add_route("GET", "/get_other", handler2, name="name")
-    assert Matches(regexp) == str(ctx.value)
+    assert str(ctx.value).startswith("Duplicate 'name', already handled by")
 
 
-def test_route_plain(router: Any) -> None:
+def test_route_plain(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     route = router.add_route("GET", "/get", handler, name="name")
     route2 = next(iter(router["name"]))
@@ -347,12 +338,12 @@ def test_route_plain(router: Any) -> None:
     assert route is route2
 
 
-def test_route_unknown_route_name(router: Any) -> None:
+def test_route_unknown_route_name(router: web.UrlDispatcher) -> None:
     with pytest.raises(KeyError):
         router["unknown"]
 
 
-def test_route_dynamic(router: Any) -> None:
+def test_route_dynamic(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     route = router.add_route("GET", "/get/{name}", handler, name="name")
 
@@ -362,7 +353,9 @@ def test_route_dynamic(router: Any) -> None:
     assert route is route2
 
 
-def test_add_static_path_checks(router: Any, tmp_path: pathlib.Path) -> None:
+def test_add_static_path_checks(
+    router: web.UrlDispatcher, tmp_path: pathlib.Path
+) -> None:
     """Test that static paths must exist and be directories."""
     with pytest.raises(ValueError, match="does not exist"):
         router.add_static("/", tmp_path / "does-not-exist")
@@ -370,14 +363,14 @@ def test_add_static_path_checks(router: Any, tmp_path: pathlib.Path) -> None:
             router.add_static("/", __file__)
 
 
-def test_add_static_path_resolution(router: Any) -> None:
+def test_add_static_path_resolution(router: web.UrlDispatcher) -> None:
     """Test that static paths are expanded and absolute."""
     res = router.add_static("/", "~/..")
     directory = str(res.get_info()["directory"])
     assert directory == str(pathlib.Path.home().parent)
 
 
-def test_add_static(router: Any) -> None:
+def test_add_static(router: web.UrlDispatcher) -> None:
     resource = router.add_static(
         "/st", pathlib.Path(aiohttp.__file__).parent, name="static"
     )
@@ -387,27 +380,31 @@ def test_add_static(router: Any) -> None:
     assert len(resource) == 2
 
 
-def test_add_static_append_version(router: Any) -> None:
+def test_add_static_append_version(router: web.UrlDispatcher) -> None:
     resource = router.add_static("/st", pathlib.Path(__file__).parent, name="static")
     url = resource.url_for(filename="/data.unknown_mime_type", append_version=True)
     expect_url = (
-        "/st/data.unknown_mime_type?" "v=aUsn8CHEhhszc81d28QmlcBW0KQpfS2F4trgQKhOYd8%3D"
+        "/st/data.unknown_mime_type?v=aUsn8CHEhhszc81d28QmlcBW0KQpfS2F4trgQKhOYd8%3D"
     )
     assert expect_url == str(url)
 
 
-def test_add_static_append_version_set_from_constructor(router: Any) -> None:
+def test_add_static_append_version_set_from_constructor(
+    router: web.UrlDispatcher,
+) -> None:
     resource = router.add_static(
         "/st", pathlib.Path(__file__).parent, append_version=True, name="static"
     )
     url = resource.url_for(filename="/data.unknown_mime_type")
     expect_url = (
-        "/st/data.unknown_mime_type?" "v=aUsn8CHEhhszc81d28QmlcBW0KQpfS2F4trgQKhOYd8%3D"
+        "/st/data.unknown_mime_type?v=aUsn8CHEhhszc81d28QmlcBW0KQpfS2F4trgQKhOYd8%3D"
     )
     assert expect_url == str(url)
 
 
-def test_add_static_append_version_override_constructor(router: Any) -> None:
+def test_add_static_append_version_override_constructor(
+    router: web.UrlDispatcher,
+) -> None:
     resource = router.add_static(
         "/st", pathlib.Path(__file__).parent, append_version=True, name="static"
     )
@@ -416,28 +413,34 @@ def test_add_static_append_version_override_constructor(router: Any) -> None:
     assert expect_url == str(url)
 
 
-def test_add_static_append_version_filename_without_slash(router: Any) -> None:
+def test_add_static_append_version_filename_without_slash(
+    router: web.UrlDispatcher,
+) -> None:
     resource = router.add_static("/st", pathlib.Path(__file__).parent, name="static")
     url = resource.url_for(filename="data.unknown_mime_type", append_version=True)
     expect_url = (
-        "/st/data.unknown_mime_type?" "v=aUsn8CHEhhszc81d28QmlcBW0KQpfS2F4trgQKhOYd8%3D"
+        "/st/data.unknown_mime_type?v=aUsn8CHEhhszc81d28QmlcBW0KQpfS2F4trgQKhOYd8%3D"
     )
     assert expect_url == str(url)
 
 
-def test_add_static_append_version_non_exists_file(router: Any) -> None:
+def test_add_static_append_version_non_exists_file(router: web.UrlDispatcher) -> None:
     resource = router.add_static("/st", pathlib.Path(__file__).parent, name="static")
     url = resource.url_for(filename="/non_exists_file", append_version=True)
     assert "/st/non_exists_file" == str(url)
 
 
-def test_add_static_append_version_non_exists_file_without_slash(router: Any) -> None:
+def test_add_static_append_version_non_exists_file_without_slash(
+    router: web.UrlDispatcher,
+) -> None:
     resource = router.add_static("/st", pathlib.Path(__file__).parent, name="static")
     url = resource.url_for(filename="non_exists_file", append_version=True)
     assert "/st/non_exists_file" == str(url)
 
 
-def test_add_static_append_version_follow_symlink(router: Any, tmp_path: Any) -> None:
+def test_add_static_append_version_follow_symlink(
+    router: web.UrlDispatcher, tmp_path: pathlib.Path
+) -> None:
     # Tests the access to a symlink, in static folder with apeend_version
     symlink_path = tmp_path / "append_version_symlink"
     symlink_target_path = pathlib.Path(__file__).parent
@@ -458,7 +461,7 @@ def test_add_static_append_version_follow_symlink(router: Any, tmp_path: Any) ->
 
 
 def test_add_static_append_version_not_follow_symlink(
-    router: Any, tmp_path: Any
+    router: web.UrlDispatcher, tmp_path: pathlib.Path
 ) -> None:
     # Tests the access to a symlink, in static folder with apeend_version
 
@@ -477,13 +480,13 @@ def test_add_static_append_version_not_follow_symlink(
     assert "/st/append_version_symlink/data.unknown_mime_type" == str(url)
 
 
-def test_add_static_quoting(router: Any) -> None:
+def test_add_static_quoting(router: web.UrlDispatcher) -> None:
     resource = router.add_static(
         "/пре %2Fфикс", pathlib.Path(aiohttp.__file__).parent, name="static"
     )
     assert router["static"] is resource
     url = resource.url_for(filename="/1 2/файл%2F.txt")
-    assert url.path == "/пре /фикс/1 2/файл%2F.txt"
+    assert url.path == "/пре %2Fфикс/1 2/файл%2F.txt"
     assert str(url) == (
         "/%D0%BF%D1%80%D0%B5%20%2F%D1%84%D0%B8%D0%BA%D1%81"
         "/1%202/%D1%84%D0%B0%D0%B9%D0%BB%252F.txt"
@@ -491,49 +494,52 @@ def test_add_static_quoting(router: Any) -> None:
     assert len(resource) == 2
 
 
-def test_plain_not_match(router: Any) -> None:
+def test_plain_not_match(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/get/path", handler, name="name")
     route = router["name"]
+    assert isinstance(route, web.Resource)
     assert route._match("/another/path") is None
 
 
-def test_dynamic_not_match(router: Any) -> None:
+def test_dynamic_not_match(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/get/{name}", handler, name="name")
     route = router["name"]
+    assert isinstance(route, web.Resource)
     assert route._match("/another/path") is None
 
 
-async def test_static_not_match(router: Any) -> None:
+async def test_static_not_match(router: web.UrlDispatcher) -> None:
     router.add_static("/pre", pathlib.Path(aiohttp.__file__).parent, name="name")
     resource = router["name"]
     ret = await resource.resolve(make_mocked_request("GET", "/another/path"))
     assert (None, set()) == ret
 
 
-def test_dynamic_with_trailing_slash(router: Any) -> None:
+def test_dynamic_with_trailing_slash(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/get/{name}/", handler, name="name")
     route = router["name"]
+    assert isinstance(route, web.Resource)
     assert {"name": "John"} == route._match("/get/John/")
 
 
-def test_len(router: Any) -> None:
+def test_len(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/get1", handler, name="name1")
     router.add_route("GET", "/get2", handler, name="name2")
     assert 2 == len(router)
 
 
-def test_iter(router: Any) -> None:
+def test_iter(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/get1", handler, name="name1")
     router.add_route("GET", "/get2", handler, name="name2")
     assert {"name1", "name2"} == set(iter(router))
 
 
-def test_contains(router: Any) -> None:
+def test_contains(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/get1", handler, name="name1")
     router.add_route("GET", "/get2", handler, name="name2")
@@ -541,37 +547,42 @@ def test_contains(router: Any) -> None:
     assert "name3" not in router
 
 
-def test_static_repr(router: Any) -> None:
+def test_static_repr(router: web.UrlDispatcher) -> None:
     router.add_static("/get", pathlib.Path(aiohttp.__file__).parent, name="name")
-    assert Matches(r"<StaticResource 'name' /get") == repr(router["name"])
+    assert repr(router["name"]).startswith("<StaticResource 'name' /get")
 
 
-def test_static_adds_slash(router: Any) -> None:
+def test_static_adds_slash(router: web.UrlDispatcher) -> None:
     route = router.add_static("/prefix", pathlib.Path(aiohttp.__file__).parent)
     assert "/prefix" == route._prefix
 
 
-def test_static_remove_trailing_slash(router: Any) -> None:
+def test_static_remove_trailing_slash(router: web.UrlDispatcher) -> None:
     route = router.add_static("/prefix/", pathlib.Path(aiohttp.__file__).parent)
     assert "/prefix" == route._prefix
 
 
-async def test_add_route_with_re(router: Any) -> None:
+@pytest.mark.parametrize(
+    "pattern,url,expected",
+    (
+        (r"{to:\d+}", r"1234", {"to": "1234"}),
+        ("{name}.html", "test.html", {"name": "test"}),
+        (r"{fn:\w+ \d+}", "abc 123", {"fn": "abc 123"}),
+        (r"{fn:\w+\s\d+}", "abc 123", {"fn": "abc 123"}),
+    ),
+)
+async def test_add_route_with_re(
+    router: web.UrlDispatcher, pattern: str, url: str, expected: Dict[str, str]
+) -> None:
     handler = make_handler()
-    router.add_route("GET", r"/handler/{to:\d+}", handler)
-
-    req = make_mocked_request("GET", "/handler/1234")
+    router.add_route("GET", f"/handler/{pattern}", handler)
+    req = make_mocked_request("GET", f"/handler/{url}")
     info = await router.resolve(req)
     assert info is not None
-    assert {"to": "1234"} == info
-
-    router.add_route("GET", r"/handler/{name}.html", handler)
-    req = make_mocked_request("GET", "/handler/test.html")
-    info = await router.resolve(req)
-    assert {"name": "test"} == info
+    assert info == expected
 
 
-async def test_add_route_with_re_and_slashes(router: Any) -> None:
+async def test_add_route_with_re_and_slashes(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", r"/handler/{to:[^/]+/?}", handler)
     req = make_mocked_request("GET", "/handler/1234/")
@@ -586,7 +597,7 @@ async def test_add_route_with_re_and_slashes(router: Any) -> None:
     assert {"to": "1234/5/6/7"} == info
 
 
-async def test_add_route_with_re_not_match(router: Any) -> None:
+async def test_add_route_with_re_not_match(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", r"/handler/{to:\d+}", handler)
 
@@ -594,11 +605,11 @@ async def test_add_route_with_re_not_match(router: Any) -> None:
     match_info = await router.resolve(req)
     assert isinstance(match_info.route, SystemRoute)
     assert {} == match_info
-    with pytest.raises(HTTPNotFound):
+    with pytest.raises(web.HTTPNotFound):
         await match_info.handler(req)
 
 
-async def test_add_route_with_re_including_slashes(router: Any) -> None:
+async def test_add_route_with_re_including_slashes(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", r"/handler/{to:.+}/tail", handler)
     req = make_mocked_request("GET", "/handler/re/with/slashes/tail")
@@ -607,7 +618,7 @@ async def test_add_route_with_re_including_slashes(router: Any) -> None:
     assert {"to": "re/with/slashes"} == info
 
 
-def test_add_route_with_invalid_re(router: Any) -> None:
+def test_add_route_with_invalid_re(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     with pytest.raises(ValueError) as ctx:
         router.add_route("GET", r"/handler/{to:+++}", handler)
@@ -622,7 +633,7 @@ def test_add_route_with_invalid_re(router: Any) -> None:
     assert ctx.value.__cause__ is None
 
 
-def test_route_dynamic_with_regex_spec(router: Any) -> None:
+def test_route_dynamic_with_regex_spec(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     route = router.add_route("GET", r"/get/{num:^\d+}", handler, name="name")
 
@@ -630,7 +641,9 @@ def test_route_dynamic_with_regex_spec(router: Any) -> None:
     assert "/get/123" == str(url)
 
 
-def test_route_dynamic_with_regex_spec_and_trailing_slash(router: Any) -> None:
+def test_route_dynamic_with_regex_spec_and_trailing_slash(
+    router: web.UrlDispatcher,
+) -> None:
     handler = make_handler()
     route = router.add_route("GET", r"/get/{num:^\d+}/", handler, name="name")
 
@@ -638,7 +651,7 @@ def test_route_dynamic_with_regex_spec_and_trailing_slash(router: Any) -> None:
     assert "/get/123/" == str(url)
 
 
-def test_route_dynamic_with_regex(router: Any) -> None:
+def test_route_dynamic_with_regex(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     route = router.add_route("GET", r"/{one}/{two:.+}", handler)
 
@@ -646,29 +659,30 @@ def test_route_dynamic_with_regex(router: Any) -> None:
     assert "/1/2" == str(url)
 
 
-def test_route_dynamic_quoting(router: Any) -> None:
+def test_route_dynamic_quoting(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     route = router.add_route("GET", r"/пре %2Fфикс/{arg}", handler)
 
     url = route.url_for(arg="1 2/текст%2F")
-    assert url.path == "/пре /фикс/1 2/текст%2F"
+    assert url.path == "/пре %2Fфикс/1 2/текст%2F"
     assert str(url) == (
         "/%D0%BF%D1%80%D0%B5%20%2F%D1%84%D0%B8%D0%BA%D1%81"
         "/1%202/%D1%82%D0%B5%D0%BA%D1%81%D1%82%252F"
     )
 
 
-async def test_regular_match_info(router: Any) -> None:
+async def test_regular_match_info(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/get/{name}", handler)
 
     req = make_mocked_request("GET", "/get/john")
     match_info = await router.resolve(req)
     assert {"name": "john"} == match_info
-    assert Matches("<MatchInfo {'name': 'john'}: .+<Dynamic.+>>") == repr(match_info)
+    assert repr(match_info).startswith("<MatchInfo {'name': 'john'}:")
+    assert "<Dynamic" in repr(match_info)
 
 
-async def test_match_info_with_plus(router: Any) -> None:
+async def test_match_info_with_plus(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/get/{version}", handler)
 
@@ -677,13 +691,13 @@ async def test_match_info_with_plus(router: Any) -> None:
     assert {"version": "1.0+test"} == match_info
 
 
-async def test_not_found_repr(router: Any) -> None:
+async def test_not_found_repr(router: web.UrlDispatcher) -> None:
     req = make_mocked_request("POST", "/path/to")
     match_info = await router.resolve(req)
     assert "<MatchInfoError 404: Not Found>" == repr(match_info)
 
 
-async def test_not_allowed_repr(router: Any) -> None:
+async def test_not_allowed_repr(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/path/to", handler)
 
@@ -695,40 +709,40 @@ async def test_not_allowed_repr(router: Any) -> None:
     assert "<MatchInfoError 405: Method Not Allowed>" == repr(match_info)
 
 
-def test_default_expect_handler(router: Any) -> None:
+def test_default_expect_handler(router: web.UrlDispatcher) -> None:
     route = router.add_route("GET", "/", make_handler())
     assert route._expect_handler is _default_expect_handler
 
 
-def test_custom_expect_handler_plain(router: Any) -> None:
-    async def handler(request):
-        pass
+def test_custom_expect_handler_plain(router: web.UrlDispatcher) -> None:
+    async def handler(request: web.Request) -> NoReturn:
+        assert False
 
     route = router.add_route("GET", "/", make_handler(), expect_handler=handler)
     assert route._expect_handler is handler
-    assert isinstance(route, ResourceRoute)
+    assert isinstance(route, web.ResourceRoute)
 
 
-def test_custom_expect_handler_dynamic(router: Any) -> None:
-    async def handler(request):
-        pass
+def test_custom_expect_handler_dynamic(router: web.UrlDispatcher) -> None:
+    async def handler(request: web.Request) -> NoReturn:
+        assert False
 
     route = router.add_route(
         "GET", "/get/{name}", make_handler(), expect_handler=handler
     )
     assert route._expect_handler is handler
-    assert isinstance(route, ResourceRoute)
+    assert isinstance(route, web.ResourceRoute)
 
 
-def test_expect_handler_non_coroutine(router: Any) -> None:
-    def handler(request):
-        pass
+def test_expect_handler_non_coroutine(router: web.UrlDispatcher) -> None:
+    def handler(request: web.Request) -> NoReturn:
+        assert False
 
     with pytest.raises(AssertionError):
         router.add_route("GET", "/", make_handler(), expect_handler=handler)
 
 
-async def test_dynamic_match_non_ascii(router: Any) -> None:
+async def test_dynamic_match_non_ascii(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/{var}", handler)
     req = make_mocked_request(
@@ -738,7 +752,7 @@ async def test_dynamic_match_non_ascii(router: Any) -> None:
     assert {"var": "рус текст"} == match_info
 
 
-async def test_dynamic_match_with_static_part(router: Any) -> None:
+async def test_dynamic_match_with_static_part(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/{name}.html", handler)
     req = make_mocked_request("GET", "/file.html")
@@ -746,7 +760,7 @@ async def test_dynamic_match_with_static_part(router: Any) -> None:
     assert {"name": "file"} == match_info
 
 
-async def test_dynamic_match_two_part2(router: Any) -> None:
+async def test_dynamic_match_two_part2(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/{name}.{ext}", handler)
     req = make_mocked_request("GET", "/file.html")
@@ -754,7 +768,7 @@ async def test_dynamic_match_two_part2(router: Any) -> None:
     assert {"name": "file", "ext": "html"} == match_info
 
 
-async def test_dynamic_match_unquoted_path(router: Any) -> None:
+async def test_dynamic_match_unquoted_path(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/{path}/{subpath}", handler)
     resource_id = "my%2Fpath%7Cwith%21some%25strange%24characters"
@@ -763,13 +777,13 @@ async def test_dynamic_match_unquoted_path(router: Any) -> None:
     assert match_info == {"path": "path", "subpath": unquote(resource_id)}
 
 
-def test_add_route_not_started_with_slash(router: Any) -> None:
+def test_add_route_not_started_with_slash(router: web.UrlDispatcher) -> None:
     with pytest.raises(ValueError):
         handler = make_handler()
         router.add_route("GET", "invalid_path", handler)
 
 
-def test_add_route_invalid_method(router: Any) -> None:
+def test_add_route_invalid_method(router: web.UrlDispatcher) -> None:
     sample_bad_methods = {
         "BAD METHOD",
         "B@D_METHOD",
@@ -785,34 +799,40 @@ def test_add_route_invalid_method(router: Any) -> None:
             router.add_route(bad_method, "/path", handler)
 
 
-def test_routes_view_len(router: Any, fill_routes: Any) -> None:
+def test_routes_view_len(
+    router: web.UrlDispatcher, fill_routes: Callable[[], List[web.AbstractRoute]]
+) -> None:
     fill_routes()
     assert 4 == len(router.routes())
 
 
-def test_routes_view_iter(router: Any, fill_routes: Any) -> None:
+def test_routes_view_iter(
+    router: web.UrlDispatcher, fill_routes: Callable[[], List[web.AbstractRoute]]
+) -> None:
     routes = fill_routes()
     assert list(routes) == list(router.routes())
 
 
-def test_routes_view_contains(router: Any, fill_routes: Any) -> None:
+def test_routes_view_contains(
+    router: web.UrlDispatcher, fill_routes: Callable[[], List[web.AbstractRoute]]
+) -> None:
     routes = fill_routes()
     for route in routes:
         assert route in router.routes()
 
 
-def test_routes_abc(router: Any) -> None:
+def test_routes_abc(router: web.UrlDispatcher) -> None:
     assert isinstance(router.routes(), Sized)
     assert isinstance(router.routes(), Iterable)
     assert isinstance(router.routes(), Container)
 
 
-def test_named_resources_abc(router: Any) -> None:
+def test_named_resources_abc(router: web.UrlDispatcher) -> None:
     assert isinstance(router.named_resources(), Mapping)
     assert not isinstance(router.named_resources(), MutableMapping)
 
 
-def test_named_resources(router: Any) -> None:
+def test_named_resources(router: web.UrlDispatcher) -> None:
     route1 = router.add_route("GET", "/plain", make_handler(), name="route1")
     route2 = router.add_route("GET", "/variable/{name}", make_handler(), name="route2")
     route3 = router.add_static(
@@ -823,13 +843,14 @@ def test_named_resources(router: Any) -> None:
     assert 3 == len(router.named_resources())
 
     for name in names:
+        assert name is not None
         assert name in router.named_resources()
-        assert isinstance(router.named_resources()[name], AbstractResource)
+        assert isinstance(router.named_resources()[name], web.AbstractResource)
 
 
-def test_resource_iter(router: Any) -> None:
-    async def handler(request):
-        pass
+def test_resource_iter(router: web.UrlDispatcher) -> None:
+    async def handler(request: web.Request) -> NoReturn:
+        assert False
 
     resource = router.add_resource("/path")
     r1 = resource.add_route("GET", handler)
@@ -838,25 +859,26 @@ def test_resource_iter(router: Any) -> None:
     assert [r1, r2] == list(resource)
 
 
-def test_view_route(router: Any) -> None:
+def test_view_route(router: web.UrlDispatcher) -> None:
     resource = router.add_resource("/path")
 
-    route = resource.add_route("*", View)
-    assert View is route.handler
+    route = resource.add_route("*", web.View)
+    assert web.View is route.handler
 
 
-def test_resource_route_match(router: Any) -> None:
-    async def handler(request):
-        pass
+def test_resource_route_match(router: web.UrlDispatcher) -> None:
+    async def handler(request: web.Request) -> NoReturn:
+        assert False
 
     resource = router.add_resource("/path")
     route = resource.add_route("GET", handler)
+    assert isinstance(route.resource, web.Resource)
     assert {} == route.resource._match("/path")
 
 
-def test_error_on_double_route_adding(router: Any) -> None:
-    async def handler(request):
-        pass
+def test_error_on_double_route_adding(router: web.UrlDispatcher) -> None:
+    async def handler(request: web.Request) -> NoReturn:
+        assert False
 
     resource = router.add_resource("/path")
 
@@ -865,9 +887,9 @@ def test_error_on_double_route_adding(router: Any) -> None:
         resource.add_route("GET", handler)
 
 
-def test_error_on_adding_route_after_wildcard(router: Any) -> None:
-    async def handler(request):
-        pass
+def test_error_on_adding_route_after_wildcard(router: web.UrlDispatcher) -> None:
+    async def handler(request: web.Request) -> NoReturn:
+        assert False
 
     resource = router.add_resource("/path")
 
@@ -876,7 +898,7 @@ def test_error_on_adding_route_after_wildcard(router: Any) -> None:
         resource.add_route("GET", handler)
 
 
-async def test_http_exception_is_none_when_resolved(router: Any) -> None:
+async def test_http_exception_is_none_when_resolved(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/", handler)
     req = make_mocked_request("GET", "/")
@@ -884,15 +906,18 @@ async def test_http_exception_is_none_when_resolved(router: Any) -> None:
     assert info.http_exception is None
 
 
-async def test_http_exception_is_not_none_when_not_resolved(router: Any) -> None:
+async def test_http_exception_is_not_none_when_not_resolved(
+    router: web.UrlDispatcher,
+) -> None:
     handler = make_handler()
     router.add_route("GET", "/", handler)
     req = make_mocked_request("GET", "/abc")
     info = await router.resolve(req)
+    assert info.http_exception is not None
     assert info.http_exception.status == 404
 
 
-async def test_match_info_get_info_plain(router: Any) -> None:
+async def test_match_info_get_info_plain(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/", handler)
     req = make_mocked_request("GET", "/")
@@ -900,7 +925,7 @@ async def test_match_info_get_info_plain(router: Any) -> None:
     assert info.get_info() == {"path": "/"}
 
 
-async def test_match_info_get_info_dynamic(router: Any) -> None:
+async def test_match_info_get_info_dynamic(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/{a}", handler)
     req = make_mocked_request("GET", "/value")
@@ -911,7 +936,7 @@ async def test_match_info_get_info_dynamic(router: Any) -> None:
     }
 
 
-async def test_match_info_get_info_dynamic2(router: Any) -> None:
+async def test_match_info_get_info_dynamic2(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/{a}/{b}", handler)
     req = make_mocked_request("GET", "/path/to")
@@ -924,17 +949,17 @@ async def test_match_info_get_info_dynamic2(router: Any) -> None:
     }
 
 
-def test_static_resource_get_info(router: Any) -> None:
+def test_static_resource_get_info(router: web.UrlDispatcher) -> None:
     directory = pathlib.Path(aiohttp.__file__).parent.resolve()
     resource = router.add_static("/st", directory)
     info = resource.get_info()
     assert len(info) == 3
     assert info["directory"] == directory
     assert info["prefix"] == "/st"
-    assert all([type(r) is ResourceRoute for r in info["routes"].values()])
+    assert all([type(r) is web.ResourceRoute for r in info["routes"].values()])
 
 
-async def test_system_route_get_info(router: Any) -> None:
+async def test_system_route_get_info(router: web.UrlDispatcher) -> None:
     handler = make_handler()
     router.add_route("GET", "/", handler)
     req = make_mocked_request("GET", "/abc")
@@ -942,20 +967,20 @@ async def test_system_route_get_info(router: Any) -> None:
     assert info.get_info()["http_exception"].status == 404
 
 
-def test_resources_view_len(router: Any) -> None:
+def test_resources_view_len(router: web.UrlDispatcher) -> None:
     router.add_resource("/plain")
     router.add_resource("/variable/{name}")
     assert 2 == len(router.resources())
 
 
-def test_resources_view_iter(router: Any) -> None:
+def test_resources_view_iter(router: web.UrlDispatcher) -> None:
     resource1 = router.add_resource("/plain")
     resource2 = router.add_resource("/variable/{name}")
     resources = [resource1, resource2]
     assert list(resources) == list(router.resources())
 
 
-def test_resources_view_contains(router: Any) -> None:
+def test_resources_view_contains(router: web.UrlDispatcher) -> None:
     resource1 = router.add_resource("/plain")
     resource2 = router.add_resource("/variable/{name}")
     resources = [resource1, resource2]
@@ -963,13 +988,13 @@ def test_resources_view_contains(router: Any) -> None:
         assert resource in router.resources()
 
 
-def test_resources_abc(router: Any) -> None:
+def test_resources_abc(router: web.UrlDispatcher) -> None:
     assert isinstance(router.resources(), Sized)
     assert isinstance(router.resources(), Iterable)
     assert isinstance(router.resources(), Container)
 
 
-def test_static_route_user_home(router: Any) -> None:
+def test_static_route_user_home(router: web.UrlDispatcher) -> None:
     here = pathlib.Path(aiohttp.__file__).parent
     try:
         static_dir = pathlib.Path("~") / here.relative_to(pathlib.Path.home())
@@ -980,25 +1005,27 @@ def test_static_route_user_home(router: Any) -> None:
     assert here == route.get_info()["directory"]
 
 
-def test_static_route_points_to_file(router: Any) -> None:
+def test_static_route_points_to_file(router: web.UrlDispatcher) -> None:
     here = pathlib.Path(aiohttp.__file__).parent / "__init__.py"
     with pytest.raises(ValueError):
         router.add_static("/st", here)
 
 
-async def test_404_for_static_resource(router: Any) -> None:
+async def test_404_for_static_resource(router: web.UrlDispatcher) -> None:
     resource = router.add_static("/st", pathlib.Path(aiohttp.__file__).parent)
     ret = await resource.resolve(make_mocked_request("GET", "/unknown/path"))
     assert (None, set()) == ret
 
 
-async def test_405_for_resource_adapter(router: Any) -> None:
+async def test_405_for_resource_adapter(router: web.UrlDispatcher) -> None:
     resource = router.add_static("/st", pathlib.Path(aiohttp.__file__).parent)
     ret = await resource.resolve(make_mocked_request("POST", "/st/abc.py"))
     assert (None, {"HEAD", "GET"}) == ret
 
 
-async def test_check_allowed_method_for_found_resource(router: Any) -> None:
+async def test_check_allowed_method_for_found_resource(
+    router: web.UrlDispatcher,
+) -> None:
     handler = make_handler()
     resource = router.add_resource("/")
     resource.add_route("GET", handler)
@@ -1007,24 +1034,24 @@ async def test_check_allowed_method_for_found_resource(router: Any) -> None:
     assert {"GET"} == ret[1]
 
 
-def test_url_for_in_static_resource(router: Any) -> None:
+def test_url_for_in_static_resource(router: web.UrlDispatcher) -> None:
     resource = router.add_static("/static", pathlib.Path(aiohttp.__file__).parent)
     assert URL("/static/file.txt") == resource.url_for(filename="file.txt")
 
 
-def test_url_for_in_static_resource_pathlib(router: Any) -> None:
+def test_url_for_in_static_resource_pathlib(router: web.UrlDispatcher) -> None:
     resource = router.add_static("/static", pathlib.Path(aiohttp.__file__).parent)
     assert URL("/static/file.txt") == resource.url_for(
         filename=pathlib.Path("file.txt")
     )
 
 
-def test_url_for_in_resource_route(router: Any) -> None:
+def test_url_for_in_resource_route(router: web.UrlDispatcher) -> None:
     route = router.add_route("GET", "/get/{name}", make_handler(), name="name")
     assert URL("/get/John") == route.url_for(name="John")
 
 
-def test_subapp_get_info(app: Any) -> None:
+def test_subapp_get_info(app: web.Application) -> None:
     subapp = web.Application()
     resource = subapp.add_subapp("/pre", subapp)
     assert resource.get_info() == {"prefix": "/pre", "app": subapp}
@@ -1040,9 +1067,9 @@ def test_subapp_get_info(app: Any) -> None:
         ("example$com", ValueError),
     ],
 )
-def test_domain_validation_error(domain: Any, error: Any) -> None:
+def test_domain_validation_error(domain: Optional[str], error: Type[Exception]) -> None:
     with pytest.raises(error):
-        Domain(domain)
+        Domain(domain)  # type: ignore[arg-type]
 
 
 def test_domain_valid() -> None:
@@ -1063,20 +1090,20 @@ def test_domain_valid() -> None:
         ("*.example.com", "example.com", False),
     ],
 )
-def test_match_domain(a: Any, b: Any, result: Any) -> None:
+def test_match_domain(a: str, b: str, result: bool) -> None:
     if "*" in a:
-        rule = MaskDomain(a)
+        rule: Domain = MaskDomain(a)
     else:
         rule = Domain(a)
     assert rule.match_domain(b) is result
 
 
-def test_add_subapp_errors(app: Any) -> None:
+def test_add_subapp_errors(app: web.Application) -> None:
     with pytest.raises(TypeError):
-        app.add_subapp(1, web.Application())
+        app.add_subapp(1, web.Application())  # type: ignore[arg-type]
 
 
-def test_subapp_rule_resource(app: Any) -> None:
+def test_subapp_rule_resource(app: web.Application) -> None:
     subapp = web.Application()
     subapp.router.add_get("/", make_handler())
     rule = Domain("example.com")
@@ -1093,13 +1120,17 @@ def test_subapp_rule_resource(app: Any) -> None:
         resource.url_for()
 
 
-async def test_add_domain_not_str(app: Any, loop: Any) -> None:
+async def test_add_domain_not_str(
+    app: web.Application, loop: asyncio.AbstractEventLoop
+) -> None:
     app = web.Application()
     with pytest.raises(TypeError):
-        app.add_domain(1, app)
+        app.add_domain(1, app)  # type: ignore[arg-type]
 
 
-async def test_add_domain(app: Any, loop: Any) -> None:
+async def test_add_domain(
+    app: web.Application, loop: asyncio.AbstractEventLoop
+) -> None:
     subapp1 = web.Application()
     h1 = make_handler()
     subapp1.router.add_get("/", h1)
@@ -1129,23 +1160,23 @@ async def test_add_domain(app: Any, loop: Any) -> None:
 
     request = make_mocked_request("POST", "/", {"host": "example.com"})
     match_info = await app.router.resolve(request)
-    assert isinstance(match_info.http_exception, HTTPMethodNotAllowed)
+    assert isinstance(match_info.http_exception, web.HTTPMethodNotAllowed)
 
 
-def test_subapp_url_for(app: Any) -> None:
+def test_subapp_url_for(app: web.Application) -> None:
     subapp = web.Application()
     resource = app.add_subapp("/pre", subapp)
     with pytest.raises(RuntimeError):
         resource.url_for()
 
 
-def test_subapp_repr(app: Any) -> None:
+def test_subapp_repr(app: web.Application) -> None:
     subapp = web.Application()
     resource = app.add_subapp("/pre", subapp)
     assert repr(resource).startswith("<PrefixedSubAppResource /pre -> <Application")
 
 
-def test_subapp_len(app: Any) -> None:
+def test_subapp_len(app: web.Application) -> None:
     subapp = web.Application()
     subapp.router.add_get("/", make_handler(), allow_head=False)
     subapp.router.add_post("/", make_handler())
@@ -1153,7 +1184,7 @@ def test_subapp_len(app: Any) -> None:
     assert len(resource) == 2
 
 
-def test_subapp_iter(app: Any) -> None:
+def test_subapp_iter(app: web.Application) -> None:
     subapp = web.Application()
     r1 = subapp.router.add_get("/", make_handler(), allow_head=False)
     r2 = subapp.router.add_post("/", make_handler())
@@ -1168,32 +1199,32 @@ def test_subapp_iter(app: Any) -> None:
         "class",
     ),
 )
-def test_invalid_route_name(router: Any, route_name: str) -> None:
+def test_invalid_route_name(router: web.UrlDispatcher, route_name: str) -> None:
     with pytest.raises(ValueError):
         router.add_get("/", make_handler(), name=route_name)
 
 
-def test_frozen_router(router: Any) -> None:
+def test_frozen_router(router: web.UrlDispatcher) -> None:
     router.freeze()
     with pytest.raises(RuntimeError):
         router.add_get("/", make_handler())
 
 
-def test_frozen_router_subapp(app: Any) -> None:
+def test_frozen_router_subapp(app: web.Application) -> None:
     subapp = web.Application()
     subapp.freeze()
     with pytest.raises(RuntimeError):
         app.add_subapp("/pre", subapp)
 
 
-def test_frozen_app_on_subapp(app: Any) -> None:
+def test_frozen_app_on_subapp(app: web.Application) -> None:
     app.freeze()
     subapp = web.Application()
     with pytest.raises(RuntimeError):
         app.add_subapp("/pre", subapp)
 
 
-def test_set_options_route(router: Any) -> None:
+def test_set_options_route(router: web.UrlDispatcher) -> None:
     resource = router.add_static("/static", pathlib.Path(aiohttp.__file__).parent)
     options = None
     for route in resource:
@@ -1210,27 +1241,32 @@ def test_set_options_route(router: Any) -> None:
         resource.set_options_route(make_handler())
 
 
-def test_dynamic_url_with_name_started_from_underscore(router: Any) -> None:
+def test_dynamic_url_with_name_started_from_underscore(
+    router: web.UrlDispatcher,
+) -> None:
     route = router.add_route("GET", "/get/{_name}", make_handler())
     assert URL("/get/John") == route.url_for(_name="John")
 
 
-def test_cannot_add_subapp_with_empty_prefix(app: Any) -> None:
+def test_cannot_add_subapp_with_empty_prefix(app: web.Application) -> None:
     subapp = web.Application()
     with pytest.raises(ValueError):
         app.add_subapp("", subapp)
 
 
-def test_cannot_add_subapp_with_slash_prefix(app: Any) -> None:
+def test_cannot_add_subapp_with_slash_prefix(app: web.Application) -> None:
     subapp = web.Application()
     with pytest.raises(ValueError):
         app.add_subapp("/", subapp)
 
 
-async def test_convert_empty_path_to_slash_on_freezing(router: Any) -> None:
+async def test_convert_empty_path_to_slash_on_freezing(
+    router: web.UrlDispatcher,
+) -> None:
     handler = make_handler()
     route = router.add_get("", handler)
     resource = route.resource
+    assert resource is not None
     assert resource.get_info() == {"path": ""}
     router.freeze()
     assert resource.get_info() == {"path": "/"}
@@ -1238,7 +1274,7 @@ async def test_convert_empty_path_to_slash_on_freezing(router: Any) -> None:
 
 def test_plain_resource_canonical() -> None:
     canonical = "/plain/path"
-    res = PlainResource(path=canonical)
+    res = web.PlainResource(path=canonical)
     assert res.canonical == canonical
 
 
@@ -1250,7 +1286,7 @@ def test_dynamic_resource_canonical() -> None:
         r"/{one}/{two:.+}": r"/{one}/{two}",
     }
     for pattern, canonical in canonicals.items():
-        res = DynamicResource(path=pattern)
+        res = web.DynamicResource(path=pattern)
         assert res.canonical == canonical
 
 
@@ -1258,18 +1294,18 @@ def test_static_resource_canonical() -> None:
     prefix = "/prefix"
     directory = str(pathlib.Path(aiohttp.__file__).parent)
     canonical = prefix
-    res = StaticResource(prefix=prefix, directory=directory)
+    res = web.StaticResource(prefix=prefix, directory=directory)
     assert res.canonical == canonical
 
 
-def test_prefixed_subapp_resource_canonical(app: Any) -> None:
+def test_prefixed_subapp_resource_canonical(app: web.Application) -> None:
     canonical = "/prefix"
     subapp = web.Application()
     res = subapp.add_subapp(canonical, subapp)
     assert res.canonical == canonical
 
 
-async def test_prefixed_subapp_overlap(app: Any) -> None:
+async def test_prefixed_subapp_overlap(app: web.Application) -> None:
     # Subapp should not overshadow other subapps with overlapping prefixes
     subapp1 = web.Application()
     handler1 = make_handler()
@@ -1294,7 +1330,7 @@ async def test_prefixed_subapp_overlap(app: Any) -> None:
     assert match_info.route.handler is handler3
 
 
-async def test_prefixed_subapp_empty_route(app: Any) -> None:
+async def test_prefixed_subapp_empty_route(app: web.Application) -> None:
     subapp = web.Application()
     handler = make_handler()
     subapp.router.add_get("", handler)
@@ -1306,7 +1342,7 @@ async def test_prefixed_subapp_empty_route(app: Any) -> None:
     assert "<MatchInfoError 404: Not Found>" == repr(match_info)
 
 
-async def test_prefixed_subapp_root_route(app: Any) -> None:
+async def test_prefixed_subapp_root_route(app: web.Application) -> None:
     subapp = web.Application()
     handler = make_handler()
     subapp.router.add_get("/", handler)
