@@ -5,7 +5,7 @@ import json
 import pathlib
 import socket
 import zlib
-from typing import Any, Optional
+from typing import Any, NoReturn, Optional
 from unittest import mock
 
 import pytest
@@ -68,7 +68,7 @@ async def test_simple_get(aiohttp_client: Any) -> None:
     txt = await resp.text()
     assert "OK" == txt
 
-    await resp.release()
+    resp.release()
 
 
 async def test_simple_get_with_text(aiohttp_client: Any) -> None:
@@ -86,7 +86,7 @@ async def test_simple_get_with_text(aiohttp_client: Any) -> None:
     txt = await resp.text()
     assert "OK" == txt
 
-    await resp.release()
+    resp.release()
 
 
 async def test_handler_returns_not_response(
@@ -103,12 +103,8 @@ async def test_handler_returns_not_response(
     server = await aiohttp_server(app, logger=logger)
     client = await aiohttp_client(server)
 
-    with pytest.raises(aiohttp.ServerDisconnectedError):
-        await client.get("/")
-
-    logger.exception.assert_called_with(
-        "Unhandled runtime exception", exc_info=mock.ANY
-    )
+    async with client.get("/") as resp:
+        assert resp.status == 500
 
 
 async def test_handler_returns_none(aiohttp_server: Any, aiohttp_client: Any) -> None:
@@ -123,13 +119,22 @@ async def test_handler_returns_none(aiohttp_server: Any, aiohttp_client: Any) ->
     server = await aiohttp_server(app, logger=logger)
     client = await aiohttp_client(server)
 
-    with pytest.raises(aiohttp.ServerDisconnectedError):
-        await client.get("/")
+    async with client.get("/") as resp:
+        assert resp.status == 500
 
-    # Actual error text is placed in exc_info
-    logger.exception.assert_called_with(
-        "Unhandled runtime exception", exc_info=mock.ANY
-    )
+
+async def test_handler_returns_not_response_after_100expect(
+    aiohttp_server: Any, aiohttp_client: Any
+) -> None:
+    async def handler(request: web.Request) -> NoReturn:
+        raise Exception("foo")
+
+    app = web.Application()
+    app.router.add_get("/", handler)
+    client = await aiohttp_client(app)
+
+    async with client.get("/", expect100=True) as resp:
+        assert resp.status == 500
 
 
 async def test_head_returns_empty_body(aiohttp_client: Any) -> None:
@@ -150,6 +155,21 @@ async def test_head_returns_empty_body(aiohttp_client: Any) -> None:
     assert resp.headers["Content-Length"] == "4"
 
 
+@pytest.mark.parametrize("status", (201, 204, 404))
+async def test_default_content_type_no_body(aiohttp_client: Any, status: int) -> None:
+    async def handler(request):
+        return web.Response(status=status)
+
+    app = web.Application()
+    app.router.add_get("/", handler)
+    client = await aiohttp_client(app)
+
+    async with client.get("/") as resp:
+        assert resp.status == status
+        assert await resp.read() == b""
+        assert "Content-Type" not in resp.headers
+
+
 async def test_response_before_complete(aiohttp_client: Any) -> None:
     async def handler(request):
         return web.Response(body=b"OK")
@@ -165,7 +185,7 @@ async def test_response_before_complete(aiohttp_client: Any) -> None:
     text = await resp.text()
     assert "OK" == text
 
-    await resp.release()
+    resp.release()
 
 
 async def test_post_form(aiohttp_client: Any) -> None:
@@ -183,7 +203,7 @@ async def test_post_form(aiohttp_client: Any) -> None:
     txt = await resp.text()
     assert "OK" == txt
 
-    await resp.release()
+    resp.release()
 
 
 async def test_post_text(aiohttp_client: Any) -> None:
@@ -203,7 +223,7 @@ async def test_post_text(aiohttp_client: Any) -> None:
     txt = await resp.text()
     assert "русский" == txt
 
-    await resp.release()
+    resp.release()
 
 
 async def test_post_json(aiohttp_client: Any) -> None:
@@ -229,7 +249,7 @@ async def test_post_json(aiohttp_client: Any) -> None:
     data = await resp.json()
     assert dct == data
 
-    await resp.release()
+    resp.release()
 
 
 async def test_multipart(aiohttp_client: Any) -> None:
@@ -263,7 +283,7 @@ async def test_multipart(aiohttp_client: Any) -> None:
 
     resp = await client.post("/", data=writer)
     assert 200 == resp.status
-    await resp.release()
+    resp.release()
 
 
 async def test_multipart_empty(aiohttp_client: Any) -> None:
@@ -283,7 +303,7 @@ async def test_multipart_empty(aiohttp_client: Any) -> None:
 
     resp = await client.post("/", data=writer)
     assert 200 == resp.status
-    await resp.release()
+    resp.release()
 
 
 async def test_multipart_content_transfer_encoding(aiohttp_client: Any) -> None:
@@ -312,7 +332,7 @@ async def test_multipart_content_transfer_encoding(aiohttp_client: Any) -> None:
 
     resp = await client.post("/", data=writer)
     assert 200 == resp.status
-    await resp.release()
+    resp.release()
 
 
 async def test_render_redirect(aiohttp_client: Any) -> None:
@@ -329,7 +349,7 @@ async def test_render_redirect(aiohttp_client: Any) -> None:
     assert "301: Moved Permanently" == txt
     assert "/path" == resp.headers["location"]
 
-    await resp.release()
+    resp.release()
 
 
 async def test_post_single_file(aiohttp_client: Any) -> None:
@@ -361,7 +381,7 @@ async def test_post_single_file(aiohttp_client: Any) -> None:
         resp = await client.post("/", data=[fd])
     assert 200 == resp.status
 
-    await resp.release()
+    resp.release()
 
 
 async def test_files_upload_with_same_key(aiohttp_client: Any) -> None:
@@ -396,7 +416,7 @@ async def test_files_upload_with_same_key(aiohttp_client: Any) -> None:
     resp = await client.post("/", data=data)
     assert 200 == resp.status
 
-    await resp.release()
+    resp.release()
 
 
 async def test_post_files(aiohttp_client: Any) -> None:
@@ -427,7 +447,7 @@ async def test_post_files(aiohttp_client: Any) -> None:
             resp = await client.post("/", data=[f1, f2])
             assert 200 == resp.status
 
-            await resp.release()
+            resp.release()
 
 
 async def test_release_post_data(aiohttp_client: Any) -> None:
@@ -444,7 +464,7 @@ async def test_release_post_data(aiohttp_client: Any) -> None:
     resp = await client.post("/", data="post text")
     assert 200 == resp.status
 
-    await resp.release()
+    resp.release()
 
 
 async def test_post_form_with_duplicate_keys(aiohttp_client: Any) -> None:
@@ -461,7 +481,7 @@ async def test_post_form_with_duplicate_keys(aiohttp_client: Any) -> None:
     resp = await client.post("/", data=MultiDict([("a", 1), ("a", 2)]))
     assert 200 == resp.status
 
-    await resp.release()
+    resp.release()
 
 
 def test_repr_for_application() -> None:
@@ -494,7 +514,7 @@ async def test_expect_default_handler_unknown(aiohttp_client: Any) -> None:
     resp = await client.post("/", headers={"Expect": "SPAM"})
     assert 417 == resp.status
 
-    await resp.release()
+    resp.release()
 
 
 async def test_100_continue(aiohttp_client: Any) -> None:
@@ -513,7 +533,7 @@ async def test_100_continue(aiohttp_client: Any) -> None:
     resp = await client.post("/", data=form, expect100=True)
     assert 200 == resp.status
 
-    await resp.release()
+    resp.release()
 
 
 async def test_100_continue_custom(aiohttp_client: Any) -> None:
@@ -538,7 +558,7 @@ async def test_100_continue_custom(aiohttp_client: Any) -> None:
     assert 200 == resp.status
     assert expect_received
 
-    await resp.release()
+    resp.release()
 
 
 async def test_100_continue_custom_response(aiohttp_client: Any) -> None:
@@ -561,12 +581,12 @@ async def test_100_continue_custom_response(aiohttp_client: Any) -> None:
     auth_err = False
     resp = await client.post("/", data=new_dummy_form(), expect100=True)
     assert 200 == resp.status
-    await resp.release()
+    resp.release()
 
     auth_err = True
     resp = await client.post("/", data=new_dummy_form(), expect100=True)
     assert 403 == resp.status
-    await resp.release()
+    resp.release()
 
 
 async def test_expect_handler_custom_response(aiohttp_client: Any) -> None:
@@ -602,7 +622,7 @@ async def test_100_continue_for_not_found(aiohttp_client: Any) -> None:
     resp = await client.post("/not_found", data="data", expect100=True)
     assert 404 == resp.status
 
-    await resp.release()
+    resp.release()
 
 
 async def test_100_continue_for_not_allowed(aiohttp_client: Any) -> None:
@@ -616,7 +636,7 @@ async def test_100_continue_for_not_allowed(aiohttp_client: Any) -> None:
     resp = await client.get("/", expect100=True)
     assert 405 == resp.status
 
-    await resp.release()
+    resp.release()
 
 
 async def test_http11_keep_alive_default(aiohttp_client: Any) -> None:
@@ -632,7 +652,7 @@ async def test_http11_keep_alive_default(aiohttp_client: Any) -> None:
     assert resp.version == HttpVersion11
     assert "Connection" not in resp.headers
 
-    await resp.release()
+    resp.release()
 
 
 @pytest.mark.xfail
@@ -649,7 +669,7 @@ async def test_http10_keep_alive_default(aiohttp_client: Any) -> None:
     assert resp.version == HttpVersion10
     assert resp.headers["Connection"] == "keep-alive"
 
-    await resp.release()
+    resp.release()
 
 
 async def test_http10_keep_alive_with_headers_close(aiohttp_client: Any) -> None:
@@ -667,7 +687,7 @@ async def test_http10_keep_alive_with_headers_close(aiohttp_client: Any) -> None
     assert resp.version == HttpVersion10
     assert "Connection" not in resp.headers
 
-    await resp.release()
+    resp.release()
 
 
 async def test_http10_keep_alive_with_headers(aiohttp_client: Any) -> None:
@@ -685,7 +705,7 @@ async def test_http10_keep_alive_with_headers(aiohttp_client: Any) -> None:
     assert resp.version == HttpVersion10
     assert resp.headers["Connection"] == "keep-alive"
 
-    await resp.release()
+    resp.release()
 
 
 async def test_upload_file(aiohttp_client: Any) -> None:
@@ -708,7 +728,7 @@ async def test_upload_file(aiohttp_client: Any) -> None:
     resp = await client.post("/", data={"file": io.BytesIO(data)})
     assert 200 == resp.status
 
-    await resp.release()
+    resp.release()
 
 
 async def test_upload_file_object(aiohttp_client: Any) -> None:
@@ -732,7 +752,7 @@ async def test_upload_file_object(aiohttp_client: Any) -> None:
         resp = await client.post("/", data={"file": f})
         assert 200 == resp.status
 
-        await resp.release()
+        resp.release()
 
 
 @pytest.mark.parametrize(
@@ -768,7 +788,7 @@ async def test_empty_content_for_query_with_body(aiohttp_client: Any) -> None:
     resp = await client.post("/", data=b"data")
     assert 200 == resp.status
 
-    await resp.release()
+    resp.release()
 
 
 async def test_get_with_empty_arg(aiohttp_client: Any) -> None:
@@ -784,7 +804,7 @@ async def test_get_with_empty_arg(aiohttp_client: Any) -> None:
     resp = await client.get("/?arg")
     assert 200 == resp.status
 
-    await resp.release()
+    resp.release()
 
 
 async def test_large_header(aiohttp_client: Any) -> None:
@@ -799,7 +819,7 @@ async def test_large_header(aiohttp_client: Any) -> None:
     resp = await client.get("/", headers=headers)
     assert 400 == resp.status
 
-    await resp.release()
+    resp.release()
 
 
 async def test_large_header_allowed(aiohttp_client: Any, aiohttp_server: Any) -> None:
@@ -815,7 +835,7 @@ async def test_large_header_allowed(aiohttp_client: Any, aiohttp_server: Any) ->
     resp = await client.post("/", headers=headers)
     assert 200 == resp.status
 
-    await resp.release()
+    resp.release()
 
 
 async def test_get_with_empty_arg_with_equal(aiohttp_client: Any) -> None:
@@ -831,7 +851,7 @@ async def test_get_with_empty_arg_with_equal(aiohttp_client: Any) -> None:
     resp = await client.get("/?arg=")
     assert 200 == resp.status
 
-    await resp.release()
+    resp.release()
 
 
 async def test_response_with_async_gen(aiohttp_client: Any, fname: Any) -> None:
@@ -861,7 +881,7 @@ async def test_response_with_async_gen(aiohttp_client: Any, fname: Any) -> None:
     assert resp_data == data
     assert resp.headers.get("Content-Length") == str(len(resp_data))
 
-    await resp.release()
+    resp.release()
 
 
 async def test_response_with_async_gen_no_params(
@@ -893,7 +913,7 @@ async def test_response_with_async_gen_no_params(
     assert resp_data == data
     assert resp.headers.get("Content-Length") == str(len(resp_data))
 
-    await resp.release()
+    resp.release()
 
 
 async def test_response_with_file(aiohttp_client: Any, fname: Any) -> None:
@@ -924,7 +944,7 @@ async def test_response_with_file(aiohttp_client: Any, fname: Any) -> None:
     assert resp.headers.get("Content-Length") == str(len(resp_data))
     assert resp.headers.get("Content-Disposition") == expected_content_disposition
 
-    await resp.release()
+    resp.release()
 
     outer_file_descriptor.close()
 
@@ -956,7 +976,7 @@ async def test_response_with_file_ctype(aiohttp_client: Any, fname: Any) -> None
     assert resp.headers.get("Content-Length") == str(len(resp_data))
     assert resp.headers.get("Content-Disposition") == expected_content_disposition
 
-    await resp.release()
+    resp.release()
 
     outer_file_descriptor.close()
 
@@ -986,7 +1006,7 @@ async def test_response_with_payload_disp(aiohttp_client: Any, fname: Any) -> No
     assert resp.headers.get("Content-Length") == str(len(resp_data))
     assert resp.headers.get("Content-Disposition") == 'inline; filename="test.txt"'
 
-    await resp.release()
+    resp.release()
 
     outer_file_descriptor.close()
 
@@ -1004,7 +1024,7 @@ async def test_response_with_payload_stringio(aiohttp_client: Any, fname: Any) -
     resp_data = await resp.read()
     assert resp_data == b"test"
 
-    await resp.release()
+    resp.release()
 
 
 @pytest.mark.parametrize(
@@ -1035,7 +1055,7 @@ async def test_response_with_precompressed_body(
     assert b"mydata" == data
     assert resp.headers.get("Content-Encoding") == encoding
 
-    await resp.release()
+    resp.release()
 
 
 async def test_response_with_precompressed_body_brotli(aiohttp_client: Any) -> None:
@@ -1053,7 +1073,7 @@ async def test_response_with_precompressed_body_brotli(aiohttp_client: Any) -> N
     assert b"mydata" == data
     assert resp.headers.get("Content-Encoding") == "br"
 
-    await resp.release()
+    resp.release()
 
 
 async def test_bad_request_payload(aiohttp_client: Any) -> None:
@@ -1072,7 +1092,7 @@ async def test_bad_request_payload(aiohttp_client: Any) -> None:
     resp = await client.post("/", data=b"test", headers={"content-encoding": "gzip"})
     assert 200 == resp.status
 
-    await resp.release()
+    resp.release()
 
 
 async def test_stream_response_multiple_chunks(aiohttp_client: Any) -> None:
@@ -1094,7 +1114,7 @@ async def test_stream_response_multiple_chunks(aiohttp_client: Any) -> None:
     data = await resp.read()
     assert b"xyz" == data
 
-    await resp.release()
+    resp.release()
 
 
 async def test_start_without_routes(aiohttp_client: Any) -> None:
@@ -1104,7 +1124,7 @@ async def test_start_without_routes(aiohttp_client: Any) -> None:
     resp = await client.get("/")
     assert 404 == resp.status
 
-    await resp.release()
+    resp.release()
 
 
 async def test_requests_count(aiohttp_client: Any) -> None:
@@ -1119,17 +1139,17 @@ async def test_requests_count(aiohttp_client: Any) -> None:
     resp = await client.get("/")
     assert 200 == resp.status
     assert client.server.handler.requests_count == 1
-    await resp.release()
+    resp.release()
 
     resp = await client.get("/")
     assert 200 == resp.status
     assert client.server.handler.requests_count == 2
-    await resp.release()
+    resp.release()
 
     resp = await client.get("/")
     assert 200 == resp.status
     assert client.server.handler.requests_count == 3
-    await resp.release()
+    resp.release()
 
 
 async def test_redirect_url(aiohttp_client: Any) -> None:
@@ -1147,7 +1167,7 @@ async def test_redirect_url(aiohttp_client: Any) -> None:
     resp = await client.get("/redirector")
     assert resp.status == 200
 
-    await resp.release()
+    resp.release()
 
 
 async def test_simple_subapp(aiohttp_client: Any) -> None:
@@ -1165,7 +1185,7 @@ async def test_simple_subapp(aiohttp_client: Any) -> None:
     txt = await resp.text()
     assert "OK" == txt
 
-    await resp.release()
+    resp.release()
 
 
 async def test_subapp_reverse_url(aiohttp_client: Any) -> None:
@@ -1188,7 +1208,7 @@ async def test_subapp_reverse_url(aiohttp_client: Any) -> None:
     assert "OK" == txt
     assert resp.url.path == "/path/final"
 
-    await resp.release()
+    resp.release()
 
 
 async def test_subapp_reverse_variable_url(aiohttp_client: Any) -> None:
@@ -1213,7 +1233,7 @@ async def test_subapp_reverse_variable_url(aiohttp_client: Any) -> None:
     assert "OK" == txt
     assert resp.url.path == "/path/final"
 
-    await resp.release()
+    resp.release()
 
 
 async def test_subapp_reverse_static_url(aiohttp_client: Any) -> None:
@@ -1237,7 +1257,7 @@ async def test_subapp_reverse_static_url(aiohttp_client: Any) -> None:
     assert resp.status == 200
     body = await resp.read()
 
-    await resp.release()
+    resp.release()
 
     with (here / fname).open("rb") as f:
         assert body == f.read()
@@ -1259,7 +1279,7 @@ async def test_subapp_app(aiohttp_client: Any) -> None:
     txt = await resp.text()
     assert "OK" == txt
 
-    await resp.release()
+    resp.release()
 
 
 async def test_subapp_not_found(aiohttp_client: Any) -> None:
@@ -1275,7 +1295,7 @@ async def test_subapp_not_found(aiohttp_client: Any) -> None:
     resp = await client.get("/path/other")
     assert resp.status == 404
 
-    await resp.release()
+    resp.release()
 
 
 async def test_subapp_not_found2(aiohttp_client: Any) -> None:
@@ -1291,7 +1311,7 @@ async def test_subapp_not_found2(aiohttp_client: Any) -> None:
     resp = await client.get("/invalid/other")
     assert resp.status == 404
 
-    await resp.release()
+    resp.release()
 
 
 async def test_subapp_not_allowed(aiohttp_client: Any) -> None:
@@ -1308,7 +1328,7 @@ async def test_subapp_not_allowed(aiohttp_client: Any) -> None:
     assert resp.status == 405
     assert resp.headers["Allow"] == "GET,HEAD"
 
-    await resp.release()
+    resp.release()
 
 
 async def test_subapp_cannot_add_app_in_handler(aiohttp_client: Any) -> None:
@@ -1325,7 +1345,7 @@ async def test_subapp_cannot_add_app_in_handler(aiohttp_client: Any) -> None:
     resp = await client.get("/path/to")
     assert resp.status == 500
 
-    await resp.release()
+    resp.release()
 
 
 async def test_old_style_subapp_middlewares(aiohttp_client: Any) -> None:
@@ -1372,7 +1392,7 @@ async def test_old_style_subapp_middlewares(aiohttp_client: Any) -> None:
         (2, "app"),
     ] == order
 
-    await resp.release()
+    resp.release()
 
 
 async def test_subapp_on_response_prepare(aiohttp_client: Any) -> None:
@@ -1402,7 +1422,7 @@ async def test_subapp_on_response_prepare(aiohttp_client: Any) -> None:
     assert resp.status == 200
     assert [app, subapp1, subapp2] == order
 
-    await resp.release()
+    resp.release()
 
 
 async def test_subapp_on_startup(aiohttp_server: Any) -> None:
@@ -1519,7 +1539,7 @@ async def test_subapp_middleware_context(
     assert "Ok" == await resp.text()
     assert expected == values
 
-    await resp.release()
+    resp.release()
 
 
 async def test_custom_date_header(aiohttp_client: Any) -> None:
@@ -1534,7 +1554,7 @@ async def test_custom_date_header(aiohttp_client: Any) -> None:
     assert 200 == resp.status
     assert resp.headers["Date"] == "Sun, 30 Oct 2016 03:13:52 GMT"
 
-    await resp.release()
+    resp.release()
 
 
 async def test_response_prepared_with_clone(aiohttp_client: Any) -> None:
@@ -1551,7 +1571,7 @@ async def test_response_prepared_with_clone(aiohttp_client: Any) -> None:
     resp = await client.get("/")
     assert 200 == resp.status
 
-    await resp.release()
+    resp.release()
 
 
 async def test_app_max_client_size(aiohttp_client: Any) -> None:
@@ -1568,14 +1588,12 @@ async def test_app_max_client_size(aiohttp_client: Any) -> None:
         resp = await client.post("/", data=data)
     assert 413 == resp.status
     resp_text = await resp.text()
-    assert (
-        "Maximum request body size 1048576 exceeded, " "actual body size" in resp_text
-    )
+    assert "Maximum request body size 1048576 exceeded, actual body size" in resp_text
     # Maximum request body size X exceeded, actual body size X
     body_size = int(resp_text.split()[-1])
     assert body_size >= max_size
 
-    await resp.release()
+    resp.release()
 
 
 async def test_app_max_client_size_adjusted(aiohttp_client: Any) -> None:
@@ -1595,21 +1613,19 @@ async def test_app_max_client_size_adjusted(aiohttp_client: Any) -> None:
     assert 200 == resp.status
     resp_text = await resp.text()
     assert "ok" == resp_text
-    await resp.release()
+    resp.release()
 
     too_large_data = {"log_string": custom_max_size * "x" + "xxx"}
     with pytest.warns(ResourceWarning):
         resp = await client.post("/", data=too_large_data)
     assert 413 == resp.status
     resp_text = await resp.text()
-    assert (
-        "Maximum request body size 2097152 exceeded, " "actual body size" in resp_text
-    )
+    assert "Maximum request body size 2097152 exceeded, actual body size" in resp_text
     # Maximum request body size X exceeded, actual body size X
     body_size = int(resp_text.split()[-1])
     assert body_size >= custom_max_size
 
-    await resp.release()
+    resp.release()
 
 
 async def test_app_max_client_size_none(aiohttp_client: Any) -> None:
@@ -1629,7 +1645,7 @@ async def test_app_max_client_size_none(aiohttp_client: Any) -> None:
     assert 200 == resp.status
     resp_text = await resp.text()
     assert "ok" == resp_text
-    await resp.release()
+    resp.release()
 
     too_large_data = {"log_string": default_max_size * 2 * "x"}
     with pytest.warns(ResourceWarning):
@@ -1637,7 +1653,7 @@ async def test_app_max_client_size_none(aiohttp_client: Any) -> None:
     assert 200 == resp.status
     resp_text = await resp.text()
     assert resp_text == "ok"
-    await resp.release()
+    resp.release()
 
 
 async def test_post_max_client_size(aiohttp_client: Any) -> None:
@@ -1661,7 +1677,7 @@ async def test_post_max_client_size(aiohttp_client: Any) -> None:
         )
         data["file"].close()
 
-        await resp.release()
+        resp.release()
 
 
 async def test_post_max_client_size_for_file(aiohttp_client: Any) -> None:
@@ -1679,7 +1695,7 @@ async def test_post_max_client_size_for_file(aiohttp_client: Any) -> None:
 
     assert 413 == resp.status
 
-    await resp.release()
+    resp.release()
 
 
 async def test_response_with_bodypart(aiohttp_client: Any) -> None:
@@ -1703,7 +1719,7 @@ async def test_response_with_bodypart(aiohttp_client: Any) -> None:
         disp = multipart.parse_content_disposition(resp.headers["content-disposition"])
         assert disp == ("attachment", {"name": "file", "filename": "file"})
 
-        await resp.release()
+        resp.release()
 
 
 async def test_response_with_bodypart_named(aiohttp_client: Any, tmp_path: Any) -> None:
@@ -1729,7 +1745,7 @@ async def test_response_with_bodypart_named(aiohttp_client: Any, tmp_path: Any) 
     disp = multipart.parse_content_disposition(resp.headers["content-disposition"])
     assert disp == ("attachment", {"name": "file", "filename": "foobar.txt"})
 
-    await resp.release()
+    resp.release()
 
 
 async def test_response_with_bodypart_invalid_name(aiohttp_client: Any) -> None:
@@ -1752,7 +1768,7 @@ async def test_response_with_bodypart_invalid_name(aiohttp_client: Any) -> None:
 
     assert "content-disposition" not in resp.headers
 
-    await resp.release()
+    resp.release()
 
 
 async def test_request_clone(aiohttp_client: Any) -> None:
@@ -1768,7 +1784,7 @@ async def test_request_clone(aiohttp_client: Any) -> None:
 
     resp = await client.get("/")
     assert 200 == resp.status
-    await resp.release()
+    resp.release()
 
 
 async def test_await(aiohttp_server: Any) -> None:
@@ -1794,7 +1810,7 @@ async def test_await(aiohttp_server: Any) -> None:
         assert resp.status == 200
         assert resp.connection is not None
         await resp.read()
-        await resp.release()
+        resp.release()
         assert resp.connection is None
 
 
@@ -1878,7 +1894,7 @@ async def test_context_manager_close_on_release(
         assert resp.connection is None
         assert proto.close.called
 
-        await resp.release()  # Trigger handler completion
+        resp.release()  # Trigger handler completion
 
 
 async def test_iter_any(aiohttp_server: Any) -> None:
@@ -1968,7 +1984,7 @@ async def test_request_tracing(aiohttp_server: Any) -> None:
     assert on_connection_create_start.called
     assert on_connection_create_end.called
 
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -1982,7 +1998,7 @@ async def test_raise_http_exception(aiohttp_client: Any) -> None:
 
     resp = await client.get("/")
     assert resp.status == 403
-    await resp.release()
+    resp.release()
 
 
 async def test_request_path(aiohttp_client: Any) -> None:
@@ -2000,7 +2016,7 @@ async def test_request_path(aiohttp_client: Any) -> None:
     assert 200 == resp.status
     txt = await resp.text()
     assert "OK" == txt
-    await resp.release()
+    resp.release()
 
 
 async def test_app_add_routes(aiohttp_client: Any) -> None:
@@ -2013,7 +2029,7 @@ async def test_app_add_routes(aiohttp_client: Any) -> None:
     client = await aiohttp_client(app)
     resp = await client.get("/get")
     assert resp.status == 200
-    await resp.release()
+    resp.release()
 
 
 async def test_request_headers_type(aiohttp_client: Any) -> None:
@@ -2027,7 +2043,7 @@ async def test_request_headers_type(aiohttp_client: Any) -> None:
     client = await aiohttp_client(app)
     resp = await client.get("/get")
     assert resp.status == 200
-    await resp.release()
+    resp.release()
 
 
 async def test_signal_on_error_handler(aiohttp_client: Any) -> None:
@@ -2041,7 +2057,7 @@ async def test_signal_on_error_handler(aiohttp_client: Any) -> None:
     resp = await client.get("/")
     assert resp.status == 404
     assert resp.headers["X-Custom"] == "val"
-    await resp.release()
+    resp.release()
 
 
 @pytest.mark.skipif(
@@ -2069,7 +2085,7 @@ async def test_read_bufsize(aiohttp_client: Any) -> None:
     resp = await client.post("/", data=b"data")
     assert resp.status == 200
     assert await resp.text() == "data (2, 4)"
-    await resp.release()
+    resp.release()
 
 
 @pytest.mark.parametrize(
@@ -2096,7 +2112,7 @@ async def test_auto_decompress(
     resp = await client.post("/", data=compressed, headers=headers)
     assert resp.status == 200
     assert await resp.text() == str(len(locals()[len_of]))
-    await resp.release()
+    resp.release()
 
 
 @pytest.mark.parametrize(
@@ -2115,7 +2131,7 @@ async def test_response_101_204_no_content_length_http11(
     resp = await client.get("/")
     assert CONTENT_LENGTH not in resp.headers
     assert TRANSFER_ENCODING not in resp.headers
-    await resp.release()
+    resp.release()
 
 
 async def test_stream_response_headers_204(aiohttp_client: Any):
@@ -2128,7 +2144,7 @@ async def test_stream_response_headers_204(aiohttp_client: Any):
     resp = await client.get("/")
     assert CONTENT_TYPE not in resp.headers
     assert TRANSFER_ENCODING not in resp.headers
-    await resp.release()
+    resp.release()
 
 
 async def test_httpfound_cookies_302(aiohttp_client: Any) -> None:
@@ -2143,7 +2159,7 @@ async def test_httpfound_cookies_302(aiohttp_client: Any) -> None:
 
     resp = await client.get("/", allow_redirects=False)
     assert "my-cookie" in resp.cookies
-    await resp.release()
+    resp.release()
 
 
 @pytest.mark.parametrize("status", (101, 204, 304))
@@ -2163,4 +2179,4 @@ async def test_no_body_for_1xx_204_304_responses(
     assert CONTENT_TYPE not in resp.headers
     assert TRANSFER_ENCODING not in resp.headers
     await resp.read() == b""
-    await resp.release()
+    resp.release()
