@@ -103,7 +103,7 @@ async def test_static_file_ok(
     assert "file content" == txt.rstrip()
     assert "application/octet-stream" == resp.headers["Content-Type"]
     assert resp.headers.get("Content-Encoding") is None
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -127,7 +127,7 @@ async def test_zero_bytes_file_ok(aiohttp_client: Any, sender: Any) -> None:
         assert "" == txt.rstrip()
         assert "application/octet-stream" == resp.headers["Content-Type"]
         assert resp.headers.get("Content-Encoding") is None
-        await resp.release()
+        resp.release()
 
     await client.close()
 
@@ -156,7 +156,7 @@ async def test_zero_bytes_file_mocked_native_sendfile(
         assert "application/octet-stream" == resp.headers["Content-Type"]
         assert resp.headers.get("Content-Encoding") is None
         assert resp.headers.get("Content-Length") == "0"
-        await resp.release()
+        resp.release()
 
     await client.close()
 
@@ -172,7 +172,7 @@ async def test_static_file_ok_string_path(
     assert "file content" == txt.rstrip()
     assert "application/octet-stream" == resp.headers["Content-Type"]
     assert resp.headers.get("Content-Encoding") is None
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -182,7 +182,7 @@ async def test_static_file_not_exists(aiohttp_client: Any) -> None:
 
     resp = await client.get("/fake")
     assert resp.status == 404
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -192,7 +192,7 @@ async def test_static_file_name_too_long(aiohttp_client: Any) -> None:
 
     resp = await client.get("/x*500")
     assert resp.status == 404
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -202,7 +202,7 @@ async def test_static_file_upper_directory(aiohttp_client: Any) -> None:
 
     resp = await client.get("/../../")
     assert resp.status == 404
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -225,7 +225,7 @@ async def test_static_file_with_content_type(aiohttp_client: Any, sender: Any) -
     assert resp.headers["Content-Type"] == "image/jpeg"
     assert resp.headers.get("Content-Encoding") is None
     resp.close()
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -250,7 +250,7 @@ async def test_static_file_custom_content_type(
     assert resp.headers["Content-Type"] == "application/pdf"
     assert await resp.read() == hello_txt.read_bytes()
     resp.close()
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -282,7 +282,7 @@ async def test_static_file_custom_content_type_compress(
     assert resp.headers["Content-Type"] == "application/pdf"
     assert await resp.read() == HELLO_AIOHTTP
     resp.close()
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -316,7 +316,7 @@ async def test_static_file_with_encoding_and_enable_compression(
     assert resp.headers["Content-Type"] == "text/plain"
     assert await resp.read() == HELLO_AIOHTTP
     resp.close()
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -348,7 +348,7 @@ async def test_static_file_with_content_encoding(
     assert await resp.read() == hello_txt.read_bytes()
     resp.close()
 
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -362,7 +362,7 @@ async def test_static_file_if_modified_since(
     lastmod = resp.headers.get("Last-Modified")
     assert lastmod is not None
     resp.close()
-    await resp.release()
+    resp.release()
 
     resp = await client.get("/", headers={"If-Modified-Since": lastmod})
     body = await resp.read()
@@ -371,7 +371,7 @@ async def test_static_file_if_modified_since(
     assert resp.headers.get("Last-Modified") == lastmod
     assert b"" == body
     resp.close()
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -386,7 +386,7 @@ async def test_static_file_if_modified_since_past_date(
     assert 200 == resp.status
     resp.close()
 
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -401,7 +401,7 @@ async def test_static_file_if_modified_since_invalid_date(
     assert 200 == resp.status
     resp.close()
 
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -420,7 +420,7 @@ async def test_static_file_if_modified_since_future_date(
     assert b"" == body
     resp.close()
 
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -438,7 +438,7 @@ async def test_static_file_if_match(
 
     assert original_etag is not None
     resp.close()
-    await resp.release()
+    resp.release()
 
     headers = {"If-Match": original_etag, "If-Unmodified-Since": if_unmodified_since}
     resp = await client.head("/", headers=headers)
@@ -448,7 +448,7 @@ async def test_static_file_if_match(
     assert resp.headers.get("Last-Modified")
     assert b"" == body
     resp.close()
-    await resp.release()
+    resp.release()
 
     await client.close()
 
@@ -478,7 +478,7 @@ async def test_static_file_if_match_custom_tags(
     assert b"" == body
     resp.close()
 
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -505,7 +505,7 @@ async def test_static_file_if_none_match(
     assert resp.headers.get("Last-Modified") is not None
     assert original_etag is not None
     resp.close()
-    await resp.release()
+    resp.release()
 
     etag = ",".join((original_etag, *additional_etags))
 
@@ -518,7 +518,7 @@ async def test_static_file_if_none_match(
     assert resp.headers.get("ETag") == original_etag
     assert b"" == body
     resp.close()
-    await resp.release()
+    resp.release()
 
     await client.close()
 
@@ -538,7 +538,7 @@ async def test_static_file_if_none_match_star(
     assert b"" == body
     resp.close()
 
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -565,7 +565,7 @@ async def test_static_file_ssl(
     assert "application/octet-stream" == ct
     assert resp.headers.get("CONTENT-ENCODING") is None
 
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -581,17 +581,17 @@ async def test_static_file_directory_traversal_attack(aiohttp_client: Any) -> No
 
     resp = await client.get("/static/" + relpath)
     assert 404 == resp.status
-    await resp.release()
+    resp.release()
 
     url_relpath2 = "/static/dir/../" + relpath
     resp = await client.get(url_relpath2)
     assert 404 == resp.status
-    await resp.release()
+    resp.release()
 
     url_abspath = "/static/" + str(full_path.resolve())
     resp = await client.get(url_abspath)
     assert 403 == resp.status
-    await resp.release()
+    resp.release()
 
     await client.close()
 
@@ -628,7 +628,7 @@ async def test_static_file_huge(aiohttp_client: Any, tmp_path: Any) -> None:
         cnt += 1
     f.close()
 
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -683,9 +683,8 @@ async def test_static_file_range(aiohttp_client: Any, sender: Any) -> None:
     responses[1].close()
     responses[2].close()
 
-    await asyncio.gather(
-        *(resp.release() for resp in responses),
-    )
+    for resp in responses:
+        resp.release()
 
     assert content == b"".join(body)
 
@@ -722,7 +721,7 @@ async def test_static_file_range_end_bigger_than_size(aiohttp_client: Any, sende
 
         assert content[54000:] == body
 
-    await response.release()
+    response.release()
     await client.close()
 
 
@@ -743,7 +742,7 @@ async def test_static_file_range_beyond_eof(aiohttp_client: Any, sender: Any) ->
         "failed 'bytes=1000000-1200000': %s" % response.reason
     )
 
-    await response.release()
+    response.release()
     await client.close()
 
 
@@ -768,7 +767,7 @@ async def test_static_file_range_tail(aiohttp_client: Any, sender: Any) -> None:
     ), "failed: Content-Range Error"
     body4 = await resp.read()
     resp.close()
-    await resp.release()
+    resp.release()
     assert content[-500:] == body4
 
     # Ensure out-of-range tails could be handled
@@ -777,7 +776,7 @@ async def test_static_file_range_tail(aiohttp_client: Any, sender: Any) -> None:
     assert (
         resp2.headers["Content-Range"] == "bytes 0-54996/54997"
     ), "failed: Content-Range Error"
-    await resp2.release()
+    resp2.release()
 
     await client.close()
 
@@ -796,37 +795,37 @@ async def test_static_file_invalid_range(aiohttp_client: Any, sender: Any) -> No
     resp = await client.get("/", headers={"Range": "blocks=0-10"})
     assert resp.status == 416, "Range must be in bytes"
     resp.close()
-    await resp.release()
+    resp.release()
 
     # start > end
     resp = await client.get("/", headers={"Range": "bytes=100-0"})
     assert resp.status == 416, "Range start can't be greater than end"
     resp.close()
-    await resp.release()
+    resp.release()
 
     # start > end
     resp = await client.get("/", headers={"Range": "bytes=10-9"})
     assert resp.status == 416, "Range start can't be greater than end"
     resp.close()
-    await resp.release()
+    resp.release()
 
     # non-number range
     resp = await client.get("/", headers={"Range": "bytes=a-f"})
     assert resp.status == 416, "Range must be integers"
     resp.close()
-    await resp.release()
+    resp.release()
 
     # double dash range
     resp = await client.get("/", headers={"Range": "bytes=0--10"})
     assert resp.status == 416, "double dash in range"
     resp.close()
-    await resp.release()
+    resp.release()
 
     # no range
     resp = await client.get("/", headers={"Range": "bytes=-"})
     assert resp.status == 416, "no range given"
     resp.close()
-    await resp.release()
+    resp.release()
 
     await client.close()
 
@@ -843,7 +842,7 @@ async def test_static_file_if_unmodified_since_past_with_range(
     )
     assert 412 == resp.status
     resp.close()
-    await resp.release()
+    resp.release()
 
     await client.close()
 
@@ -862,7 +861,7 @@ async def test_static_file_if_unmodified_since_future_with_range(
     assert resp.headers["Content-Range"] == "bytes 2-12/13"
     assert resp.headers["Content-Length"] == "11"
     resp.close()
-    await resp.release()
+    resp.release()
 
     await client.close()
 
@@ -878,7 +877,7 @@ async def test_static_file_if_range_past_with_range(
     assert 200 == resp.status
     assert resp.headers["Content-Length"] == "13"
     resp.close()
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -895,7 +894,7 @@ async def test_static_file_if_range_future_with_range(
     assert resp.headers["Content-Length"] == "11"
     resp.close()
 
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -910,7 +909,7 @@ async def test_static_file_if_unmodified_since_past_without_range(
     assert 412 == resp.status
     resp.close()
 
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -926,7 +925,7 @@ async def test_static_file_if_unmodified_since_future_without_range(
     assert resp.headers["Content-Length"] == "13"
     resp.close()
 
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -942,7 +941,7 @@ async def test_static_file_if_range_past_without_range(
     assert resp.headers["Content-Length"] == "13"
     resp.close()
 
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -958,7 +957,7 @@ async def test_static_file_if_range_future_without_range(
     assert resp.headers["Content-Length"] == "13"
     resp.close()
 
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -973,7 +972,7 @@ async def test_static_file_if_unmodified_since_invalid_date(
     assert 200 == resp.status
     resp.close()
 
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -987,7 +986,7 @@ async def test_static_file_if_range_invalid_date(
     resp = await client.get("/", headers={"If-Range": lastmod})
     assert 200 == resp.status
     resp.close()
-    await resp.release()
+    resp.release()
 
     await client.close()
 
@@ -1011,7 +1010,7 @@ async def test_static_file_compression(aiohttp_client: Any, sender: Any) -> None
     assert expected_body == await resp.read()
     assert "application/octet-stream" == resp.headers["Content-Type"]
     assert resp.headers.get("Content-Encoding") == "deflate"
-    await resp.release()
+    resp.release()
 
     await client.close()
 
@@ -1053,7 +1052,7 @@ async def test_static_file_huge_cancel(aiohttp_client: Any, tmp_path: Any) -> No
             break
     assert len(data) < 1024 * 1024 * 20
 
-    await resp.release()
+    resp.release()
     await client.close()
 
 
@@ -1083,5 +1082,5 @@ async def test_static_file_huge_error(aiohttp_client: Any, tmp_path: Any) -> Non
     # raise an exception on server side
     resp.close()
 
-    await resp.release()
+    resp.release()
     await client.close()
