@@ -1,9 +1,8 @@
-# type: ignore
+import asyncio
 import pickle
 import random
 import struct
 import zlib
-from typing import Any
 from unittest import mock
 
 import pytest
@@ -23,13 +22,13 @@ from aiohttp.http_websocket import (
 
 
 def build_frame(
-    message: Any,
-    opcode: Any,
+    message: bytes,
+    opcode: int,
     use_mask: bool = False,
     noheader: bool = False,
     is_fin: bool = True,
     compress: bool = False,
-):
+) -> bytes:
     # Send a frame over the websocket with message as its payload.
     if compress:
         compressobj = zlib.compressobj(wbits=-9)
@@ -59,8 +58,8 @@ def build_frame(
         header = PACK_LEN3(header_first_byte, 127 | mask_bit, msg_length)
 
     if use_mask:  # pragma: no cover
-        mask = random.randrange(0, 0xFFFFFFFF)
-        mask = mask.to_bytes(4, "big")
+        maski = random.randrange(0, 0xFFFFFFFF)
+        mask = maski.to_bytes(4, "big")
         message = bytearray(message)
         _websocket_mask(mask, message)
         if noheader:
@@ -74,7 +73,9 @@ def build_frame(
             return header + message
 
 
-def build_close_frame(code: int = 1000, message: bytes = b"", noheader: bool = False):
+def build_close_frame(
+    code: int = 1000, message: bytes = b"", noheader: bool = False
+) -> bytes:
     # Close the websocket, sending the specified code and message.
     if isinstance(message, str):  # pragma: no cover
         message = message.encode("utf-8")
@@ -84,16 +85,16 @@ def build_close_frame(code: int = 1000, message: bytes = b"", noheader: bool = F
 
 
 @pytest.fixture()
-def out(loop: Any):
+def out(loop: asyncio.AbstractEventLoop) -> aiohttp.DataQueue[WSMessage]:
     return aiohttp.DataQueue(loop)
 
 
 @pytest.fixture()
-def parser(out: Any):
+def parser(out: aiohttp.DataQueue[WSMessage]) -> WebSocketReader:
     return WebSocketReader(out, 4 * 1024 * 1024)
 
 
-def test_parse_frame(parser: Any) -> None:
+def test_parse_frame(parser: WebSocketReader) -> None:
     parser.parse_frame(struct.pack("!BB", 0b00000001, 0b00000001))
     res = parser.parse_frame(b"1")
     fin, opcode, payload, compress = res[0]
@@ -101,7 +102,7 @@ def test_parse_frame(parser: Any) -> None:
     assert (0, 1, b"1", False) == (fin, opcode, payload, not not compress)
 
 
-def test_parse_frame_length0(parser: Any) -> None:
+def test_parse_frame_length0(parser: WebSocketReader) -> None:
     fin, opcode, payload, compress = parser.parse_frame(
         struct.pack("!BB", 0b00000001, 0b00000000)
     )[0]
@@ -109,7 +110,7 @@ def test_parse_frame_length0(parser: Any) -> None:
     assert (0, 1, b"", False) == (fin, opcode, payload, not not compress)
 
 
-def test_parse_frame_length2(parser: Any) -> None:
+def test_parse_frame_length2(parser: WebSocketReader) -> None:
     parser.parse_frame(struct.pack("!BB", 0b00000001, 126))
     parser.parse_frame(struct.pack("!H", 4))
     res = parser.parse_frame(b"1234")
@@ -118,7 +119,7 @@ def test_parse_frame_length2(parser: Any) -> None:
     assert (0, 1, b"1234", False) == (fin, opcode, payload, not not compress)
 
 
-def test_parse_frame_length4(parser: Any) -> None:
+def test_parse_frame_length4(parser: WebSocketReader) -> None:
     parser.parse_frame(struct.pack("!BB", 0b00000001, 127))
     parser.parse_frame(struct.pack("!Q", 4))
     fin, opcode, payload, compress = parser.parse_frame(b"1234")[0]
@@ -126,7 +127,7 @@ def test_parse_frame_length4(parser: Any) -> None:
     assert (0, 1, b"1234", False) == (fin, opcode, payload, not not compress)
 
 
-def test_parse_frame_mask(parser: Any) -> None:
+def test_parse_frame_mask(parser: WebSocketReader) -> None:
     parser.parse_frame(struct.pack("!BB", 0b00000001, 0b10000001))
     parser.parse_frame(b"0001")
     fin, opcode, payload, compress = parser.parse_frame(b"1")[0]
@@ -134,76 +135,90 @@ def test_parse_frame_mask(parser: Any) -> None:
     assert (0, 1, b"\x01", False) == (fin, opcode, payload, not not compress)
 
 
-def test_parse_frame_header_reversed_bits(out: Any, parser: Any) -> None:
+def test_parse_frame_header_reversed_bits(
+    out: aiohttp.DataQueue[WSMessage], parser: WebSocketReader
+) -> None:
     with pytest.raises(WebSocketError):
         parser.parse_frame(struct.pack("!BB", 0b01100000, 0b00000000))
-        raise out.exception()
 
 
-def test_parse_frame_header_control_frame(out: Any, parser: Any) -> None:
+def test_parse_frame_header_control_frame(
+    out: aiohttp.DataQueue[WSMessage], parser: WebSocketReader
+) -> None:
     with pytest.raises(WebSocketError):
         parser.parse_frame(struct.pack("!BB", 0b00001000, 0b00000000))
-        raise out.exception()
 
 
-def _test_parse_frame_header_new_data_err(out, parser):
+@pytest.mark.xfail()
+def test_parse_frame_header_new_data_err(
+    out: aiohttp.DataQueue[WSMessage], parser: WebSocketReader
+) -> None:
     with pytest.raises(WebSocketError):
         parser.parse_frame(struct.pack("!BB", 0b000000000, 0b00000000))
-        raise out.exception()
 
 
-def test_parse_frame_header_payload_size(out: Any, parser: Any) -> None:
+def test_parse_frame_header_payload_size(
+    out: aiohttp.DataQueue[WSMessage], parser: WebSocketReader
+) -> None:
     with pytest.raises(WebSocketError):
         parser.parse_frame(struct.pack("!BB", 0b10001000, 0b01111110))
-        raise out.exception()
 
 
-def test_ping_frame(out: Any, parser: Any) -> None:
-    parser.parse_frame = mock.Mock()
-    parser.parse_frame.return_value = [(1, WSMsgType.PING, b"data", False)]
+def test_ping_frame(out: aiohttp.DataQueue[WSMessage], parser: WebSocketReader) -> None:
+    with mock.patch.object(parser, "parse_frame", autospec=True) as m:
+        m.return_value = [(1, WSMsgType.PING, b"data", False)]
 
-    parser.feed_data(b"")
-    res = out._buffer[0]
-    assert res == (WSMsgType.PING, b"data", "")
-
-
-def test_pong_frame(out: Any, parser: Any) -> None:
-    parser.parse_frame = mock.Mock()
-    parser.parse_frame.return_value = [(1, WSMsgType.PONG, b"data", False)]
-
-    parser.feed_data(b"")
-    res = out._buffer[0]
-    assert res == (WSMsgType.PONG, b"data", "")
+        parser.feed_data(b"")
+        res = out._buffer[0]
+        assert res == (WSMsgType.PING, b"data", "")
 
 
-def test_close_frame(out: Any, parser: Any) -> None:
-    parser.parse_frame = mock.Mock()
-    parser.parse_frame.return_value = [(1, WSMsgType.CLOSE, b"", False)]
+def test_pong_frame(out: aiohttp.DataQueue[WSMessage], parser: WebSocketReader) -> None:
+    with mock.patch.object(parser, "parse_frame", autospec=True) as m:
+        m.return_value = [(1, WSMsgType.PONG, b"data", False)]
 
-    parser.feed_data(b"")
-    res = out._buffer[0]
-    assert res == (WSMsgType.CLOSE, 0, "")
-
-
-def test_close_frame_info(out: Any, parser: Any) -> None:
-    parser.parse_frame = mock.Mock()
-    parser.parse_frame.return_value = [(1, WSMsgType.CLOSE, b"0112345", False)]
-
-    parser.feed_data(b"")
-    res = out._buffer[0]
-    assert res == (WSMessage(WSMsgType.CLOSE, 12337, "12345"))
+        parser.feed_data(b"")
+        res = out._buffer[0]
+        assert res == (WSMsgType.PONG, b"data", "")
 
 
-def test_close_frame_invalid(out: Any, parser: Any) -> None:
-    parser.parse_frame = mock.Mock()
-    parser.parse_frame.return_value = [(1, WSMsgType.CLOSE, b"1", False)]
-    parser.feed_data(b"")
+def test_close_frame(
+    out: aiohttp.DataQueue[WSMessage], parser: WebSocketReader
+) -> None:
+    with mock.patch.object(parser, "parse_frame", autospec=True) as m:
+        m.return_value = [(1, WSMsgType.CLOSE, b"", False)]
 
-    assert isinstance(out.exception(), WebSocketError)
-    assert out.exception().code == WSCloseCode.PROTOCOL_ERROR
+        parser.feed_data(b"")
+        res = out._buffer[0]
+        assert res == (WSMsgType.CLOSE, 0, "")
 
 
-def test_close_frame_invalid_2(out: Any, parser: Any) -> None:
+def test_close_frame_info(
+    out: aiohttp.DataQueue[WSMessage], parser: WebSocketReader
+) -> None:
+    with mock.patch.object(parser, "parse_frame", autospec=True) as m:
+        m.return_value = [(1, WSMsgType.CLOSE, b"0112345", False)]
+
+        parser.feed_data(b"")
+        res = out._buffer[0]
+        assert res == WSMessage(WSMsgType.CLOSE, 12337, "12345")
+
+
+def test_close_frame_invalid(
+    out: aiohttp.DataQueue[WSMessage], parser: WebSocketReader
+) -> None:
+    with mock.patch.object(parser, "parse_frame", autospec=True) as m:
+        m.return_value = [(1, WSMsgType.CLOSE, b"1", False)]
+        parser.feed_data(b"")
+
+        exc = out.exception()
+        assert isinstance(exc, WebSocketError)
+        assert exc.code == WSCloseCode.PROTOCOL_ERROR
+
+
+def test_close_frame_invalid_2(
+    out: aiohttp.DataQueue[WSMessage], parser: WebSocketReader
+) -> None:
     data = build_close_frame(code=1)
 
     with pytest.raises(WebSocketError) as ctx:
@@ -212,7 +227,7 @@ def test_close_frame_invalid_2(out: Any, parser: Any) -> None:
     assert ctx.value.code == WSCloseCode.PROTOCOL_ERROR
 
 
-def test_close_frame_unicode_err(parser: Any) -> None:
+def test_close_frame_unicode_err(parser: WebSocketReader) -> None:
     data = build_close_frame(code=1000, message=b"\xf4\x90\x80\x80")
 
     with pytest.raises(WebSocketError) as ctx:
@@ -221,23 +236,26 @@ def test_close_frame_unicode_err(parser: Any) -> None:
     assert ctx.value.code == WSCloseCode.INVALID_TEXT
 
 
-def test_unknown_frame(out: Any, parser: Any) -> None:
-    parser.parse_frame = mock.Mock()
-    parser.parse_frame.return_value = [(1, WSMsgType.CONTINUATION, b"", False)]
+def test_unknown_frame(
+    out: aiohttp.DataQueue[WSMessage], parser: WebSocketReader
+) -> None:
+    with mock.patch.object(parser, "parse_frame", autospec=True) as m:
+        m.return_value = [(1, WSMsgType.CONTINUATION, b"", False)]
 
-    with pytest.raises(WebSocketError):
         parser.feed_data(b"")
-        raise out.exception()
+        assert isinstance(out.exception(), WebSocketError)
 
 
-def test_simple_text(out: Any, parser: Any) -> None:
+def test_simple_text(
+    out: aiohttp.DataQueue[WSMessage], parser: WebSocketReader
+) -> None:
     data = build_frame(b"text", WSMsgType.TEXT)
     parser._feed_data(data)
     res = out._buffer[0]
     assert res == (WSMsgType.TEXT, "text", "")
 
 
-def test_simple_text_unicode_err(parser: Any) -> None:
+def test_simple_text_unicode_err(parser: WebSocketReader) -> None:
     data = build_frame(b"\xf4\x90\x80\x80", WSMsgType.TEXT)
 
     with pytest.raises(WebSocketError) as ctx:
@@ -246,16 +264,20 @@ def test_simple_text_unicode_err(parser: Any) -> None:
     assert ctx.value.code == WSCloseCode.INVALID_TEXT
 
 
-def test_simple_binary(out: Any, parser: Any) -> None:
-    parser.parse_frame = mock.Mock()
-    parser.parse_frame.return_value = [(1, WSMsgType.BINARY, b"binary", False)]
+def test_simple_binary(
+    out: aiohttp.DataQueue[WSMessage], parser: WebSocketReader
+) -> None:
+    with mock.patch.object(parser, "parse_frame", autospec=True) as m:
+        m.return_value = [(1, WSMsgType.BINARY, b"binary", False)]
 
-    parser.feed_data(b"")
-    res = out._buffer[0]
-    assert res == (WSMsgType.BINARY, b"binary", "")
+        parser.feed_data(b"")
+        res = out._buffer[0]
+        assert res == (WSMsgType.BINARY, b"binary", "")
 
 
-def test_fragmentation_header(out: Any, parser: Any) -> None:
+def test_fragmentation_header(
+    out: aiohttp.DataQueue[WSMessage], parser: WebSocketReader
+) -> None:
     data = build_frame(b"a", WSMsgType.TEXT)
     parser._feed_data(data[:1])
     parser._feed_data(data[1:])
@@ -264,7 +286,9 @@ def test_fragmentation_header(out: Any, parser: Any) -> None:
     assert res == (WSMessage(WSMsgType.TEXT, "a", ""))
 
 
-def test_continuation(out: Any, parser: Any) -> None:
+def test_continuation(
+    out: aiohttp.DataQueue[WSMessage], parser: WebSocketReader
+) -> None:
     data1 = build_frame(b"line1", WSMsgType.TEXT, is_fin=False)
     parser._feed_data(data1)
 
@@ -275,121 +299,140 @@ def test_continuation(out: Any, parser: Any) -> None:
     assert res == (WSMessage(WSMsgType.TEXT, "line1line2", ""))
 
 
-def test_continuation_with_ping(out: Any, parser: Any) -> None:
-    parser.parse_frame = mock.Mock()
-    parser.parse_frame.return_value = [
-        (0, WSMsgType.TEXT, b"line1", False),
-        (0, WSMsgType.PING, b"", False),
-        (1, WSMsgType.CONTINUATION, b"line2", False),
-    ]
+def test_continuation_with_ping(
+    out: aiohttp.DataQueue[WSMessage], parser: WebSocketReader
+) -> None:
+    with mock.patch.object(parser, "parse_frame", autospec=True) as m:
+        m.return_value = [
+            (0, WSMsgType.TEXT, b"line1", False),
+            (0, WSMsgType.PING, b"", False),
+            (1, WSMsgType.CONTINUATION, b"line2", False),
+        ]
 
-    data1 = build_frame(b"line1", WSMsgType.TEXT, is_fin=False)
-    parser._feed_data(data1)
+        data1 = build_frame(b"line1", WSMsgType.TEXT, is_fin=False)
+        parser._feed_data(data1)
 
-    data2 = build_frame(b"", WSMsgType.PING)
-    parser._feed_data(data2)
+        data2 = build_frame(b"", WSMsgType.PING)
+        parser._feed_data(data2)
 
-    data3 = build_frame(b"line2", WSMsgType.CONTINUATION)
-    parser._feed_data(data3)
+        data3 = build_frame(b"line2", WSMsgType.CONTINUATION)
+        parser._feed_data(data3)
 
-    res = out._buffer[0]
-    assert res == (WSMessage(WSMsgType.PING, b"", ""))
-    res = out._buffer[1]
-    assert res == (WSMessage(WSMsgType.TEXT, "line1line2", ""))
-
-
-def test_continuation_err(out: Any, parser: Any) -> None:
-    parser.parse_frame = mock.Mock()
-    parser.parse_frame.return_value = [
-        (0, WSMsgType.TEXT, b"line1", False),
-        (1, WSMsgType.TEXT, b"line2", False),
-    ]
-
-    with pytest.raises(WebSocketError):
-        parser._feed_data(b"")
+        res = out._buffer[0]
+        assert res == WSMessage(WSMsgType.PING, b"", "")
+        res = out._buffer[1]
+        assert res == WSMessage(WSMsgType.TEXT, "line1line2", "")
 
 
-def test_continuation_with_close(out: Any, parser: Any) -> None:
-    parser.parse_frame = mock.Mock()
-    parser.parse_frame.return_value = [
-        (0, WSMsgType.TEXT, b"line1", False),
-        (0, WSMsgType.CLOSE, build_close_frame(1002, b"test", noheader=True), False),
-        (1, WSMsgType.CONTINUATION, b"line2", False),
-    ]
+def test_continuation_err(
+    out: aiohttp.DataQueue[WSMessage], parser: WebSocketReader
+) -> None:
+    with mock.patch.object(parser, "parse_frame", autospec=True) as m:
+        m.return_value = [
+            (0, WSMsgType.TEXT, b"line1", False),
+            (1, WSMsgType.TEXT, b"line2", False),
+        ]
 
-    parser.feed_data(b"")
-    res = out._buffer[0]
-    assert res, WSMessage(WSMsgType.CLOSE, 1002, "test")
-    res = out._buffer[1]
-    assert res == (WSMessage(WSMsgType.TEXT, "line1line2", ""))
+        with pytest.raises(WebSocketError):
+            parser._feed_data(b"")
 
 
-def test_continuation_with_close_unicode_err(out: Any, parser: Any) -> None:
-    parser.parse_frame = mock.Mock()
-    parser.parse_frame.return_value = [
-        (0, WSMsgType.TEXT, b"line1", False),
-        (
-            0,
-            WSMsgType.CLOSE,
-            build_close_frame(1000, b"\xf4\x90\x80\x80", noheader=True),
-            False,
-        ),
-        (1, WSMsgType.CONTINUATION, b"line2", False),
-    ]
+def test_continuation_with_close(
+    out: aiohttp.DataQueue[WSMessage], parser: WebSocketReader
+) -> None:
+    with mock.patch.object(parser, "parse_frame", autospec=True) as m:
+        m.return_value = [
+            (0, WSMsgType.TEXT, b"line1", False),
+            (
+                0,
+                WSMsgType.CLOSE,
+                build_close_frame(1002, b"test", noheader=True),
+                False,
+            ),
+            (1, WSMsgType.CONTINUATION, b"line2", False),
+        ]
 
-    with pytest.raises(WebSocketError) as ctx:
-        parser._feed_data(b"")
-
-    assert ctx.value.code == WSCloseCode.INVALID_TEXT
-
-
-def test_continuation_with_close_bad_code(out: Any, parser: Any) -> None:
-    parser.parse_frame = mock.Mock()
-    parser.parse_frame.return_value = [
-        (0, WSMsgType.TEXT, b"line1", False),
-        (0, WSMsgType.CLOSE, build_close_frame(1, b"test", noheader=True), False),
-        (1, WSMsgType.CONTINUATION, b"line2", False),
-    ]
-
-    with pytest.raises(WebSocketError) as ctx:
-        parser._feed_data(b"")
-
-    assert ctx.value.code == WSCloseCode.PROTOCOL_ERROR
+        parser.feed_data(b"")
+        res = out._buffer[0]
+        assert res, WSMessage(WSMsgType.CLOSE, 1002, "test")
+        res = out._buffer[1]
+        assert res == WSMessage(WSMsgType.TEXT, "line1line2", "")
 
 
-def test_continuation_with_close_bad_payload(out: Any, parser: Any) -> None:
-    parser.parse_frame = mock.Mock()
-    parser.parse_frame.return_value = [
-        (0, WSMsgType.TEXT, b"line1", False),
-        (0, WSMsgType.CLOSE, b"1", False),
-        (1, WSMsgType.CONTINUATION, b"line2", False),
-    ]
+def test_continuation_with_close_unicode_err(
+    out: aiohttp.DataQueue[WSMessage], parser: WebSocketReader
+) -> None:
+    with mock.patch.object(parser, "parse_frame", autospec=True) as m:
+        m.return_value = [
+            (0, WSMsgType.TEXT, b"line1", False),
+            (
+                0,
+                WSMsgType.CLOSE,
+                build_close_frame(1000, b"\xf4\x90\x80\x80", noheader=True),
+                False,
+            ),
+            (1, WSMsgType.CONTINUATION, b"line2", False),
+        ]
 
-    with pytest.raises(WebSocketError) as ctx:
-        parser._feed_data(b"")
+        with pytest.raises(WebSocketError) as ctx:
+            parser._feed_data(b"")
 
-    assert ctx.value.code, WSCloseCode.PROTOCOL_ERROR
+        assert ctx.value.code == WSCloseCode.INVALID_TEXT
 
 
-def test_continuation_with_close_empty(out: Any, parser: Any) -> None:
-    parser.parse_frame = mock.Mock()
-    parser.parse_frame.return_value = [
-        (0, WSMsgType.TEXT, b"line1", False),
-        (0, WSMsgType.CLOSE, b"", False),
-        (1, WSMsgType.CONTINUATION, b"line2", False),
-    ]
+def test_continuation_with_close_bad_code(
+    out: aiohttp.DataQueue[WSMessage], parser: WebSocketReader
+) -> None:
+    with mock.patch.object(parser, "parse_frame", autospec=True) as m:
+        m.return_value = [
+            (0, WSMsgType.TEXT, b"line1", False),
+            (0, WSMsgType.CLOSE, build_close_frame(1, b"test", noheader=True), False),
+            (1, WSMsgType.CONTINUATION, b"line2", False),
+        ]
 
-    parser.feed_data(b"")
-    res = out._buffer[0]
-    assert res, WSMessage(WSMsgType.CLOSE, 0, "")
-    res = out._buffer[1]
-    assert res == (WSMessage(WSMsgType.TEXT, "line1line2", ""))
+        with pytest.raises(WebSocketError) as ctx:
+            parser._feed_data(b"")
+
+        assert ctx.value.code == WSCloseCode.PROTOCOL_ERROR
+
+
+def test_continuation_with_close_bad_payload(
+    out: aiohttp.DataQueue[WSMessage], parser: WebSocketReader
+) -> None:
+    with mock.patch.object(parser, "parse_frame", autospec=True) as m:
+        m.return_value = [
+            (0, WSMsgType.TEXT, b"line1", False),
+            (0, WSMsgType.CLOSE, b"1", False),
+            (1, WSMsgType.CONTINUATION, b"line2", False),
+        ]
+
+        with pytest.raises(WebSocketError) as ctx:
+            parser._feed_data(b"")
+
+        assert ctx.value.code, WSCloseCode.PROTOCOL_ERROR
+
+
+def test_continuation_with_close_empty(
+    out: aiohttp.DataQueue[WSMessage], parser: WebSocketReader
+) -> None:
+    with mock.patch.object(parser, "parse_frame", autospec=True) as m:
+        m.return_value = [
+            (0, WSMsgType.TEXT, b"line1", False),
+            (0, WSMsgType.CLOSE, b"", False),
+            (1, WSMsgType.CONTINUATION, b"line2", False),
+        ]
+
+        parser.feed_data(b"")
+        res = out._buffer[0]
+        assert res, WSMessage(WSMsgType.CLOSE, 0, "")
+        res = out._buffer[1]
+        assert res == WSMessage(WSMsgType.TEXT, "line1line2", "")
 
 
 websocket_mask_data: bytes = b"some very long data for masking by websocket"
 websocket_mask_mask: bytes = b"1234"
 websocket_mask_masked: bytes = (
-    b"B]^Q\x11DVFH\x12_[_U\x13PPFR\x14W]A\x14\\S@_X" b"\\T\x14SK\x13CTP@[RYV@"
+    b"B]^Q\x11DVFH\x12_[_U\x13PPFR\x14W]A\x14\\S@_X\\T\x14SK\x13CTP@[RYV@"
 )
 
 
@@ -404,7 +447,7 @@ def test_websocket_mask_python() -> None:
 )
 def test_websocket_mask_cython() -> None:
     message = bytearray(websocket_mask_data)
-    http_websocket._websocket_mask_cython(websocket_mask_mask, message)
+    http_websocket._websocket_mask_cython(websocket_mask_mask, message)  # type: ignore[attr-defined]
     assert message == websocket_mask_masked
 
 
@@ -419,11 +462,11 @@ def test_websocket_mask_python_empty() -> None:
 )
 def test_websocket_mask_cython_empty() -> None:
     message = bytearray()
-    http_websocket._websocket_mask_cython(websocket_mask_mask, message)
+    http_websocket._websocket_mask_cython(websocket_mask_mask, message)  # type: ignore[attr-defined]
     assert message == bytearray()
 
 
-def test_parse_compress_frame_single(parser: Any) -> None:
+def test_parse_compress_frame_single(parser: WebSocketReader) -> None:
     parser.parse_frame(struct.pack("!BB", 0b11000001, 0b00000001))
     res = parser.parse_frame(b"1")
     fin, opcode, payload, compress = res[0]
@@ -431,7 +474,7 @@ def test_parse_compress_frame_single(parser: Any) -> None:
     assert (1, 1, b"1", True) == (fin, opcode, payload, not not compress)
 
 
-def test_parse_compress_frame_multi(parser: Any) -> None:
+def test_parse_compress_frame_multi(parser: WebSocketReader) -> None:
     parser.parse_frame(struct.pack("!BB", 0b01000001, 126))
     parser.parse_frame(struct.pack("!H", 4))
     res = parser.parse_frame(b"1234")
@@ -451,7 +494,7 @@ def test_parse_compress_frame_multi(parser: Any) -> None:
     assert (1, 1, b"1234", False) == (fin, opcode, payload, not not compress)
 
 
-def test_parse_compress_error_frame(parser: Any) -> None:
+def test_parse_compress_error_frame(parser: WebSocketReader) -> None:
     parser.parse_frame(struct.pack("!BB", 0b01000001, 0b00000001))
     parser.parse_frame(b"1")
 
@@ -462,7 +505,7 @@ def test_parse_compress_error_frame(parser: Any) -> None:
     assert ctx.value.code == WSCloseCode.PROTOCOL_ERROR
 
 
-def test_parse_no_compress_frame_single() -> None:
+def test_parse_no_compress_frame_single(out: aiohttp.DataQueue[WSMessage]) -> None:
     parser_no_compress = WebSocketReader(out, 0, compress=False)
     with pytest.raises(WebSocketError) as ctx:
         parser_no_compress.parse_frame(struct.pack("!BB", 0b11000001, 0b00000001))
@@ -471,7 +514,7 @@ def test_parse_no_compress_frame_single() -> None:
     assert ctx.value.code == WSCloseCode.PROTOCOL_ERROR
 
 
-def test_msg_too_large(out: Any) -> None:
+def test_msg_too_large(out: aiohttp.DataQueue[WSMessage]) -> None:
     parser = WebSocketReader(out, 256, compress=False)
     data = build_frame(b"text" * 256, WSMsgType.TEXT)
     with pytest.raises(WebSocketError) as ctx:
@@ -479,7 +522,7 @@ def test_msg_too_large(out: Any) -> None:
     assert ctx.value.code == WSCloseCode.MESSAGE_TOO_BIG
 
 
-def test_msg_too_large_not_fin(out: Any) -> None:
+def test_msg_too_large_not_fin(out: aiohttp.DataQueue[WSMessage]) -> None:
     parser = WebSocketReader(out, 256, compress=False)
     data = build_frame(b"text" * 256, WSMsgType.TEXT, is_fin=False)
     with pytest.raises(WebSocketError) as ctx:
@@ -487,7 +530,7 @@ def test_msg_too_large_not_fin(out: Any) -> None:
     assert ctx.value.code == WSCloseCode.MESSAGE_TOO_BIG
 
 
-def test_compressed_msg_too_large(out: Any) -> None:
+def test_compressed_msg_too_large(out: aiohttp.DataQueue[WSMessage]) -> None:
     parser = WebSocketReader(out, 256, compress=True)
     data = build_frame(b"aaa" * 256, WSMsgType.TEXT, compress=True)
     with pytest.raises(WebSocketError) as ctx:
@@ -503,7 +546,7 @@ class TestWebSocketError:
 
     def test_pickle(self) -> None:
         err = WebSocketError(WSCloseCode.PROTOCOL_ERROR, "Something invalid")
-        err.foo = "bar"
+        err.foo = "bar"  # type: ignore[attr-defined]
         for proto in range(pickle.HIGHEST_PROTOCOL + 1):
             pickled = pickle.dumps(err, proto)
             err2 = pickle.loads(pickled)
